Check observations for uniqueness of grouping variables with respect to reference_date and report_date

This function checks that the input data is stratified by reference_date, report_date, and .group. It does this by counting the number of observations for each combination of these variables, and throwing a warning if any combination has more than one observation.

Usage

check_group_date_unique(obs)

Arguments

obs

An object that will be coerce_dtd in place, that contains .group, reference_date, and report_date columns.

See also