Constructs random effect terms

Constructs random effect terms

Usage

construct_re(re, data)

Arguments

re

A random effect as defined using re() which itself takes random effects specified in a model formula using the lme4 syntax.

data

A data.frame of observations used to define the random effects. Must contain the variables specified in the re() term.

Value

A list containing the transformed data ("data"), fixed effects terms ("terms") and a data.frame specifying the random effect structure between these terms (effects). Note that if the specified random effect was not a factor it will have been converted into one.

See also

Examples

form <- epinowcast:::parse_formula(~ 1 + (1 | day_of_week))
data <- enw_example("prepr")$metareference[[1]]
random_effect <- re(form$random[[1]])
epinowcast:::construct_re(random_effect, data)
#> $data
#>           date group location age_group delay day_of_week week month
#>  1: 2021-07-13     1       DE       00+     0     Tuesday    0     0
#>  2: 2021-07-14     1       DE       00+     0   Wednesday    0     0
#>  3: 2021-07-15     1       DE       00+     0    Thursday    0     0
#>  4: 2021-07-16     1       DE       00+     0      Friday    1     0
#>  5: 2021-07-17     1       DE       00+     0    Saturday    1     0
#>  6: 2021-07-18     1       DE       00+     0      Sunday    1     0
#>  7: 2021-07-19     1       DE       00+     0      Monday    1     0
#>  8: 2021-07-20     1       DE       00+     0     Tuesday    1     0
#>  9: 2021-07-21     1       DE       00+     0   Wednesday    1     0
#> 10: 2021-07-22     1       DE       00+     0    Thursday    1     0
#> 11: 2021-07-23     1       DE       00+     0      Friday    2     0
#> 12: 2021-07-24     1       DE       00+     0    Saturday    2     0
#> 13: 2021-07-25     1       DE       00+     0      Sunday    2     0
#> 14: 2021-07-26     1       DE       00+     0      Monday    2     0
#> 15: 2021-07-27     1       DE       00+     0     Tuesday    2     0
#> 16: 2021-07-28     1       DE       00+     0   Wednesday    2     0
#> 17: 2021-07-29     1       DE       00+     0    Thursday    2     0
#> 18: 2021-07-30     1       DE       00+     0      Friday    3     0
#> 19: 2021-07-31     1       DE       00+     0    Saturday    3     0
#> 20: 2021-08-01     1       DE       00+     0      Sunday    3     1
#> 21: 2021-08-02     1       DE       00+     0      Monday    3     1
#> 22: 2021-08-03     1       DE       00+     0     Tuesday    3     1
#> 23: 2021-08-04     1       DE       00+     0   Wednesday    3     1
#> 24: 2021-08-05     1       DE       00+     0    Thursday    3     1
#> 25: 2021-08-06     1       DE       00+     0      Friday    4     1
#> 26: 2021-08-07     1       DE       00+     0    Saturday    4     1
#> 27: 2021-08-08     1       DE       00+     0      Sunday    4     1
#> 28: 2021-08-09     1       DE       00+     0      Monday    4     1
#> 29: 2021-08-10     1       DE       00+     0     Tuesday    4     1
#> 30: 2021-08-11     1       DE       00+     0   Wednesday    4     1
#> 31: 2021-08-12     1       DE       00+     0    Thursday    4     1
#> 32: 2021-08-13     1       DE       00+     0      Friday    5     1
#> 33: 2021-08-14     1       DE       00+     0    Saturday    5     1
#> 34: 2021-08-15     1       DE       00+     0      Sunday    5     1
#> 35: 2021-08-16     1       DE       00+     0      Monday    5     1
#> 36: 2021-08-17     1       DE       00+     0     Tuesday    5     1
#> 37: 2021-08-18     1       DE       00+     0   Wednesday    5     1
#> 38: 2021-08-19     1       DE       00+     0    Thursday    5     1
#> 39: 2021-08-20     1       DE       00+     0      Friday    6     1
#> 40: 2021-08-21     1       DE       00+     0    Saturday    6     1
#> 41: 2021-08-22     1       DE       00+     0      Sunday    6     1
#>           date group location age_group delay day_of_week week month
#> 
#> $terms
#> [1] "day_of_week"
#> 
#> $effects
#>                 effects fixed day_of_week
#> 1:    day_of_weekFriday     0           1
#> 2:    day_of_weekMonday     0           1
#> 3:  day_of_weekSaturday     0           1
#> 4:    day_of_weekSunday     0           1
#> 5:  day_of_weekThursday     0           1
#> 6:   day_of_weekTuesday     0           1
#> 7: day_of_weekWednesday     0           1
#>