FUNCTION_TITLE

FUNCTION_DESCRIPTION

Usage

enw_add_metaobs_features(metaobs, holidays = c(), holidays_to = "Sunday")

Arguments

metaobs

PARAM_DESCRIPTION

holidays

PARAM_DESCRIPTION

holidays_to

A character string to assign to holidays when present. Replaces the day of the week and defaults to Sunday.

Value

OUTPUT_DESCRIPTION

See also