Expectation model module

Expectation model module

Usage

enw_expectation(formula = ~rw(day, .group), order = 1, data)

Arguments

formula

A formula (as implemented in enw_formula()) describing the generative process used for expected incidence. This can use features defined by reference date as defined in metareference as produced by enw_preprocess_data(). By default this is set to use a daily group specific random walk. Note that the daily group specific random walk is currently the only option supported by epinowcast().

order

Integer, defaults to 1. The order of the expectation process with 1 being a simple log scale generative process, 2 being a log scale generative process where each time point is offset by the log count from previous time points.

data

Output from enw_preprocess_data().

Value

A list containing the supplied formulas, data passed into a list describing the models, a data.frame describing the priors used, and a function that takes the output data and priors and returns a function that can be used to sample from a tightened version of the prior distribution.

See also

Examples

enw_expectation(data = enw_example("preprocessed"))
#> A grouped random walk using .group is not possible as this variable has fewer than 2 unique values.
#> $formula
#> $formula$expectation
#> [1] "~rw(day, .group)"
#> 
#> 
#> $data
#> $data$exp_fdesign
#>    (Intercept) cday1 cday2 cday3 cday4 cday5 cday6 cday7 cday8 cday9 cday10
#> 1            1     0     0     0     0     0     0     0     0     0      0
#> 2            1     1     0     0     0     0     0     0     0     0      0
#> 3            1     1     1     0     0     0     0     0     0     0      0
#> 4            1     1     1     1     0     0     0     0     0     0      0
#> 5            1     1     1     1     1     0     0     0     0     0      0
#> 6            1     1     1     1     1     1     0     0     0     0      0
#> 7            1     1     1     1     1     1     1     0     0     0      0
#> 8            1     1     1     1     1     1     1     1     0     0      0
#> 9            1     1     1     1     1     1     1     1     1     0      0
#> 10           1     1     1     1     1     1     1     1     1     1      0
#> 11           1     1     1     1     1     1     1     1     1     1      1
#> 12           1     1     1     1     1     1     1     1     1     1      1
#> 13           1     1     1     1     1     1     1     1     1     1      1
#> 14           1     1     1     1     1     1     1     1     1     1      1
#> 15           1     1     1     1     1     1     1     1     1     1      1
#> 16           1     1     1     1     1     1     1     1     1     1      1
#> 17           1     1     1     1     1     1     1     1     1     1      1
#> 18           1     1     1     1     1     1     1     1     1     1      1
#> 19           1     1     1     1     1     1     1     1     1     1      1
#> 20           1     1     1     1     1     1     1     1     1     1      1
#> 21           1     1     1     1     1     1     1     1     1     1      1
#> 22           1     1     1     1     1     1     1     1     1     1      1
#> 23           1     1     1     1     1     1     1     1     1     1      1
#> 24           1     1     1     1     1     1     1     1     1     1      1
#> 25           1     1     1     1     1     1     1     1     1     1      1
#> 26           1     1     1     1     1     1     1     1     1     1      1
#> 27           1     1     1     1     1     1     1     1     1     1      1
#> 28           1     1     1     1     1     1     1     1     1     1      1
#> 29           1     1     1     1     1     1     1     1     1     1      1
#> 30           1     1     1     1     1     1     1     1     1     1      1
#> 31           1     1     1     1     1     1     1     1     1     1      1
#> 32           1     1     1     1     1     1     1     1     1     1      1
#> 33           1     1     1     1     1     1     1     1     1     1      1
#> 34           1     1     1     1     1     1     1     1     1     1      1
#> 35           1     1     1     1     1     1     1     1     1     1      1
#> 36           1     1     1     1     1     1     1     1     1     1      1
#> 37           1     1     1     1     1     1     1     1     1     1      1
#> 38           1     1     1     1     1     1     1     1     1     1      1
#> 39           1     1     1     1     1     1     1     1     1     1      1
#> 40           1     1     1     1     1     1     1     1     1     1      1
#> 41           1     1     1     1     1     1     1     1     1     1      1
#>    cday11 cday12 cday13 cday14 cday15 cday16 cday17 cday18 cday19 cday20 cday21
#> 1       0      0      0      0      0      0      0      0      0      0      0
#> 2       0      0      0      0      0      0      0      0      0      0      0
#> 3       0      0      0      0      0      0      0      0      0      0      0
#> 4       0      0      0      0      0      0      0      0      0      0      0
#> 5       0      0      0      0      0      0      0      0      0      0      0
#> 6       0      0      0      0      0      0      0      0      0      0      0
#> 7       0      0      0      0      0      0      0      0      0      0      0
#> 8       0      0      0      0      0      0      0      0      0      0      0
#> 9       0      0      0      0      0      0      0      0      0      0      0
#> 10      0      0      0      0      0      0      0      0      0      0      0
#> 11      0      0      0      0      0      0      0      0      0      0      0
#> 12      1      0      0      0      0      0      0      0      0      0      0
#> 13      1      1      0      0      0      0      0      0      0      0      0
#> 14      1      1      1      0      0      0      0      0      0      0      0
#> 15      1      1      1      1      0      0      0      0      0      0      0
#> 16      1      1      1      1      1      0      0      0      0      0      0
#> 17      1      1      1      1      1      1      0      0      0      0      0
#> 18      1      1      1      1      1      1      1      0      0      0      0
#> 19      1      1      1      1      1      1      1      1      0      0      0
#> 20      1      1      1      1      1      1      1      1      1      0      0
#> 21      1      1      1      1      1      1      1      1      1      1      0
#> 22      1      1      1      1      1      1      1      1      1      1      1
#> 23      1      1      1      1      1      1      1      1      1      1      1
#> 24      1      1      1      1      1      1      1      1      1      1      1
#> 25      1      1      1      1      1      1      1      1      1      1      1
#> 26      1      1      1      1      1      1      1      1      1      1      1
#> 27      1      1      1      1      1      1      1      1      1      1      1
#> 28      1      1      1      1      1      1      1      1      1      1      1
#> 29      1      1      1      1      1      1      1      1      1      1      1
#> 30      1      1      1      1      1      1      1      1      1      1      1
#> 31      1      1      1      1      1      1      1      1      1      1      1
#> 32      1      1      1      1      1      1      1      1      1      1      1
#> 33      1      1      1      1      1      1      1      1      1      1      1
#> 34      1      1      1      1      1      1      1      1      1      1      1
#> 35      1      1      1      1      1      1      1      1      1      1      1
#> 36      1      1      1      1      1      1      1      1      1      1      1
#> 37      1      1      1      1      1      1      1      1      1      1      1
#> 38      1      1      1      1      1      1      1      1      1      1      1
#> 39      1      1      1      1      1      1      1      1      1      1      1
#> 40      1      1      1      1      1      1      1      1      1      1      1
#> 41      1      1      1      1      1      1      1      1      1      1      1
#>    cday22 cday23 cday24 cday25 cday26 cday27 cday28 cday29 cday30 cday31 cday32
#> 1       0      0      0      0      0      0      0      0      0      0      0
#> 2       0      0      0      0      0      0      0      0      0      0      0
#> 3       0      0      0      0      0      0      0      0      0      0      0
#> 4       0      0      0      0      0      0      0      0      0      0      0
#> 5       0      0      0      0      0      0      0      0      0      0      0
#> 6       0      0      0      0      0      0      0      0      0      0      0
#> 7       0      0      0      0      0      0      0      0      0      0      0
#> 8       0      0      0      0      0      0      0      0      0      0      0
#> 9       0      0      0      0      0      0      0      0      0      0      0
#> 10      0      0      0      0      0      0      0      0      0      0      0
#> 11      0      0      0      0      0      0      0      0      0      0      0
#> 12      0      0      0      0      0      0      0      0      0      0      0
#> 13      0      0      0      0      0      0      0      0      0      0      0
#> 14      0      0      0      0      0      0      0      0      0      0      0
#> 15      0      0      0      0      0      0      0      0      0      0      0
#> 16      0      0      0      0      0      0      0      0      0      0      0
#> 17      0      0      0      0      0      0      0      0      0      0      0
#> 18      0      0      0      0      0      0      0      0      0      0      0
#> 19      0      0      0      0      0      0      0      0      0      0      0
#> 20      0      0      0      0      0      0      0      0      0      0      0
#> 21      0      0      0      0      0      0      0      0      0      0      0
#> 22      0      0      0      0      0      0      0      0      0      0      0
#> 23      1      0      0      0      0      0      0      0      0      0      0
#> 24      1      1      0      0      0      0      0      0      0      0      0
#> 25      1      1      1      0      0      0      0      0      0      0      0
#> 26      1      1      1      1      0      0      0      0      0      0      0
#> 27      1      1      1      1      1      0      0      0      0      0      0
#> 28      1      1      1      1      1      1      0      0      0      0      0
#> 29      1      1      1      1      1      1      1      0      0      0      0
#> 30      1      1      1      1      1      1      1      1      0      0      0
#> 31      1      1      1      1      1      1      1      1      1      0      0
#> 32      1      1      1      1      1      1      1      1      1      1      0
#> 33      1      1      1      1      1      1      1      1      1      1      1
#> 34      1      1      1      1      1      1      1      1      1      1      1
#> 35      1      1      1      1      1      1      1      1      1      1      1
#> 36      1      1      1      1      1      1      1      1      1      1      1
#> 37      1      1      1      1      1      1      1      1      1      1      1
#> 38      1      1      1      1      1      1      1      1      1      1      1
#> 39      1      1      1      1      1      1      1      1      1      1      1
#> 40      1      1      1      1      1      1      1      1      1      1      1
#> 41      1      1      1      1      1      1      1      1      1      1      1
#>    cday33 cday34 cday35 cday36 cday37 cday38 cday39 cday40
#> 1       0      0      0      0      0      0      0      0
#> 2       0      0      0      0      0      0      0      0
#> 3       0      0      0      0      0      0      0      0
#> 4       0      0      0      0      0      0      0      0
#> 5       0      0      0      0      0      0      0      0
#> 6       0      0      0      0      0      0      0      0
#> 7       0      0      0      0      0      0      0      0
#> 8       0      0      0      0      0      0      0      0
#> 9       0      0      0      0      0      0      0      0
#> 10      0      0      0      0      0      0      0      0
#> 11      0      0      0      0      0      0      0      0
#> 12      0      0      0      0      0      0      0      0
#> 13      0      0      0      0      0      0      0      0
#> 14      0      0      0      0      0      0      0      0
#> 15      0      0      0      0      0      0      0      0
#> 16      0      0      0      0      0      0      0      0
#> 17      0      0      0      0      0      0      0      0
#> 18      0      0      0      0      0      0      0      0
#> 19      0      0      0      0      0      0      0      0
#> 20      0      0      0      0      0      0      0      0
#> 21      0      0      0      0      0      0      0      0
#> 22      0      0      0      0      0      0      0      0
#> 23      0      0      0      0      0      0      0      0
#> 24      0      0      0      0      0      0      0      0
#> 25      0      0      0      0      0      0      0      0
#> 26      0      0      0      0      0      0      0      0
#> 27      0      0      0      0      0      0      0      0
#> 28      0      0      0      0      0      0      0      0
#> 29      0      0      0      0      0      0      0      0
#> 30      0      0      0      0      0      0      0      0
#> 31      0      0      0      0      0      0      0      0
#> 32      0      0      0      0      0      0      0      0
#> 33      0      0      0      0      0      0      0      0
#> 34      1      0      0      0      0      0      0      0
#> 35      1      1      0      0      0      0      0      0
#> 36      1      1      1      0      0      0      0      0
#> 37      1      1      1      1      0      0      0      0
#> 38      1      1      1      1      1      0      0      0
#> 39      1      1      1      1      1      1      0      0
#> 40      1      1      1      1      1      1      1      0
#> 41      1      1      1      1      1      1      1      1
#> attr(,"assign")
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> [26] 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $data$exp_fnrow
#> [1] 41
#> 
#> $data$exp_findex
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41
#> 
#> $data$exp_fnindex
#> [1] 41
#> 
#> $data$exp_fncol
#> [1] 40
#> 
#> $data$exp_rdesign
#>    fixed day
#> 1      0   1
#> 2      0   1
#> 3      0   1
#> 4      0   1
#> 5      0   1
#> 6      0   1
#> 7      0   1
#> 8      0   1
#> 9      0   1
#> 10     0   1
#> 11     0   1
#> 12     0   1
#> 13     0   1
#> 14     0   1
#> 15     0   1
#> 16     0   1
#> 17     0   1
#> 18     0   1
#> 19     0   1
#> 20     0   1
#> 21     0   1
#> 22     0   1
#> 23     0   1
#> 24     0   1
#> 25     0   1
#> 26     0   1
#> 27     0   1
#> 28     0   1
#> 29     0   1
#> 30     0   1
#> 31     0   1
#> 32     0   1
#> 33     0   1
#> 34     0   1
#> 35     0   1
#> 36     0   1
#> 37     0   1
#> 38     0   1
#> 39     0   1
#> 40     0   1
#> attr(,"assign")
#> [1] 1 2
#> 
#> $data$exp_rncol
#> [1] 1
#> 
#> $data$exp_order
#> [1] 1
#> 
#> 
#> $priors
#>       variable                                             description
#> 1: exp_beta_sd Standard deviation of scaled pooled expectation effects
#> 2:    eobs_lsd      Standard deviation for expected final observations
#>             distribution mean sd
#> 1: Zero truncated normal    0  1
#> 2: Zero truncated normal    0  1
#> 
#> $inits
#> function (data, priors) 
#> {
#>     priors <- enw_priors_as_data_list(priors)
#>     fn <- function() {
#>         init <- list(exp_beta = numeric(0), exp_beta_sd = numeric(0), 
#>             leobs_init = array(purrr::map_dbl(data$latest_obs[1, 
#>                 ] + 1, ~rnorm(1, log(.), 1))), eobs_lsd = array(abs(rnorm(data$g, 
#>                 priors$eobs_lsd_p[1], priors$eobs_lsd_p[2]/10))), 
#>             leobs_resids = array(rnorm((data$t - 1) * data$g, 
#>                 0, 0.01), dim = c(data$t - 1, data$g)))
#>         if (data$exp_fncol > 0) {
#>             init$exp_beta <- rnorm(data$exp_fncol, 0, 0.01)
#>         }
#>         if (data$exp_rncol > 0) {
#>             init$exp_beta_sd <- abs(rnorm(data$exp_rncol, priors$exp_beta_sd_p[1], 
#>                 priors$exp_beta_sd_p[2]/10))
#>         }
#>         return(init)
#>     }
#>     return(fn)
#> }
#> <bytecode: 0x55692ae86a30>
#> <environment: 0x55692ae8c838>
#>