Filter observations to restrict the maximum reporting delay

Filter observations to restrict the maximum reporting delay

Usage

enw_filter_obs(obs, max_delay)

Arguments

obs

A data frame containing at least the following variables: reference date (index date of interest), report_date (report date for observations), confirm (cumulative observations by reference and report date), and group (as added by enw_assign_group()).

max_delay

Numeric defaults to 20. The maximum number of days to include in the delay distribution. Computation scales non-linearly with this setting so consider what maximum makes sense for your data carefully.

Value

A data frame filtered so that dates by report are less than or equal the reference date plus the maximum delay.

See also

Examples

obs <- enw_example("preprocessed")$obs[[1]]
enw_filter_obs(obs, max_delay = 2)
#>     reference_date group report_date location age_group confirm max_confirm
#>  1:     2021-07-13     1  2021-07-13       DE       00+      21          59
#>  2:     2021-07-13     1  2021-07-14       DE       00+      33          59
#>  3:     2021-07-14     1  2021-07-14       DE       00+      22          72
#>  4:     2021-07-14     1  2021-07-15       DE       00+      34          72
#>  5:     2021-07-15     1  2021-07-15       DE       00+      28          69
#>  6:     2021-07-15     1  2021-07-16       DE       00+      43          69
#>  7:     2021-07-16     1  2021-07-16       DE       00+      19          47
#>  8:     2021-07-16     1  2021-07-17       DE       00+      32          47
#>  9:     2021-07-17     1  2021-07-17       DE       00+      20          65
#> 10:     2021-07-17     1  2021-07-18       DE       00+      27          65
#> 11:     2021-07-18     1  2021-07-18       DE       00+       9          50
#> 12:     2021-07-18     1  2021-07-19       DE       00+      15          50
#> 13:     2021-07-19     1  2021-07-19       DE       00+       3          36
#> 14:     2021-07-19     1  2021-07-20       DE       00+      19          36
#> 15:     2021-07-20     1  2021-07-20       DE       00+      36          94
#> 16:     2021-07-20     1  2021-07-21       DE       00+      55          94
#> 17:     2021-07-21     1  2021-07-21       DE       00+      28          91
#> 18:     2021-07-21     1  2021-07-22       DE       00+      46          91
#> 19:     2021-07-22     1  2021-07-22       DE       00+      34          99
#> 20:     2021-07-22     1  2021-07-23       DE       00+      53          99
#> 21:     2021-07-23     1  2021-07-23       DE       00+      30          86
#> 22:     2021-07-23     1  2021-07-24       DE       00+      42          86
#> 23:     2021-07-24     1  2021-07-24       DE       00+      31          93
#> 24:     2021-07-24     1  2021-07-25       DE       00+      39          93
#> 25:     2021-07-25     1  2021-07-25       DE       00+       8          74
#> 26:     2021-07-25     1  2021-07-26       DE       00+      12          74
#> 27:     2021-07-26     1  2021-07-26       DE       00+       9          28
#> 28:     2021-07-26     1  2021-07-27       DE       00+      15          28
#> 29:     2021-07-27     1  2021-07-27       DE       00+      35          78
#> 30:     2021-07-27     1  2021-07-28       DE       00+      46          78
#> 31:     2021-07-28     1  2021-07-28       DE       00+      51         156
#> 32:     2021-07-28     1  2021-07-29       DE       00+      79         156
#> 33:     2021-07-29     1  2021-07-29       DE       00+      47         135
#> 34:     2021-07-29     1  2021-07-30       DE       00+      84         135
#> 35:     2021-07-30     1  2021-07-30       DE       00+      36         114
#> 36:     2021-07-30     1  2021-07-31       DE       00+      56         114
#> 37:     2021-07-31     1  2021-07-31       DE       00+      38         126
#> 38:     2021-07-31     1  2021-08-01       DE       00+      54         126
#> 39:     2021-08-01     1  2021-08-01       DE       00+       7          77
#> 40:     2021-08-01     1  2021-08-02       DE       00+      12          77
#> 41:     2021-08-02     1  2021-08-02       DE       00+      13          59
#> 42:     2021-08-02     1  2021-08-03       DE       00+      26          59
#> 43:     2021-08-03     1  2021-08-03       DE       00+      51         149
#> 44:     2021-08-03     1  2021-08-04       DE       00+      94         149
#> 45:     2021-08-04     1  2021-08-04       DE       00+      51         166
#> 46:     2021-08-04     1  2021-08-05       DE       00+      94         166
#> 47:     2021-08-05     1  2021-08-05       DE       00+      45         133
#> 48:     2021-08-05     1  2021-08-06       DE       00+      66         133
#> 49:     2021-08-06     1  2021-08-06       DE       00+      47         137
#> 50:     2021-08-06     1  2021-08-07       DE       00+      78         137
#> 51:     2021-08-07     1  2021-08-07       DE       00+      40         139
#> 52:     2021-08-07     1  2021-08-08       DE       00+      55         139
#> 53:     2021-08-08     1  2021-08-08       DE       00+      13          97
#> 54:     2021-08-08     1  2021-08-09       DE       00+      27          97
#> 55:     2021-08-09     1  2021-08-09       DE       00+      14          58
#> 56:     2021-08-09     1  2021-08-10       DE       00+      37          58
#> 57:     2021-08-10     1  2021-08-10       DE       00+      78         175
#> 58:     2021-08-10     1  2021-08-11       DE       00+     121         175
#> 59:     2021-08-11     1  2021-08-11       DE       00+      80         233
#> 60:     2021-08-11     1  2021-08-12       DE       00+     133         233
#> 61:     2021-08-12     1  2021-08-12       DE       00+      89         237
#> 62:     2021-08-12     1  2021-08-13       DE       00+     137         237
#> 63:     2021-08-13     1  2021-08-13       DE       00+      86         204
#> 64:     2021-08-13     1  2021-08-14       DE       00+     130         204
#> 65:     2021-08-14     1  2021-08-14       DE       00+      79         189
#> 66:     2021-08-14     1  2021-08-15       DE       00+     115         189
#> 67:     2021-08-15     1  2021-08-15       DE       00+      22         125
#> 68:     2021-08-15     1  2021-08-16       DE       00+      46         125
#> 69:     2021-08-16     1  2021-08-16       DE       00+      23          98
#> 70:     2021-08-16     1  2021-08-17       DE       00+      55          98
#> 71:     2021-08-17     1  2021-08-17       DE       00+      96         242
#> 72:     2021-08-17     1  2021-08-18       DE       00+     181         242
#> 73:     2021-08-18     1  2021-08-18       DE       00+      92         223
#> 74:     2021-08-18     1  2021-08-19       DE       00+     178         223
#> 75:     2021-08-19     1  2021-08-19       DE       00+      84         202
#> 76:     2021-08-19     1  2021-08-20       DE       00+     171         202
#> 77:     2021-08-20     1  2021-08-20       DE       00+      98         171
#> 78:     2021-08-20     1  2021-08-21       DE       00+     159         171
#> 79:     2021-08-21     1  2021-08-21       DE       00+      69         112
#> 80:     2021-08-21     1  2021-08-22       DE       00+     112         112
#> 81:     2021-08-22     1  2021-08-22       DE       00+      45          45
#>     reference_date group report_date location age_group confirm max_confirm
#>     cum_prop_reported delay
#>  1:        0.35593220     0
#>  2:        0.55932203     1
#>  3:        0.30555556     0
#>  4:        0.47222222     1
#>  5:        0.40579710     0
#>  6:        0.62318841     1
#>  7:        0.40425532     0
#>  8:        0.68085106     1
#>  9:        0.30769231     0
#> 10:        0.41538462     1
#> 11:        0.18000000     0
#> 12:        0.30000000     1
#> 13:        0.08333333     0
#> 14:        0.52777778     1
#> 15:        0.38297872     0
#> 16:        0.58510638     1
#> 17:        0.30769231     0
#> 18:        0.50549451     1
#> 19:        0.34343434     0
#> 20:        0.53535354     1
#> 21:        0.34883721     0
#> 22:        0.48837209     1
#> 23:        0.33333333     0
#> 24:        0.41935484     1
#> 25:        0.10810811     0
#> 26:        0.16216216     1
#> 27:        0.32142857     0
#> 28:        0.53571429     1
#> 29:        0.44871795     0
#> 30:        0.58974359     1
#> 31:        0.32692308     0
#> 32:        0.50641026     1
#> 33:        0.34814815     0
#> 34:        0.62222222     1
#> 35:        0.31578947     0
#> 36:        0.49122807     1
#> 37:        0.30158730     0
#> 38:        0.42857143     1
#> 39:        0.09090909     0
#> 40:        0.15584416     1
#> 41:        0.22033898     0
#> 42:        0.44067797     1
#> 43:        0.34228188     0
#> 44:        0.63087248     1
#> 45:        0.30722892     0
#> 46:        0.56626506     1
#> 47:        0.33834586     0
#> 48:        0.49624060     1
#> 49:        0.34306569     0
#> 50:        0.56934307     1
#> 51:        0.28776978     0
#> 52:        0.39568345     1
#> 53:        0.13402062     0
#> 54:        0.27835052     1
#> 55:        0.24137931     0
#> 56:        0.63793103     1
#> 57:        0.44571429     0
#> 58:        0.69142857     1
#> 59:        0.34334764     0
#> 60:        0.57081545     1
#> 61:        0.37552743     0
#> 62:        0.57805907     1
#> 63:        0.42156863     0
#> 64:        0.63725490     1
#> 65:        0.41798942     0
#> 66:        0.60846561     1
#> 67:        0.17600000     0
#> 68:        0.36800000     1
#> 69:        0.23469388     0
#> 70:        0.56122449     1
#> 71:        0.39669421     0
#> 72:        0.74793388     1
#> 73:        0.41255605     0
#> 74:        0.79820628     1
#> 75:        0.41584158     0
#> 76:        0.84653465     1
#> 77:        0.57309942     0
#> 78:        0.92982456     1
#> 79:        0.61607143     0
#> 80:        1.00000000     1
#> 81:        1.00000000     0
#>     cum_prop_reported delay