Retrieve Stan cache location

Retrieves the user set cache location for Stan models. This path can be set through the enw_cache_location function call. If no environmental variable is available the output from tempdir() will be returned.

Usage

enw_get_cache()

Value

A string representing the file path for the cache location

See also