Missing reference data model module

Missing reference data model module

Usage

enw_missing(formula = ~1, data)

Arguments

formula

A formula (as implemented in enw_formula()) describing the missing data proportion on the logit scale by reference date. This can use features defined by reference date as defined in metareference as produced by enw_preprocess_data().

data

Output from enw_preprocess_data().

Value

A list containing the supplied formulas, data passed into a list describing the models, a data.frame describing the priors used, and a function that takes the output data and priors and returns a function that can be used to sample from a tightened version of the prior distribution.

See also

Examples

enw_missing(data = enw_example("preprocessed"))
#> $formula
#> [1] "~1"
#> 
#> $data
#> $data$miss_fdesign
#>    (Intercept)
#> 1            1
#> 2            1
#> 3            1
#> 4            1
#> 5            1
#> 6            1
#> 7            1
#> 8            1
#> 9            1
#> 10           1
#> 11           1
#> 12           1
#> 13           1
#> 14           1
#> 15           1
#> 16           1
#> 17           1
#> 18           1
#> 19           1
#> 20           1
#> 21           1
#> 22           1
#> 23           1
#> 24           1
#> 25           1
#> 26           1
#> 27           1
#> 28           1
#> 29           1
#> 30           1
#> 31           1
#> 32           1
#> 33           1
#> 34           1
#> 35           1
#> 36           1
#> 37           1
#> 38           1
#> 39           1
#> 40           1
#> 41           1
#> attr(,"assign")
#> [1] 0
#> 
#> $data$miss_fnrow
#> [1] 41
#> 
#> $data$miss_findex
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41
#> 
#> $data$miss_fnindex
#> [1] 41
#> 
#> $data$miss_fncol
#> [1] 1
#> 
#> $data$miss_rdesign
#>      (Intercept)
#> attr(,"assign")
#> [1] 0
#> 
#> $data$miss_rncol
#> [1] 0
#> 
#> $data$missing_ref
#>      2021-07-13 2021-07-14 2021-07-15 2021-07-16 2021-07-17 2021-07-18
#> [1,]          0          0          0          0          0          0
#>      2021-07-19 2021-07-20 2021-07-21 2021-07-22 2021-07-23 2021-07-24
#> [1,]          0          0          0          0          0          0
#>      2021-07-25 2021-07-26 2021-07-27 2021-07-28 2021-07-29 2021-07-30
#> [1,]          0          0          0          0          0          0
#>      2021-07-31 2021-08-01 2021-08-02 2021-08-03 2021-08-04 2021-08-05
#> [1,]          0          0          0          0          0          0
#>      2021-08-06 2021-08-07 2021-08-08 2021-08-09 2021-08-10 2021-08-11
#> [1,]          0          0          0          0          0          0
#>      2021-08-12 2021-08-13 2021-08-14 2021-08-15 2021-08-16 2021-08-17
#> [1,]          0          0          0          0          0          0
#>      2021-08-18 2021-08-19 2021-08-20 2021-08-21 2021-08-22
#> [1,]          0          0          0          0          0
#> 
#> $data$model_missing
#> [1] 1
#> 
#> 
#> $priors
#>        variable
#> 1: miss_beta_sd
#>                                                                          description
#> 1: Standard deviation of scaled pooled logit missing\n        reference date effects
#>             distribution mean sd
#> 1: Zero truncated normal    0  1
#> 
#> $inits
#> function (data, priors) 
#> {
#>     priors <- enw_priors_as_data_list(priors)
#>     fn <- function() {
#>         init <- list(miss_beta = numeric(0), miss_beta_sd = numeric(0))
#>         if (data$miss_fncol > 0) {
#>             init$miss_beta <- rnorm(data$miss_fncol, 0, 0.01)
#>         }
#>         if (data$miss_rncol > 0) {
#>             init$miss_beta_sd <- abs(rnorm(data$miss_rncol, priors$miss_beta_sd_p[1], 
#>                 priors$miss_beta_sd_p[2]/10))
#>         }
#>         return(init)
#>     }
#>     return(fn)
#> }
#> <bytecode: 0x55e6aa8fed98>
#> <environment: 0x55e6aa903888>
#>