Calculate incidence of new reports from cumulative reports

Calculate incidence of new reports from cumulative reports

Usage

enw_new_reports(obs, set_negatives_to_zero = TRUE)

Arguments

obs

A data frame containing at least the following variables: reference date (index date of interest), report_date (report date for observations), confirm (cumulative observations by reference and report date), and .group (as added by enw_assign_group()).

set_negatives_to_zero

Logical, defaults to TRUE. Should negative counts (for calculated incidence of observations) be set to zero. Currently downstream modelling does not support negative counts and so setting must be TRUE if intending to use epinowcast().

Value

The input data frame with a new variable new_confirm. If max_confirm was present in the data frame then the proportion reported on each day (prop_reported) is also added.

See also

Examples

# Default reconstruct incidence
dt <- enw_assign_group(
  germany_covid19_hosp[location == "DE"],
  by = "age_group"
)
enw_new_reports(dt)
#>        .group age_group reference_date location confirm report_date new_confirm
#>     1:      2     00-04     2021-04-06       DE       4  2021-04-06           4
#>     2:      2     00-04     2021-04-06       DE       6  2021-04-07           2
#>     3:      2     00-04     2021-04-06       DE       6  2021-04-08           0
#>     4:      2     00-04     2021-04-06       DE       6  2021-04-09           0
#>     5:      2     00-04     2021-04-06       DE       6  2021-04-10           0
#>    ---                                                                         
#> 90401:      7       80+     2021-10-18       DE      42  2021-10-19          25
#> 90402:      7       80+     2021-10-18       DE      50  2021-10-20           8
#> 90403:      7       80+     2021-10-19       DE      70  2021-10-19          70
#> 90404:      7       80+     2021-10-19       DE     128  2021-10-20          58
#> 90405:      7       80+     2021-10-20       DE      71  2021-10-20          71
#>        delay
#>     1:     0
#>     2:     1
#>     3:     2
#>     4:     3
#>     5:     4
#>    ---      
#> 90401:     1
#> 90402:     2
#> 90403:     0
#> 90404:     1
#> 90405:     0

# Make use of maximum reported to calculate empirical daily reporting
dt <- enw_add_max_reported(dt)
enw_new_reports(dt)
#>        .group reference_date max_confirm age_group location confirm report_date
#>     1:      2     2021-04-06          11     00-04       DE       4  2021-04-06
#>     2:      2     2021-04-06          11     00-04       DE       6  2021-04-07
#>     3:      2     2021-04-06          11     00-04       DE       6  2021-04-08
#>     4:      2     2021-04-06          11     00-04       DE       6  2021-04-09
#>     5:      2     2021-04-06          11     00-04       DE       6  2021-04-10
#>    ---                                                                         
#> 90401:      7     2021-10-18          50       80+       DE      42  2021-10-19
#> 90402:      7     2021-10-18          50       80+       DE      50  2021-10-20
#> 90403:      7     2021-10-19         128       80+       DE      70  2021-10-19
#> 90404:      7     2021-10-19         128       80+       DE     128  2021-10-20
#> 90405:      7     2021-10-20          71       80+       DE      71  2021-10-20
#>        cum_prop_reported new_confirm delay prop_reported
#>     1:         0.3636364           4     0     0.3636364
#>     2:         0.5454545           2     1     0.1818182
#>     3:         0.5454545           0     2     0.0000000
#>     4:         0.5454545           0     3     0.0000000
#>     5:         0.5454545           0     4     0.0000000
#>    ---                                                  
#> 90401:         0.8400000          25     1     0.5000000
#> 90402:         1.0000000           8     2     0.1600000
#> 90403:         0.5468750          70     0     0.5468750
#> 90404:         1.0000000          58     1     0.4531250
#> 90405:         1.0000000          71     0     1.0000000