Format model options for use with stan

Format model options for use with stan

Usage

enw_opts_as_data_list(
  dist = "lognormal",
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE
)

Arguments

dist

Character string indicating the type of distribution to use for reference date effects. The default is to use a lognormal but other options available include: gamma distributed ("gamma").

nowcast

Logical, defaults to TRUE. Should a nowcast be made using posterior predictions of the unobserved future reported notifications.

pp

Logical, defaults to FALSE. Should posterior predictions be made for observed data. Useful for evaluating the performance of the model.

likelihood

Logical, defaults to TRUE. Should the likelihood be included in the model

debug

Logical, defaults to FALSE. Should within model debug information be returned.

output_loglik

Logical, defaults to FALSE. Should the log-likelihood be output. Disabling this will speed up fitting if evaluating the model fit is not required.

Value

A list as required by stan.

See also