Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean       median        sd       mad         q5
#>            <char>       <num>        <num>     <num>     <num>      <num>
#>  1:  expr_beta[1] -0.32092558 -0.302995770 0.4974060 0.5003899 -1.1751767
#>  2:  expr_beta[2] -0.74265782 -0.753756680 0.4890380 0.4876114 -1.5490582
#>  3:  expr_beta[3]  0.49075793  0.485267860 0.4909068 0.4815642 -0.3189382
#>  4:  expr_beta[4] -0.56077281 -0.557551560 0.4827831 0.5089833 -1.3395942
#>  5:  expr_beta[5] -0.38091095 -0.394056945 0.5247593 0.5420244 -1.2090794
#>  6:  expr_beta[6]  1.79486180  1.789772550 0.5316671 0.5501319  0.9399786
#>  7:  expr_beta[7]  0.24656509  0.238040255 0.4489022 0.4530981 -0.4740480
#>  8:  expr_beta[8]  0.18508535  0.167452445 0.4515854 0.4299909 -0.5038096
#>  9:  expr_beta[9] -0.31261059 -0.317163115 0.4463138 0.4614627 -1.0475074
#> 10: expr_beta[10] -0.01486819 -0.010347709 0.4220814 0.3813205 -0.7009876
#> 11: expr_beta[11] -0.61675220 -0.587198200 0.4589874 0.4705479 -1.4044386
#> 12: expr_beta[12] -1.42790779 -1.421195000 0.5044782 0.4973346 -2.2509056
#> 13: expr_beta[13]  1.62719872  1.622999750 0.5315766 0.5111194  0.7765763
#> 14: expr_beta[14]  1.79896425  1.763649550 0.4751534 0.4575447  1.0507029
#> 15: expr_beta[15] -0.25842477 -0.284254145 0.4146232 0.4247106 -0.9031172
#> 16: expr_beta[16] -0.21802551 -0.214714005 0.4273798 0.4200640 -0.9205409
#> 17: expr_beta[17] -0.04914702 -0.034840848 0.4204239 0.4048612 -0.7464496
#> 18: expr_beta[18] -0.79394694 -0.783196240 0.4063712 0.3807600 -1.4677365
#> 19: expr_beta[19] -0.51678077 -0.508444485 0.4687769 0.4601203 -1.2958871
#> 20: expr_beta[20]  1.84954623  1.852946300 0.4604705 0.4493234  1.1068935
#> 21: expr_beta[21]  0.43830304  0.429870680 0.3943223 0.4093633 -0.1897291
#> 22: expr_beta[22] -0.37514479 -0.383338120 0.3877576 0.3791090 -1.0321690
#> 23: expr_beta[23] -0.16177802 -0.161673335 0.3808587 0.3499998 -0.8074587
#> 24: expr_beta[24] -0.00725609 -0.007511365 0.3737199 0.3847101 -0.6169439
#> 25: expr_beta[25] -0.89265300 -0.872465965 0.4432475 0.4634984 -1.6111248
#> 26: expr_beta[26] -0.86827671 -0.867037245 0.4458622 0.4330594 -1.5976094
#> 27: expr_beta[27]  2.19781269  2.186659050 0.4764435 0.4831766  1.4849359
#> 28: expr_beta[28]  1.41009094  1.376479950 0.4308574 0.4191553  0.7662879
#> 29: expr_beta[29]  0.04979937  0.047622414 0.3741277 0.3795555 -0.5551349
#> 30: expr_beta[30] -0.52065898 -0.507807690 0.3997635 0.3944245 -1.1884466
#> 31: expr_beta[31] -0.13778692 -0.122675075 0.4232148 0.4005677 -0.8108834
#> 32: expr_beta[32] -0.74039628 -0.730091135 0.4162849 0.4148147 -1.4607499
#> 33: expr_beta[33] -0.38393827 -0.382522510 0.4451681 0.4618817 -1.1170769
#> 34: expr_beta[34]  1.78265989  1.769618250 0.4557591 0.4744966  1.0625023
#> 35: expr_beta[35]  0.15483523  0.148572520 0.4288613 0.4337991 -0.5316621
#> 36: expr_beta[36]  0.03436145  0.019842342 0.4642230 0.4439917 -0.6827261
#> 37: expr_beta[37]  0.01997313  0.039526770 0.4972613 0.5074665 -0.7911249
#> 38: expr_beta[38]  0.15482011  0.147784385 0.5264489 0.4984904 -0.6974510
#> 39: expr_beta[39] -0.04957734 -0.067841933 0.6669359 0.6651630 -1.0722428
#>          variable        mean       median        sd       mad         q5
#>            <char>       <num>        <num>     <num>     <num>      <num>
#>             q20          q80         q95      rhat  ess_bulk ess_tail
#>           <num>        <num>       <num>     <num>     <num>    <num>
#>  1: -0.73639250  0.090287902  0.48502811 1.0018321  878.4473 706.1199
#>  2: -1.13676816 -0.344636374  0.06348241 1.0043698  741.5364 691.1377
#>  3:  0.10044581  0.907438526  1.24770518 0.9982835  733.0923 665.0849
#>  4: -0.97341661 -0.150753668  0.22795212 1.0010690  860.5830 583.4544
#>  5: -0.81852404  0.058032422  0.46030230 1.0009878 1090.5546 797.8204
#>  6:  1.35440418  2.267011520  2.68265889 1.0015818  841.2275 735.3280
#>  7: -0.12999056  0.624345948  0.99579804 1.0009268  772.3739 798.7478
#>  8: -0.18120249  0.550547224  0.95366276 1.0045412  870.8106 750.1877
#>  9: -0.68543293  0.072643654  0.41431620 1.0028455  747.4746 684.8390
#> 10: -0.34238558  0.297214288  0.65848327 1.0009370  768.7739 689.4648
#> 11: -1.00488118 -0.217497066  0.08972751 1.0001749  982.4667 699.1044
#> 12: -1.85951502 -1.012967420 -0.60127481 1.0015900 1051.7232 689.9978
#> 13:  1.17686242  2.062967660  2.52119996 1.0046222  821.0012 796.5544
#> 14:  1.41514588  2.175821040  2.58542339 1.0045549  875.3790 794.9830
#> 15: -0.61021848  0.111923740  0.40275218 1.0037413 1189.3795 703.3240
#> 16: -0.56055053  0.128765930  0.43360164 1.0032162  975.0190 582.6757
#> 17: -0.39980843  0.305642726  0.62747661 1.0026723 1022.9964 605.9811
#> 18: -1.12668856 -0.476163362 -0.15166687 1.0084161  828.9399 724.5456
#> 19: -0.91194215 -0.132182392  0.26215119 1.0049323  849.5924 556.6769
#> 20:  1.46947150  2.216933900  2.58423808 1.0018926  821.4537 766.5134
#> 21:  0.09463508  0.781410200  1.08987256 1.0017265  653.4625 680.4379
#> 22: -0.69134303 -0.039566003  0.26008569 1.0020710  848.7761 673.0280
#> 23: -0.46726042  0.151998248  0.45985306 0.9992550  993.7858 703.3272
#> 24: -0.31956196  0.315795620  0.60201213 1.0033079  911.5141 755.1578
#> 25: -1.28310830 -0.513215950 -0.15434040 1.0021771  910.8511 824.6721
#> 26: -1.22998792 -0.512188760 -0.10110044 1.0029128  922.3267 746.6852
#> 27:  1.76485550  2.582021120  3.03506279 1.0074510 1044.4111 796.6966
#> 28:  1.04795938  1.760060540  2.20097286 1.0047873  768.8989 663.5001
#> 29: -0.26491786  0.353596022  0.66010952 1.0042307 1133.0614 725.7444
#> 30: -0.85210609 -0.197091604  0.15555015 1.0020936  914.4906 657.1161
#> 31: -0.48909516  0.190286410  0.55627997 0.9990510 1190.4347 885.7016
#> 32: -1.07899738 -0.384859026 -0.11416102 1.0000829  816.3915 574.7499
#> 33: -0.75352633  0.003764149  0.33175456 0.9988494  778.8376 750.3215
#> 34:  1.39398302  2.176802840  2.53932863 0.9994602  830.5158 741.4103
#> 35: -0.21077467  0.517067238  0.86714956 1.0007156  812.8916 630.8140
#> 36: -0.34905981  0.400178566  0.86290725 1.0043016  850.4438 573.1620
#> 37: -0.41276909  0.449751784  0.82059578 1.0024291 1186.3154 644.8023
#> 38: -0.27921994  0.572035036  1.08701899 0.9986116 1157.5990 765.7358
#> 39: -0.60852177  0.498290006  1.04315013 1.0138393 1072.8518 600.3256
#>             q20          q80         q95      rhat  ess_bulk ess_tail
#>           <num>        <num>       <num>     <num>     <num>    <num>