Construct a lookup of references dates by report

Construct a lookup of references dates by report

enw_reference_by_report(
  missing_reference,
  reps_with_complete_refs,
  metareference,
  max_delay
)

Arguments

missing_reference

missing_reference data.frame output from enw_preprocess_data().

reps_with_complete_refs

A data.frame of report dates with complete (i.e fully reported) reference dates as produced using enw_reps_with_complete_refs().

metareference

metareference data.frame output from enw_preprocess_data().

max_delay

The maximum number of days to model in the delay distribution. Must be an integer greater than or equal to 1. Observations with delays larger then the maximum delay will be dropped. If the specified maximum delay is too short, nowcasts can be biased as important parts of the true delay distribution are cut off. At the same time, computational cost scales non-linearly with this setting, so you want the maximum delay to be as long as necessary, but not much longer. Consider what delays are realistic for your application, and when in doubt, check if increasing the maximum delay noticeably changes the delay distribution or nowcasts as estimated by epinowcast. If it does, your maximum delay may still be too short. Note that delays are zero indexed and so include the reference date and max_delay - 1 other days (i.e. a max_delay of 1 corresponds to no delay). You can use check_max_delay() to check the coverage of a delay distribution for different maximum delays.

Value

A wide data.frame with each row being a complete report date and' the columns being the observation index for each reporting delay

See also