epinowcast: Flexible Hierarchical Nowcasting

Tools to enable flexible and efficient hierarchical nowcasting of right-truncated epidemiological time-series using a semi-mechanistic Bayesian model with support for a range of reporting and generative processes. Nowcasting, in this context, is gaining situational awareness using currently available observations and the reporting patterns of historical observations. This can be useful when tracking the spread of infectious disease in real-time: without nowcasting, changes in trends can be obfuscated by partial reporting or their detection may be delayed due to the use of simpler methods like truncation. While the package has been designed with epidemiological applications in mind, it could be applied to any set of right-truncated time-series count data.

See also

Author

Maintainer: Sam Abbott contact@samabbott.co.uk (ORCID)

Authors:

Adrian Lison adrian.lison@bsse.ethz.ch (ORCID)

Sebastian Funk sebastian.funk@lshtm.ac.uk

Carl Pearson carl.ab.pearson@gmail.com (ORCID)

Hugo Gruson hugo.gruson@normalesup.org (ORCID)

Felix Guenther felixguenther1@gmx.de (ORCID)

Michael DeWitt me.dewitt.jr@gmail.com (ORCID)

Other contributors:

Hannah Choi hannah.choi1@lshtm.ac.uk [contributor]

Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [contributor]

Joel Hellewell joel@ebi.ac.uk (ORCID) [contributor]

Luis Rivas luisnicolasrivas@gmail.com [contributor]

Sang Woo Park swp2@princeton.edu (ORCID) [contributor]

Nathan McIntosh natemcintosh@gmail.com [contributor]

James Mba Azam james.azam@lshtm.ac.uk (ORCID) [contributor]

Kath Sherratt katharine.sherratt@lshtm.ac.uk (ORCID) [contributor]

Nikos Bosse nikos.bosse@lshtm.ac.uk (ORCID) [contributor]