Remove profiling statements from a character vector representing stan code

Remove profiling statements from a character vector representing stan code

Usage

remove_profiling(s)

Arguments

s

Character vector representing stan code

Value

A character vector of the stan code without profiling statements

See also