Add latest observations to nowcast output

Add the latest observations to the nowcast output. This is useful for plotting the nowcast against the latest observations.

Usage

enw_add_latest_obs_to_nowcast(nowcast, obs)

Arguments

nowcast

A data.frame of nowcast output from enw_nowcast_summary().

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

Value

A data.frame of nowcast output with the latest observations added.

See also

Examples

fit <- enw_example("nowcast")
obs <- enw_example("obs")
nowcast <- summary(fit, type = "nowcast")
enw_add_latest_obs_to_nowcast(nowcast, obs)
#>     reference_date .group latest_confirm confirm report_date max_confirm
#>  1:     2021-08-03      1            156     149  2021-08-22         149
#>  2:     2021-08-04      1            183     166  2021-08-22         166
#>  3:     2021-08-05      1            147     133  2021-08-22         133
#>  4:     2021-08-06      1            155     137  2021-08-22         137
#>  5:     2021-08-07      1            159     139  2021-08-22         139
#>  6:     2021-08-08      1            119      97  2021-08-22          97
#>  7:     2021-08-09      1             65      58  2021-08-22          58
#>  8:     2021-08-10      1            204     175  2021-08-22         175
#>  9:     2021-08-11      1            275     233  2021-08-22         233
#> 10:     2021-08-12      1            273     237  2021-08-22         237
#> 11:     2021-08-13      1            270     204  2021-08-22         204
#> 12:     2021-08-14      1            262     189  2021-08-22         189
#> 13:     2021-08-15      1            192     125  2021-08-22         125
#> 14:     2021-08-16      1            140      98  2021-08-22          98
#> 15:     2021-08-17      1            323     242  2021-08-22         242
#> 16:     2021-08-18      1            409     223  2021-08-22         223
#> 17:     2021-08-19      1            370     202  2021-08-22         202
#> 18:     2021-08-20      1            361     171  2021-08-22         171
#> 19:     2021-08-21      1            339     112  2021-08-22         112
#> 20:     2021-08-22      1            258      45  2021-08-22          45
#>     location age_group cum_prop_reported delay prop_reported    mean median
#>  1:       DE       00+                 1    19   0.000000000 149.000    149
#>  2:       DE       00+                 1    18   0.000000000 167.510    167
#>  3:       DE       00+                 1    17   0.000000000 135.699    135
#>  4:       DE       00+                 1    16   0.000000000 141.110    141
#>  5:       DE       00+                 1    15   0.007194245 145.919    146
#>  6:       DE       00+                 1    14   0.000000000 103.761    103
#>  7:       DE       00+                 1    13   0.000000000  62.843     63
#>  8:       DE       00+                 1    12   0.000000000 185.143    185
#>  9:       DE       00+                 1    11   0.000000000 256.242    256
#> 10:       DE       00+                 1    10   0.004219409 267.754    267
#> 11:       DE       00+                 1     9   0.000000000 236.311    235
#> 12:       DE       00+                 1     8   0.015873016 230.777    230
#> 13:       DE       00+                 1     7   0.040000000 165.677    164
#> 14:       DE       00+                 1     6   0.010204082 129.764    129
#> 15:       DE       00+                 1     5   0.012396694 293.502    292
#> 16:       DE       00+                 1     4   0.017937220 293.818    292
#> 17:       DE       00+                 1     3   0.019801980 292.588    291
#> 18:       DE       00+                 1     2   0.070175439 293.411    291
#> 19:       DE       00+                 1     1   0.383928571 310.284    305
#> 20:       DE       00+                 1     0   1.000000000 381.715    366
#>             sd     mad     q5   q20 q35 q50 q65   q80    q95      rhat
#>  1:   0.000000  0.0000 149.00 149.0 149 149 149 149.0 149.00        NA
#>  2:   1.477187  1.4826 166.00 166.0 167 167 168 169.0 170.00 1.0021059
#>  3:   1.821460  1.4826 133.00 134.0 135 135 136 137.0 139.00 1.0026309
#>  4:   2.278397  2.9652 138.00 139.0 140 141 142 143.0 145.05 0.9985305
#>  5:   3.102590  2.9652 141.95 143.0 144 146 147 148.0 152.00 1.0024283
#>  6:   3.001814  2.9652  99.00 101.0 102 103 105 106.0 109.00 1.0002988
#>  7:   2.447108  2.9652  59.00  61.0  62  63  64  65.0  67.00 1.0050496
#>  8:   3.892776  4.4478 179.00 182.0 184 185 186 188.0 192.00 0.9987252
#>  9:   6.614922  5.9304 246.00 251.0 253 256 259 262.0 267.00 0.9997534
#> 10:   7.815406  7.4130 256.95 261.0 264 267 270 274.0 282.00 1.0015893
#> 11:   8.261994  7.4130 225.00 229.0 232 235 239 243.0 251.00 1.0009977
#> 12:   9.846330  8.8956 216.00 223.0 227 230 234 238.0 249.00 0.9981689
#> 13:  10.046871 10.3782 150.00 157.0 161 164 169 174.0 184.00 0.9992937
#> 14:   8.397549  7.4130 118.00 123.0 126 129 132 136.0 145.00 0.9992580
#> 15:  12.324443 11.8608 275.95 283.0 288 292 297 303.0 316.00 0.9997709
#> 16:  16.721797 14.8260 269.00 280.0 287 292 298 306.0 323.00 0.9987247
#> 17:  21.275500 22.2390 260.95 274.0 282 291 299 310.0 333.00 0.9996703
#> 18:  28.361531 26.6868 253.00 269.8 280 291 301 313.2 342.00 1.0014503
#> 19:  51.245443 46.7019 240.00 267.0 288 305 322 348.2 403.00 0.9996016
#> 20: 105.802296 94.8864 245.95 291.8 331 366 403 459.0 580.05 0.9991110
#>      ess_bulk ess_tail
#>  1:        NA       NA
#>  2:  916.5251 961.2656
#>  3:  865.5792 980.4149
#>  4:  964.6971 916.0647
#>  5:  982.1855 878.7477
#>  6: 1079.3848 873.5648
#>  7: 1015.4992 831.1428
#>  8: 1035.5304 854.2485
#>  9: 1020.3568 936.2525
#> 10: 1073.7290 947.6912
#> 11: 1024.8633 849.3204
#> 12: 1305.3677 998.3510
#> 13: 1114.1136 874.1649
#> 14: 1020.4703 799.7645
#> 15: 1097.9740 908.7449
#> 16: 1025.9249 971.7279
#> 17: 1243.7915 969.4998
#> 18: 1167.6104 910.7979
#> 19: 1107.4854 874.2575
#> 20: 1298.2575 864.7724