Add latest observations to nowcast output

Add the latest observations to the nowcast output. This is useful for plotting the nowcast against the latest observations.

Usage

enw_add_latest_obs_to_nowcast(nowcast, obs)

Arguments

nowcast

A data.frame of nowcast output from enw_nowcast_summary().

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

Value

A data.frame of nowcast output with the latest observations added.

See also

Examples

fit <- enw_example("nowcast")
obs <- enw_example("obs")
nowcast <- summary(fit, type = "nowcast")
enw_add_latest_obs_to_nowcast(nowcast, obs)
#> Key: <reference_date, .group>
#>     reference_date .group latest_confirm confirm report_date max_confirm
#>             <IDat>  <num>          <int>   <int>      <IDat>       <int>
#>  1:     2021-08-03      1            156     149  2021-08-22         149
#>  2:     2021-08-04      1            183     166  2021-08-22         166
#>  3:     2021-08-05      1            147     133  2021-08-22         133
#>  4:     2021-08-06      1            155     137  2021-08-22         137
#>  5:     2021-08-07      1            159     139  2021-08-22         139
#>  6:     2021-08-08      1            119      97  2021-08-22          97
#>  7:     2021-08-09      1             65      58  2021-08-22          58
#>  8:     2021-08-10      1            204     175  2021-08-22         175
#>  9:     2021-08-11      1            275     233  2021-08-22         233
#> 10:     2021-08-12      1            273     237  2021-08-22         237
#> 11:     2021-08-13      1            270     204  2021-08-22         204
#> 12:     2021-08-14      1            262     189  2021-08-22         189
#> 13:     2021-08-15      1            192     125  2021-08-22         125
#> 14:     2021-08-16      1            140      98  2021-08-22          98
#> 15:     2021-08-17      1            323     242  2021-08-22         242
#> 16:     2021-08-18      1            409     223  2021-08-22         223
#> 17:     2021-08-19      1            370     202  2021-08-22         202
#> 18:     2021-08-20      1            361     171  2021-08-22         171
#> 19:     2021-08-21      1            339     112  2021-08-22         112
#> 20:     2021-08-22      1            258      45  2021-08-22          45
#>     reference_date .group latest_confirm confirm report_date max_confirm
#>     location age_group cum_prop_reported delay prop_reported    mean median
#>       <fctr>    <fctr>             <num> <num>         <num>   <num>  <num>
#>  1:       DE       00+                 1    19   0.000000000 149.000  149.0
#>  2:       DE       00+                 1    18   0.000000000 167.424  167.0
#>  3:       DE       00+                 1    17   0.000000000 135.760  135.5
#>  4:       DE       00+                 1    16   0.000000000 141.290  141.0
#>  5:       DE       00+                 1    15   0.007194245 145.842  145.5
#>  6:       DE       00+                 1    14   0.000000000 103.852  104.0
#>  7:       DE       00+                 1    13   0.000000000  62.653   62.0
#>  8:       DE       00+                 1    12   0.000000000 185.247  185.0
#>  9:       DE       00+                 1    11   0.000000000 256.324  256.0
#> 10:       DE       00+                 1    10   0.004219409 267.501  267.0
#> 11:       DE       00+                 1     9   0.000000000 236.484  236.0
#> 12:       DE       00+                 1     8   0.015873016 231.475  231.0
#> 13:       DE       00+                 1     7   0.040000000 165.095  164.0
#> 14:       DE       00+                 1     6   0.010204082 130.024  129.0
#> 15:       DE       00+                 1     5   0.012396694 292.568  292.0
#> 16:       DE       00+                 1     4   0.017937220 292.827  292.0
#> 17:       DE       00+                 1     3   0.019801980 291.069  289.0
#> 18:       DE       00+                 1     2   0.070175439 292.768  290.0
#> 19:       DE       00+                 1     1   0.383928571 310.151  303.0
#> 20:       DE       00+                 1     0   1.000000000 374.970  359.0
#>     location age_group cum_prop_reported delay prop_reported    mean median
#>             sd     mad     q5   q20    q35   q50    q65   q80    q95      rhat
#>          <num>   <num>  <num> <num>  <num> <num>  <num> <num>  <num>     <num>
#>  1:   0.000000  0.0000 149.00   149 149.00 149.0 149.00   149 149.00        NA
#>  2:   1.293018  1.4826 166.00   166 167.00 167.0 168.00   168 170.00 0.9987896
#>  3:   1.798787  2.2239 133.00   134 135.00 135.5 136.00   137 139.00 1.0008364
#>  4:   2.340807  2.9652 138.00   139 140.00 141.0 142.00   143 145.00 0.9996689
#>  5:   3.092345  3.7065 142.00   143 144.00 145.5 147.00   148 152.00 1.0038336
#>  6:   3.069775  2.9652  99.00   101 102.00 104.0 105.00   106 109.00 1.0006551
#>  7:   2.382912  2.9652  59.00    61  62.00  62.0  63.00    64  67.00 1.0010114
#>  8:   3.988722  4.4478 179.00   182 183.00 185.0 186.00   189 192.00 0.9997242
#>  9:   6.795970  5.9304 246.00   251 253.00 256.0 258.00   262 268.00 0.9997658
#> 10:   7.764038  7.4130 256.00   261 264.00 267.0 270.00   274 281.00 0.9996562
#> 11:   8.613358  7.4130 223.00   229 233.00 236.0 239.00   243 252.00 1.0009820
#> 12:  10.333642  8.8956 216.00   223 227.00 231.0 234.35   240 250.00 1.0012825
#> 13:   9.970927 10.3782 150.00   157 161.00 164.0 168.00   173 182.05 1.0020750
#> 14:   8.503278  8.8956 117.00   123 126.00 129.0 133.00   137 144.00 1.0003030
#> 15:  11.588255 11.1195 275.00   283 288.00 292.0 296.00   302 313.00 1.0002163
#> 16:  15.568798 14.8260 269.95   279 287.00 292.0 297.00   305 320.00 0.9993465
#> 17:  20.306417 19.2738 262.00   274 281.65 289.0 297.00   307 326.00 1.0053979
#> 18:  28.243546 26.6868 253.00   269 280.00 290.0 300.00   314 344.00 1.0017054
#> 19:  50.513957 45.9606 240.00   267 286.00 303.0 322.00   348 403.00 0.9990505
#> 20: 105.934202 93.4038 237.00   290 324.00 359.0 400.00   450 555.25 1.0005813
#>             sd     mad     q5   q20    q35   q50    q65   q80    q95      rhat
#>      ess_bulk  ess_tail
#>         <num>     <num>
#>  1:        NA        NA
#>  2: 1188.2125 1063.3471
#>  3:  963.5041  951.3370
#>  4: 1102.2710  925.0777
#>  5:  987.5966  824.7528
#>  6: 1101.8532  991.0711
#>  7:  751.4622  773.2821
#>  8: 1059.1415  883.2260
#>  9: 1039.1865  998.6170
#> 10: 1039.1951  936.1291
#> 11: 1187.7267 1046.0769
#> 12:  863.1990  964.5881
#> 13: 1113.7200  714.2937
#> 14: 1148.2821  886.6410
#> 15: 1089.7858 1059.6357
#> 16: 1014.5908  974.8269
#> 17:  946.5257  909.0082
#> 18: 1208.7533  912.9634
#> 19: 1404.1200  860.4009
#> 20: 1458.2840  771.2588
#>      ess_bulk  ess_tail