Add latest observations to nowcast output

Add the latest observations to the nowcast output. This is useful for plotting the nowcast against the latest observations.

Usage

enw_add_latest_obs_to_nowcast(nowcast, obs)

Arguments

nowcast

A data.frame of nowcast output from enw_nowcast_summary().

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

Value

A data.frame of nowcast output with the latest observations added.

See also

Examples

fit <- enw_example("nowcast")
obs <- enw_example("obs")
nowcast <- summary(fit, type = "nowcast")
enw_add_latest_obs_to_nowcast(nowcast, obs)
#>     reference_date .group latest_confirm confirm report_date max_confirm
#>  1:     2021-08-03      1            156     149  2021-08-22         149
#>  2:     2021-08-04      1            183     166  2021-08-22         166
#>  3:     2021-08-05      1            147     133  2021-08-22         133
#>  4:     2021-08-06      1            155     137  2021-08-22         137
#>  5:     2021-08-07      1            159     139  2021-08-22         139
#>  6:     2021-08-08      1            119      97  2021-08-22          97
#>  7:     2021-08-09      1             65      58  2021-08-22          58
#>  8:     2021-08-10      1            204     175  2021-08-22         175
#>  9:     2021-08-11      1            275     233  2021-08-22         233
#> 10:     2021-08-12      1            273     237  2021-08-22         237
#> 11:     2021-08-13      1            270     204  2021-08-22         204
#> 12:     2021-08-14      1            262     189  2021-08-22         189
#> 13:     2021-08-15      1            192     125  2021-08-22         125
#> 14:     2021-08-16      1            140      98  2021-08-22          98
#> 15:     2021-08-17      1            323     242  2021-08-22         242
#> 16:     2021-08-18      1            409     223  2021-08-22         223
#> 17:     2021-08-19      1            370     202  2021-08-22         202
#> 18:     2021-08-20      1            361     171  2021-08-22         171
#> 19:     2021-08-21      1            339     112  2021-08-22         112
#> 20:     2021-08-22      1            258      45  2021-08-22          45
#>     location age_group cum_prop_reported delay prop_reported    mean median
#>  1:       DE       00+                 1    19   0.000000000 149.000    149
#>  2:       DE       00+                 1    18   0.000000000 167.374    167
#>  3:       DE       00+                 1    17   0.000000000 135.815    136
#>  4:       DE       00+                 1    16   0.000000000 141.216    141
#>  5:       DE       00+                 1    15   0.007194245 145.893    145
#>  6:       DE       00+                 1    14   0.000000000 103.790    103
#>  7:       DE       00+                 1    13   0.000000000  62.841     62
#>  8:       DE       00+                 1    12   0.000000000 184.990    185
#>  9:       DE       00+                 1    11   0.000000000 256.377    256
#> 10:       DE       00+                 1    10   0.004219409 267.731    267
#> 11:       DE       00+                 1     9   0.000000000 236.758    236
#> 12:       DE       00+                 1     8   0.015873016 231.538    231
#> 13:       DE       00+                 1     7   0.040000000 165.269    164
#> 14:       DE       00+                 1     6   0.010204082 129.858    129
#> 15:       DE       00+                 1     5   0.012396694 293.125    292
#> 16:       DE       00+                 1     4   0.017937220 293.747    292
#> 17:       DE       00+                 1     3   0.019801980 291.156    289
#> 18:       DE       00+                 1     2   0.070175439 294.404    291
#> 19:       DE       00+                 1     1   0.383928571 311.311    307
#> 20:       DE       00+                 1     0   1.000000000 372.540    360
#>             sd     mad     q5 q20 q35 q50    q65   q80    q95      rhat
#>  1:   0.000000  0.0000 149.00 149 149 149 149.00 149.0 149.00        NA
#>  2:   1.289103  1.4826 166.00 166 167 167 168.00 168.0 170.00 0.9988559
#>  3:   1.837975  1.4826 133.00 134 135 136 136.00 137.0 139.00 1.0000435
#>  4:   2.345815  2.9652 138.00 139 140 141 142.00 143.0 145.00 1.0024112
#>  5:   3.111468  2.9652 141.00 143 144 145 147.00 149.0 151.05 0.9986451
#>  6:   3.037619  2.9652  99.00 101 102 103 105.00 106.0 109.00 0.9998714
#>  7:   2.658719  2.9652  59.00  61  62  62  64.00  65.0  68.00 1.0015584
#>  8:   3.790814  2.9652 179.00 182 183 185 186.00 188.0 192.00 1.0001197
#>  9:   6.672285  5.9304 247.00 251 254 256 258.00 261.0 268.00 1.0002656
#> 10:   8.319727  8.8956 255.00 261 264 267 270.00 275.0 282.00 1.0047052
#> 11:   8.200285  7.4130 224.00 230 233 236 239.00 243.2 251.00 0.9995461
#> 12:  10.108053 10.3782 216.00 223 227 231 235.00 240.0 250.00 1.0004555
#> 13:  10.333800 10.3782 149.95 156 161 164 168.00 173.0 183.00 1.0004271
#> 14:   8.445660  7.4130 117.00 123 126 129 132.00 137.0 145.00 1.0014760
#> 15:  12.075811 11.8608 275.00 283 288 292 297.00 302.2 315.00 1.0010481
#> 16:  16.541784 14.8260 269.00 280 286 292 297.35 306.2 325.05 0.9994023
#> 17:  21.078804 19.2738 260.00 274 282 289 298.00 307.0 327.00 0.9995880
#> 18:  29.526710 28.1694 251.00 269 281 291 303.00 316.0 350.05 1.0006337
#> 19:  50.224743 47.4432 238.00 269 288 307 325.00 349.2 401.05 1.0015331
#> 20: 105.260887 91.9212 232.95 287 323 360 394.00 443.2 565.05 0.9997072
#>      ess_bulk  ess_tail
#>  1:        NA        NA
#>  2:  924.1472  912.2693
#>  3:  771.7362  922.9924
#>  4:  898.8139  983.5991
#>  5:  794.3513  874.6964
#>  6:  988.0289  880.0392
#>  7:  938.9066  820.2890
#>  8:  869.3789  985.1324
#>  9: 1005.6559  921.4996
#> 10:  966.8434  952.7493
#> 11:  956.9055  887.3885
#> 12: 1107.9597 1004.2316
#> 13: 1165.4725  816.2192
#> 14: 1218.5668  759.7062
#> 15: 1147.2772  939.7924
#> 16: 1133.2367 1022.1212
#> 17: 1047.4570  878.3429
#> 18: 1349.2910  894.1807
#> 19: 1222.3760  832.5573
#> 20: 1389.5766  880.9037