Add latest observations to nowcast output

Add the latest observations to the nowcast output. This is useful for plotting the nowcast against the latest observations.

Usage

enw_add_latest_obs_to_nowcast(nowcast, obs)

Arguments

nowcast

A data.frame of nowcast output from enw_nowcast_summary().

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

Value

A data.frame of nowcast output with the latest observations added.

See also

Examples

fit <- enw_example("nowcast")
obs <- enw_example("obs")
nowcast <- summary(fit, type = "nowcast")
enw_add_latest_obs_to_nowcast(nowcast, obs)
#>     reference_date .group latest_confirm confirm report_date max_confirm
#>  1:     2021-08-03      1            156     149  2021-08-22         149
#>  2:     2021-08-04      1            183     166  2021-08-22         166
#>  3:     2021-08-05      1            147     133  2021-08-22         133
#>  4:     2021-08-06      1            155     137  2021-08-22         137
#>  5:     2021-08-07      1            159     139  2021-08-22         139
#>  6:     2021-08-08      1            119      97  2021-08-22          97
#>  7:     2021-08-09      1             65      58  2021-08-22          58
#>  8:     2021-08-10      1            204     175  2021-08-22         175
#>  9:     2021-08-11      1            275     233  2021-08-22         233
#> 10:     2021-08-12      1            273     237  2021-08-22         237
#> 11:     2021-08-13      1            270     204  2021-08-22         204
#> 12:     2021-08-14      1            262     189  2021-08-22         189
#> 13:     2021-08-15      1            192     125  2021-08-22         125
#> 14:     2021-08-16      1            140      98  2021-08-22          98
#> 15:     2021-08-17      1            323     242  2021-08-22         242
#> 16:     2021-08-18      1            409     223  2021-08-22         223
#> 17:     2021-08-19      1            370     202  2021-08-22         202
#> 18:     2021-08-20      1            361     171  2021-08-22         171
#> 19:     2021-08-21      1            339     112  2021-08-22         112
#> 20:     2021-08-22      1            258      45  2021-08-22          45
#>     location age_group cum_prop_reported delay prop_reported    mean median
#>  1:       DE       00+                 1    19   0.000000000 149.000  149.0
#>  2:       DE       00+                 1    18   0.000000000 167.440  167.0
#>  3:       DE       00+                 1    17   0.000000000 135.804  136.0
#>  4:       DE       00+                 1    16   0.000000000 141.039  141.0
#>  5:       DE       00+                 1    15   0.007194245 145.916  146.0
#>  6:       DE       00+                 1    14   0.000000000 103.674  103.0
#>  7:       DE       00+                 1    13   0.000000000  63.014   63.0
#>  8:       DE       00+                 1    12   0.000000000 185.131  185.0
#>  9:       DE       00+                 1    11   0.000000000 255.938  256.0
#> 10:       DE       00+                 1    10   0.004219409 267.445  267.0
#> 11:       DE       00+                 1     9   0.000000000 236.189  235.5
#> 12:       DE       00+                 1     8   0.015873016 230.969  230.0
#> 13:       DE       00+                 1     7   0.040000000 165.393  165.0
#> 14:       DE       00+                 1     6   0.010204082 129.419  129.0
#> 15:       DE       00+                 1     5   0.012396694 293.442  292.0
#> 16:       DE       00+                 1     4   0.017937220 292.867  292.0
#> 17:       DE       00+                 1     3   0.019801980 292.416  291.0
#> 18:       DE       00+                 1     2   0.070175439 294.685  291.0
#> 19:       DE       00+                 1     1   0.383928571 312.347  307.0
#> 20:       DE       00+                 1     0   1.000000000 377.870  363.0
#>             sd     mad     q5 q20    q35   q50    q65 q80    q95      rhat
#>  1:   0.000000  0.0000 149.00 149 149.00 149.0 149.00 149 149.00        NA
#>  2:   1.306946  1.4826 166.00 166 167.00 167.0 168.00 168 170.00 0.9984731
#>  3:   1.910297  1.4826 133.00 134 135.00 136.0 136.00 137 139.00 1.0001748
#>  4:   2.336009  1.4826 138.00 139 140.00 141.0 142.00 143 145.00 0.9985918
#>  5:   3.245223  2.9652 141.00 143 145.00 146.0 147.00 148 152.00 1.0014625
#>  6:   3.004455  2.9652  99.00 101 102.00 103.0 105.00 106 109.00 1.0008170
#>  7:   2.633769  2.9652  59.00  61  62.00  63.0  64.00  65  68.00 1.0001596
#>  8:   3.861703  4.4478 179.00 182 183.00 185.0 186.00 188 192.00 1.0016498
#>  9:   6.338796  5.9304 246.00 251 253.00 256.0 258.00 261 267.00 1.0008653
#> 10:   7.795495  7.4130 256.00 261 264.00 267.0 270.00 274 281.00 1.0020121
#> 11:   8.328183  8.1543 224.00 229 233.00 235.5 239.00 243 251.00 0.9993738
#> 12:   9.851248  9.6369 216.00 223 227.00 230.0 234.00 239 247.00 0.9997526
#> 13:  10.147883 10.3782 150.00 157 161.00 165.0 168.00 173 184.00 0.9989137
#> 14:   8.341524  8.8956 116.00 122 126.00 129.0 132.00 136 144.00 1.0002835
#> 15:  12.342730 11.8608 276.00 283 288.00 292.0 297.00 303 316.00 0.9993443
#> 16:  15.235795 14.8260 271.00 279 286.00 292.0 297.00 305 320.05 1.0000165
#> 17:  20.600906 19.2738 263.00 275 283.00 291.0 298.35 309 331.00 1.0007083
#> 18:  29.652944 28.1694 251.95 270 282.00 291.0 302.00 318 347.05 1.0031517
#> 19:  52.500656 50.4084 237.00 269 287.65 307.0 327.00 351 403.15 1.0029817
#> 20: 106.473272 93.4038 239.95 292 327.00 363.0 401.00 451 576.15 1.0034715
#>      ess_bulk  ess_tail
#>  1:        NA        NA
#>  2: 1072.5191 1068.0041
#>  3: 1040.7197 1025.8570
#>  4: 1021.8261  829.9055
#>  5: 1036.5263  915.7683
#>  6: 1084.2564  888.0770
#>  7: 1170.0810  845.5932
#>  8:  826.3595  989.6300
#>  9:  976.1162  943.8214
#> 10: 1016.2724  988.2011
#> 11: 1119.8109  785.8675
#> 12: 1067.0426  980.9775
#> 13: 1252.9097  836.7735
#> 14: 1111.6597  956.2202
#> 15: 1076.8727  782.2336
#> 16: 1040.5391  767.7477
#> 17:  950.4381  731.4047
#> 18: 1230.4660  982.6347
#> 19: 1076.9075  871.5941
#> 20: 1394.6975  915.4979