Load a package example

Loads examples of nowcasts produce using example scripts. Used to streamline examples, in package tests and to enable users to explore package functionality without needing to install cmdstanr.

Usage

enw_example(type = "nowcast")

Arguments

type

A character string indicating the example to load. Supported options are "nowcast, "preprocessed_observations", "observations", and "script" which are the output of epinowcast()], enw_preprocess_data(), and enw_latest_data() applied to the germany_covid19_hosp package dataset), and the script used to generate these examples respectively.

Value

A data.table of summarised output

See also

Examples

# Load the nowcast
enw_example(type = "nowcast")
#>                    obs          new_confirm              latest
#> 1: <data.table[671x9]> <data.table[630x11]> <data.table[41x10]>
#>     missing_reference  reporting_triangle      metareference
#> 1: <data.table[41x6]> <data.table[41x22]> <data.table[41x9]>
#>             metareport          metadelay time snapshots by groups max_delay
#> 1: <data.table[60x12]> <data.table[20x4]>   41        41         1        20
#>      max_date               fit       data  fit_args samples max_rhat
#> 1: 2021-08-22 <CmdStanMCMC[32]> <list[52]> <list[6]>    1000     1.01
#>    divergent_transitions per_divergent_transitions max_treedepth
#> 1:                     0                         0             8
#>    no_at_max_treedepth per_at_max_treedepth run_time
#> 1:                 115                0.115     79.6

# Load the preprocessed observations
enw_example(type = "preprocessed_observations")
#>                    obs          new_confirm              latest
#> 1: <data.table[671x9]> <data.table[630x11]> <data.table[41x10]>
#>     missing_reference  reporting_triangle      metareference
#> 1: <data.table[41x6]> <data.table[41x22]> <data.table[41x9]>
#>             metareport          metadelay time snapshots by groups max_delay
#> 1: <data.table[60x12]> <data.table[20x4]>   41        41         1        20
#>      max_date
#> 1: 2021-08-22

# Load the latest observations
enw_example(type = "observations")
#>     reference_date location age_group confirm report_date
#>  1:     2021-08-02       DE       00+      59  2021-10-01
#>  2:     2021-08-03       DE       00+     156  2021-10-01
#>  3:     2021-08-04       DE       00+     183  2021-10-01
#>  4:     2021-08-05       DE       00+     147  2021-10-01
#>  5:     2021-08-06       DE       00+     155  2021-10-01
#>  6:     2021-08-07       DE       00+     159  2021-10-01
#>  7:     2021-08-08       DE       00+     119  2021-10-01
#>  8:     2021-08-09       DE       00+      65  2021-10-01
#>  9:     2021-08-10       DE       00+     204  2021-10-01
#> 10:     2021-08-11       DE       00+     275  2021-10-01
#> 11:     2021-08-12       DE       00+     273  2021-10-01
#> 12:     2021-08-13       DE       00+     270  2021-10-01
#> 13:     2021-08-14       DE       00+     262  2021-10-01
#> 14:     2021-08-15       DE       00+     192  2021-10-01
#> 15:     2021-08-16       DE       00+     140  2021-10-01
#> 16:     2021-08-17       DE       00+     323  2021-10-01
#> 17:     2021-08-18       DE       00+     409  2021-10-01
#> 18:     2021-08-19       DE       00+     370  2021-10-01
#> 19:     2021-08-20       DE       00+     361  2021-10-01
#> 20:     2021-08-21       DE       00+     339  2021-10-01
#> 21:     2021-08-22       DE       00+     258  2021-10-01
#>     reference_date location age_group confirm report_date

# Load the script used to generate these examples
# Optionally source this script to regenerate the example
readLines(enw_example(type = "script"))
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former
#> character(0)