Define a model using a formula interface

This function allows models to be defined using a flexible formula interface that supports fixed effects, random effects (using lme4 syntax). Note that the returned fixed effects design matrix is sparse and so the index supplied is required to link observations to the appropriate design matrix row.

Usage

enw_formula(formula, data, sparse = TRUE)

Arguments

formula

A model formula that may use standard fixed effects, random effects using lme4 syntax (see re()), and random walks defined using the rw() helper function.

data

A data.frame of observations. It must include all variables used in the supplied formula.

sparse

Logical, defaults to TRUE. Should the fixed effects design matrix be sparely defined.

Value

A list containing the following:

formula: The user supplied formula

parsed_formula: The formula as parsed by parse_formula()

extended_formula: The flattened version of the formula with both user supplied terms and terms added for the user supplied complex model components.

fixed: A list containing the fixed effect formula, sparse design matrix, and the index linking the design matrix with observations.

random: A list containing the random effect formula, sparse design matrix, and the index linking the design matrix with random effects.

See also

Examples

# Use meta data for references dates from the Germany COVID-19
# hospitalisation data.
obs <- enw_filter_report_dates(
  germany_covid19_hosp[location == "DE"],
  remove_days = 40
)
obs <- enw_filter_reference_dates(obs, include_days = 40)
pobs <- enw_preprocess_data(obs, by = c("age_group", "location"))
data <- pobs$metareference[[1]]

# Model with fixed effects for age group
enw_formula(~ 1 + age_group, data)
#> $formula
#> [1] "~1 + age_group"
#> 
#> $parsed_formula
#> $parsed_formula$fixed
#> [1] "1"         "age_group"
#> 
#> $parsed_formula$random
#> NULL
#> 
#> $parsed_formula$rw
#> character(0)
#> 
#> 
#> $expanded_formula
#> [1] "~1 + age_group"
#> 
#> $fixed
#> $fixed$formula
#> [1] "~1 + age_group"
#> 
#> $fixed$design
#>     (Intercept) age_group00+ age_group05-14 age_group15-34 age_group35-59
#> 1             1            0              0              0              0
#> 42            1            1              0              0              0
#> 83            1            0              1              0              0
#> 124           1            0              0              1              0
#> 165           1            0              0              0              1
#> 206           1            0              0              0              0
#> 247           1            0              0              0              0
#>     age_group60-79 age_group80+
#> 1                0            0
#> 42               0            0
#> 83               0            0
#> 124              0            0
#> 165              0            0
#> 206              1            0
#> 247              0            1
#> 
#> $fixed$index
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [112] 3 3 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4
#> [149] 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5
#> [186] 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6
#> [223] 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 7 7 7 7 7 7 7 7 7 7 7 7 7
#> [260] 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7
#> 
#> 
#> $random
#> $random$formula
#> [1] "~1"
#> 
#> $random$design
#>   (Intercept)
#> 1           1
#> 2           1
#> 3           1
#> 4           1
#> 5           1
#> 6           1
#> attr(,"assign")
#> [1] 0
#> 
#> $random$index
#> [1] 1 2 3 4 5 6
#> 
#> 
#> attr(,"class")
#> [1] "enw_formula" "list"       

# Model with random effects for age group
enw_formula(~ 1 + (1 | age_group), data)
#> $formula
#> [1] "~1 + (1 | age_group)"
#> 
#> $parsed_formula
#> $parsed_formula$fixed
#> [1] "1"
#> 
#> $parsed_formula$random
#> $parsed_formula$random[[1]]
#> 1 | age_group
#> 
#> 
#> $parsed_formula$rw
#> character(0)
#> 
#> 
#> $expanded_formula
#> [1] "~1 + age_group"
#> 
#> $fixed
#> $fixed$formula
#> [1] "~1 + age_group"
#> 
#> $fixed$design
#>     (Intercept) age_group00-04 age_group00+ age_group05-14 age_group15-34
#> 1             1              1            0              0              0
#> 42            1              0            1              0              0
#> 83            1              0            0              1              0
#> 124           1              0            0              0              1
#> 165           1              0            0              0              0
#> 206           1              0            0              0              0
#> 247           1              0            0              0              0
#>     age_group35-59 age_group60-79 age_group80+
#> 1                0              0            0
#> 42               0              0            0
#> 83               0              0            0
#> 124              0              0            0
#> 165              1              0            0
#> 206              0              1            0
#> 247              0              0            1
#> 
#> $fixed$index
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [112] 3 3 3 3 3 3 3 3 3 3 3 3 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4
#> [149] 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5
#> [186] 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6
#> [223] 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 7 7 7 7 7 7 7 7 7 7 7 7 7
#> [260] 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7 7
#> 
#> 
#> $random
#> $random$formula
#> [1] "~0 + fixed + age_group"
#> 
#> $random$design
#>   fixed age_group
#> 1     0         1
#> 2     0         1
#> 3     0         1
#> 4     0         1
#> 5     0         1
#> 6     0         1
#> 7     0         1
#> attr(,"assign")
#> [1] 1 2
#> 
#> $random$index
#> [1] 1 2 3 4 5 6 7
#> 
#> 
#> attr(,"class")
#> [1] "enw_formula" "list"       

# Model with a random effect for age group and a random walk
enw_formula(~ 1 + (1 | age_group) + rw(week), data)
#> $formula
#> [1] "~1 + (1 | age_group) + rw(week)"
#> 
#> $parsed_formula
#> $parsed_formula$fixed
#> [1] "1"
#> 
#> $parsed_formula$random
#> $parsed_formula$random[[1]]
#> 1 | age_group
#> 
#> 
#> $parsed_formula$rw
#> [1] "rw(week)"
#> 
#> 
#> $expanded_formula
#> [1] "~1 + cweek1 + cweek2 + cweek3 + cweek4 + cweek5 + cweek6 + age_group"
#> 
#> $fixed
#> $fixed$formula
#> [1] "~1 + cweek1 + cweek2 + cweek3 + cweek4 + cweek5 + cweek6 + age_group"
#> 
#> $fixed$design
#>     (Intercept) cweek1 cweek2 cweek3 cweek4 cweek5 cweek6 age_group00-04
#> 1             1      0      0      0      0      0      0              1
#> 6             1      1      0      0      0      0      0              1
#> 13            1      1      1      0      0      0      0              1
#> 20            1      1      1      1      0      0      0              1
#> 27            1      1      1      1      1      0      0              1
#> 34            1      1      1      1      1      1      0              1
#> 41            1      1      1      1      1      1      1              1
#> 42            1      0      0      0      0      0      0              0
#> 47            1      1      0      0      0      0      0              0
#> 54            1      1      1      0      0      0      0              0
#> 61            1      1      1      1      0      0      0              0
#> 68            1      1      1      1      1      0      0              0
#> 75            1      1      1      1      1      1      0              0
#> 82            1      1      1      1      1      1      1              0
#> 83            1      0      0      0      0      0      0              0
#> 88            1      1      0      0      0      0      0              0
#> 95            1      1      1      0      0      0      0              0
#> 102           1      1      1      1      0      0      0              0
#> 109           1      1      1      1      1      0      0              0
#> 116           1      1      1      1      1      1      0              0
#> 123           1      1      1      1      1      1      1              0
#> 124           1      0      0      0      0      0      0              0
#> 129           1      1      0      0      0      0      0              0
#> 136           1      1      1      0      0      0      0              0
#> 143           1      1      1      1      0      0      0              0
#> 150           1      1      1      1      1      0      0              0
#> 157           1      1      1      1      1      1      0              0
#> 164           1      1      1      1      1      1      1              0
#> 165           1      0      0      0      0      0      0              0
#> 170           1      1      0      0      0      0      0              0
#> 177           1      1      1      0      0      0      0              0
#> 184           1      1      1      1      0      0      0              0
#> 191           1      1      1      1      1      0      0              0
#> 198           1      1      1      1      1      1      0              0
#> 205           1      1      1      1      1      1      1              0
#> 206           1      0      0      0      0      0      0              0
#> 211           1      1      0      0      0      0      0              0
#> 218           1      1      1      0      0      0      0              0
#> 225           1      1      1      1      0      0      0              0
#> 232           1      1      1      1      1      0      0              0
#> 239           1      1      1      1      1      1      0              0
#> 246           1      1      1      1      1      1      1              0
#> 247           1      0      0      0      0      0      0              0
#> 252           1      1      0      0      0      0      0              0
#> 259           1      1      1      0      0      0      0              0
#> 266           1      1      1      1      0      0      0              0
#> 273           1      1      1      1      1      0      0              0
#> 280           1      1      1      1      1      1      0              0
#> 287           1      1      1      1      1      1      1              0
#>     age_group00+ age_group05-14 age_group15-34 age_group35-59 age_group60-79
#> 1              0              0              0              0              0
#> 6              0              0              0              0              0
#> 13             0              0              0              0              0
#> 20             0              0              0              0              0
#> 27             0              0              0              0              0
#> 34             0              0              0              0              0
#> 41             0              0              0              0              0
#> 42             1              0              0              0              0
#> 47             1              0              0              0              0
#> 54             1              0              0              0              0
#> 61             1              0              0              0              0
#> 68             1              0              0              0              0
#> 75             1              0              0              0              0
#> 82             1              0              0              0              0
#> 83             0              1              0              0              0
#> 88             0              1              0              0              0
#> 95             0              1              0              0              0
#> 102            0              1              0              0              0
#> 109            0              1              0              0              0
#> 116            0              1              0              0              0
#> 123            0              1              0              0              0
#> 124            0              0              1              0              0
#> 129            0              0              1              0              0
#> 136            0              0              1              0              0
#> 143            0              0              1              0              0
#> 150            0              0              1              0              0
#> 157            0              0              1              0              0
#> 164            0              0              1              0              0
#> 165            0              0              0              1              0
#> 170            0              0              0              1              0
#> 177            0              0              0              1              0
#> 184            0              0              0              1              0
#> 191            0              0              0              1              0
#> 198            0              0              0              1              0
#> 205            0              0              0              1              0
#> 206            0              0              0              0              1
#> 211            0              0              0              0              1
#> 218            0              0              0              0              1
#> 225            0              0              0              0              1
#> 232            0              0              0              0              1
#> 239            0              0              0              0              1
#> 246            0              0              0              0              1
#> 247            0              0              0              0              0
#> 252            0              0              0              0              0
#> 259            0              0              0              0              0
#> 266            0              0              0              0              0
#> 273            0              0              0              0              0
#> 280            0              0              0              0              0
#> 287            0              0              0              0              0
#>     age_group80+
#> 1              0
#> 6              0
#> 13             0
#> 20             0
#> 27             0
#> 34             0
#> 41             0
#> 42             0
#> 47             0
#> 54             0
#> 61             0
#> 68             0
#> 75             0
#> 82             0
#> 83             0
#> 88             0
#> 95             0
#> 102            0
#> 109            0
#> 116            0
#> 123            0
#> 124            0
#> 129            0
#> 136            0
#> 143            0
#> 150            0
#> 157            0
#> 164            0
#> 165            0
#> 170            0
#> 177            0
#> 184            0
#> 191            0
#> 198            0
#> 205            0
#> 206            0
#> 211            0
#> 218            0
#> 225            0
#> 232            0
#> 239            0
#> 246            0
#> 247            1
#> 252            1
#> 259            1
#> 266            1
#> 273            1
#> 280            1
#> 287            1
#> 
#> $fixed$index
#>   [1]  1  1  1  1  1  2  2  2  2  2  2  2  3  3  3  3  3  3  3  4  4  4  4  4  4
#>  [26]  4  5  5  5  5  5  5  5  6  6  6  6  6  6  6  7  8  8  8  8  8  9  9  9  9
#>  [51]  9  9  9 10 10 10 10 10 10 10 11 11 11 11 11 11 11 12 12 12 12 12 12 12 13
#>  [76] 13 13 13 13 13 13 14 15 15 15 15 15 16 16 16 16 16 16 16 17 17 17 17 17 17
#> [101] 17 18 18 18 18 18 18 18 19 19 19 19 19 19 19 20 20 20 20 20 20 20 21 22 22
#> [126] 22 22 22 23 23 23 23 23 23 23 24 24 24 24 24 24 24 25 25 25 25 25 25 25 26
#> [151] 26 26 26 26 26 26 27 27 27 27 27 27 27 28 29 29 29 29 29 30 30 30 30 30 30
#> [176] 30 31 31 31 31 31 31 31 32 32 32 32 32 32 32 33 33 33 33 33 33 33 34 34 34
#> [201] 34 34 34 34 35 36 36 36 36 36 37 37 37 37 37 37 37 38 38 38 38 38 38 38 39
#> [226] 39 39 39 39 39 39 40 40 40 40 40 40 40 41 41 41 41 41 41 41 42 43 43 43 43
#> [251] 43 44 44 44 44 44 44 44 45 45 45 45 45 45 45 46 46 46 46 46 46 46 47 47 47
#> [276] 47 47 47 47 48 48 48 48 48 48 48 49
#> 
#> 
#> $random
#> $random$formula
#> [1] "~0 + fixed + age_group + week"
#> 
#> $random$design
#>    fixed age_group week
#> 1      0         1    0
#> 2      0         1    0
#> 3      0         1    0
#> 4      0         1    0
#> 5      0         1    0
#> 6      0         1    0
#> 7      0         1    0
#> 8      0         0    1
#> 9      0         0    1
#> 10     0         0    1
#> 11     0         0    1
#> 12     0         0    1
#> 13     0         0    1
#> attr(,"assign")
#> [1] 1 2 3
#> 
#> $random$index
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13
#> 
#> 
#> attr(,"class")
#> [1] "enw_formula" "list"       

# Model defined without a sparse fixed effects design matrix
enw_formula(~1, data[1:20, ])
#> $formula
#> [1] "~1"
#> 
#> $parsed_formula
#> $parsed_formula$fixed
#> [1] "1"
#> 
#> $parsed_formula$random
#> NULL
#> 
#> $parsed_formula$rw
#> character(0)
#> 
#> 
#> $expanded_formula
#> [1] "~1"
#> 
#> $fixed
#> $fixed$formula
#> [1] "~1"
#> 
#> $fixed$design
#>   (Intercept)
#> 1           1
#> 
#> $fixed$index
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> 
#> $random
#> $random$formula
#> [1] "~1"
#> 
#> $random$design
#>      (Intercept)
#> attr(,"assign")
#> [1] 0
#> 
#> $random$index
#> integer(0)
#> 
#> 
#> attr(,"class")
#> [1] "enw_formula" "list"