Summarise the posterior nowcast prediction

A generic wrapper around enw_posterior() with opinionated defaults to extract the posterior prediction for the nowcast ("pp_inf_obs" from the stan code). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_nowcast_summary(
  fit,
  obs,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  timestep = "day"
)

Arguments

fit

A cmdstanr fit object.

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

timestep

The timestep to used. This can be a string ("day", "week", "month") or a numeric whole number representing the number of days.

Value

A data.frame summarising the model posterior nowcast prediction. This uses observed data where available and the posterior prediction where not.

See also

Examples

fit <- enw_example("nowcast")
enw_nowcast_summary(fit$fit[[1]], fit$latest[[1]])
#>     reference_date report_date .group max_confirm location age_group confirm
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>  1:                 1    19   0.000000000 149.000    149   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.510    167   1.477187  1.4826
#>  3:                 1    17   0.000000000 135.699    135   1.821460  1.4826
#>  4:                 1    16   0.000000000 141.110    141   2.278397  2.9652
#>  5:                 1    15   0.007194245 145.919    146   3.102590  2.9652
#>  6:                 1    14   0.000000000 103.761    103   3.001814  2.9652
#>  7:                 1    13   0.000000000  62.843     63   2.447108  2.9652
#>  8:                 1    12   0.000000000 185.143    185   3.892776  4.4478
#>  9:                 1    11   0.000000000 256.242    256   6.614922  5.9304
#> 10:                 1    10   0.004219409 267.754    267   7.815406  7.4130
#> 11:                 1     9   0.000000000 236.311    235   8.261994  7.4130
#> 12:                 1     8   0.015873016 230.777    230   9.846330  8.8956
#> 13:                 1     7   0.040000000 165.677    164  10.046871 10.3782
#> 14:                 1     6   0.010204082 129.764    129   8.397549  7.4130
#> 15:                 1     5   0.012396694 293.502    292  12.324443 11.8608
#> 16:                 1     4   0.017937220 293.818    292  16.721797 14.8260
#> 17:                 1     3   0.019801980 292.588    291  21.275500 22.2390
#> 18:                 1     2   0.070175439 293.411    291  28.361531 26.6868
#> 19:                 1     1   0.383928571 310.284    305  51.245443 46.7019
#> 20:                 1     0   1.000000000 381.715    366 105.802296 94.8864
#>         q5   q20 q35 q50 q65   q80    q95      rhat  ess_bulk ess_tail
#>  1: 149.00 149.0 149 149 149 149.0 149.00        NA        NA       NA
#>  2: 166.00 166.0 167 167 168 169.0 170.00 1.0021059  916.5251 961.2656
#>  3: 133.00 134.0 135 135 136 137.0 139.00 1.0026309  865.5792 980.4149
#>  4: 138.00 139.0 140 141 142 143.0 145.05 0.9985305  964.6971 916.0647
#>  5: 141.95 143.0 144 146 147 148.0 152.00 1.0024283  982.1855 878.7477
#>  6:  99.00 101.0 102 103 105 106.0 109.00 1.0002988 1079.3848 873.5648
#>  7:  59.00  61.0  62  63  64  65.0  67.00 1.0050496 1015.4992 831.1428
#>  8: 179.00 182.0 184 185 186 188.0 192.00 0.9987252 1035.5304 854.2485
#>  9: 246.00 251.0 253 256 259 262.0 267.00 0.9997534 1020.3568 936.2525
#> 10: 256.95 261.0 264 267 270 274.0 282.00 1.0015893 1073.7290 947.6912
#> 11: 225.00 229.0 232 235 239 243.0 251.00 1.0009977 1024.8633 849.3204
#> 12: 216.00 223.0 227 230 234 238.0 249.00 0.9981689 1305.3677 998.3510
#> 13: 150.00 157.0 161 164 169 174.0 184.00 0.9992937 1114.1136 874.1649
#> 14: 118.00 123.0 126 129 132 136.0 145.00 0.9992580 1020.4703 799.7645
#> 15: 275.95 283.0 288 292 297 303.0 316.00 0.9997709 1097.9740 908.7449
#> 16: 269.00 280.0 287 292 298 306.0 323.00 0.9987247 1025.9249 971.7279
#> 17: 260.95 274.0 282 291 299 310.0 333.00 0.9996703 1243.7915 969.4998
#> 18: 253.00 269.8 280 291 301 313.2 342.00 1.0014503 1167.6104 910.7979
#> 19: 240.00 267.0 288 305 322 348.2 403.00 0.9996016 1107.4854 874.2575
#> 20: 245.95 291.8 331 366 403 459.0 580.05 0.9991110 1298.2575 864.7724