Summarise the posterior nowcast prediction

A generic wrapper around enw_posterior() with opinionated defaults to extract the posterior prediction for the nowcast ("pp_inf_obs" from the stan code). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_nowcast_summary(fit, obs, probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95))

Arguments

fit

A cmdstanr fit object.

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

Value

A data.frame summarising the model posterior nowcast prediction. This uses observed data where available and the posterior prediction where not.

See also

Examples

fit <- enw_example("nowcast")
enw_nowcast_summary(fit$fit[[1]], fit$latest[[1]])
#>     reference_date report_date .group max_confirm location age_group confirm
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>  1:                 1    19   0.000000000 149.000    149   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.374    167   1.289103  1.4826
#>  3:                 1    17   0.000000000 135.815    136   1.837975  1.4826
#>  4:                 1    16   0.000000000 141.216    141   2.345815  2.9652
#>  5:                 1    15   0.007194245 145.893    145   3.111468  2.9652
#>  6:                 1    14   0.000000000 103.790    103   3.037619  2.9652
#>  7:                 1    13   0.000000000  62.841     62   2.658719  2.9652
#>  8:                 1    12   0.000000000 184.990    185   3.790814  2.9652
#>  9:                 1    11   0.000000000 256.377    256   6.672285  5.9304
#> 10:                 1    10   0.004219409 267.731    267   8.319727  8.8956
#> 11:                 1     9   0.000000000 236.758    236   8.200285  7.4130
#> 12:                 1     8   0.015873016 231.538    231  10.108053 10.3782
#> 13:                 1     7   0.040000000 165.269    164  10.333800 10.3782
#> 14:                 1     6   0.010204082 129.858    129   8.445660  7.4130
#> 15:                 1     5   0.012396694 293.125    292  12.075811 11.8608
#> 16:                 1     4   0.017937220 293.747    292  16.541784 14.8260
#> 17:                 1     3   0.019801980 291.156    289  21.078804 19.2738
#> 18:                 1     2   0.070175439 294.404    291  29.526710 28.1694
#> 19:                 1     1   0.383928571 311.311    307  50.224743 47.4432
#> 20:                 1     0   1.000000000 372.540    360 105.260887 91.9212
#>         q5 q20 q35 q50    q65   q80    q95      rhat  ess_bulk  ess_tail
#>  1: 149.00 149 149 149 149.00 149.0 149.00        NA        NA        NA
#>  2: 166.00 166 167 167 168.00 168.0 170.00 0.9988559  924.1472  912.2693
#>  3: 133.00 134 135 136 136.00 137.0 139.00 1.0000435  771.7362  922.9924
#>  4: 138.00 139 140 141 142.00 143.0 145.00 1.0024112  898.8139  983.5991
#>  5: 141.00 143 144 145 147.00 149.0 151.05 0.9986451  794.3513  874.6964
#>  6:  99.00 101 102 103 105.00 106.0 109.00 0.9998714  988.0289  880.0392
#>  7:  59.00  61  62  62  64.00  65.0  68.00 1.0015584  938.9066  820.2890
#>  8: 179.00 182 183 185 186.00 188.0 192.00 1.0001197  869.3789  985.1324
#>  9: 247.00 251 254 256 258.00 261.0 268.00 1.0002656 1005.6559  921.4996
#> 10: 255.00 261 264 267 270.00 275.0 282.00 1.0047052  966.8434  952.7493
#> 11: 224.00 230 233 236 239.00 243.2 251.00 0.9995461  956.9055  887.3885
#> 12: 216.00 223 227 231 235.00 240.0 250.00 1.0004555 1107.9597 1004.2316
#> 13: 149.95 156 161 164 168.00 173.0 183.00 1.0004271 1165.4725  816.2192
#> 14: 117.00 123 126 129 132.00 137.0 145.00 1.0014760 1218.5668  759.7062
#> 15: 275.00 283 288 292 297.00 302.2 315.00 1.0010481 1147.2772  939.7924
#> 16: 269.00 280 286 292 297.35 306.2 325.05 0.9994023 1133.2367 1022.1212
#> 17: 260.00 274 282 289 298.00 307.0 327.00 0.9995880 1047.4570  878.3429
#> 18: 251.00 269 281 291 303.00 316.0 350.05 1.0006337 1349.2910  894.1807
#> 19: 238.00 269 288 307 325.00 349.2 401.05 1.0015331 1222.3760  832.5573
#> 20: 232.95 287 323 360 394.00 443.2 565.05 0.9997072 1389.5766  880.9037