Summarise the posterior nowcast prediction

A generic wrapper around enw_posterior() with opinionated defaults to extract the posterior prediction for the nowcast ("pp_inf_obs" from the stan code). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_nowcast_summary(
  fit,
  obs,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  timestep = "day"
)

Arguments

fit

A cmdstanr fit object.

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

timestep

The timestep to used. This can be a string ("day", "week", "month") or a numeric whole number representing the number of days.

Value

A data.frame summarising the model posterior nowcast prediction. This uses observed data where available and the posterior prediction where not.

See also

Examples

fit <- enw_example("nowcast")
enw_nowcast_summary(fit$fit[[1]], fit$latest[[1]])
#>     reference_date report_date .group max_confirm location age_group confirm
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>  1:                 1    19   0.000000000 149.000  149.0   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.440  167.0   1.306946  1.4826
#>  3:                 1    17   0.000000000 135.804  136.0   1.910297  1.4826
#>  4:                 1    16   0.000000000 141.039  141.0   2.336009  1.4826
#>  5:                 1    15   0.007194245 145.916  146.0   3.245223  2.9652
#>  6:                 1    14   0.000000000 103.674  103.0   3.004455  2.9652
#>  7:                 1    13   0.000000000  63.014   63.0   2.633769  2.9652
#>  8:                 1    12   0.000000000 185.131  185.0   3.861703  4.4478
#>  9:                 1    11   0.000000000 255.938  256.0   6.338796  5.9304
#> 10:                 1    10   0.004219409 267.445  267.0   7.795495  7.4130
#> 11:                 1     9   0.000000000 236.189  235.5   8.328183  8.1543
#> 12:                 1     8   0.015873016 230.969  230.0   9.851248  9.6369
#> 13:                 1     7   0.040000000 165.393  165.0  10.147883 10.3782
#> 14:                 1     6   0.010204082 129.419  129.0   8.341524  8.8956
#> 15:                 1     5   0.012396694 293.442  292.0  12.342730 11.8608
#> 16:                 1     4   0.017937220 292.867  292.0  15.235795 14.8260
#> 17:                 1     3   0.019801980 292.416  291.0  20.600906 19.2738
#> 18:                 1     2   0.070175439 294.685  291.0  29.652944 28.1694
#> 19:                 1     1   0.383928571 312.347  307.0  52.500656 50.4084
#> 20:                 1     0   1.000000000 377.870  363.0 106.473272 93.4038
#>         q5 q20    q35   q50    q65 q80    q95      rhat  ess_bulk  ess_tail
#>  1: 149.00 149 149.00 149.0 149.00 149 149.00        NA        NA        NA
#>  2: 166.00 166 167.00 167.0 168.00 168 170.00 0.9984731 1072.5191 1068.0041
#>  3: 133.00 134 135.00 136.0 136.00 137 139.00 1.0001748 1040.7197 1025.8570
#>  4: 138.00 139 140.00 141.0 142.00 143 145.00 0.9985918 1021.8261  829.9055
#>  5: 141.00 143 145.00 146.0 147.00 148 152.00 1.0014625 1036.5263  915.7683
#>  6:  99.00 101 102.00 103.0 105.00 106 109.00 1.0008170 1084.2564  888.0770
#>  7:  59.00  61  62.00  63.0  64.00  65  68.00 1.0001596 1170.0810  845.5932
#>  8: 179.00 182 183.00 185.0 186.00 188 192.00 1.0016498  826.3595  989.6300
#>  9: 246.00 251 253.00 256.0 258.00 261 267.00 1.0008653  976.1162  943.8214
#> 10: 256.00 261 264.00 267.0 270.00 274 281.00 1.0020121 1016.2724  988.2011
#> 11: 224.00 229 233.00 235.5 239.00 243 251.00 0.9993738 1119.8109  785.8675
#> 12: 216.00 223 227.00 230.0 234.00 239 247.00 0.9997526 1067.0426  980.9775
#> 13: 150.00 157 161.00 165.0 168.00 173 184.00 0.9989137 1252.9097  836.7735
#> 14: 116.00 122 126.00 129.0 132.00 136 144.00 1.0002835 1111.6597  956.2202
#> 15: 276.00 283 288.00 292.0 297.00 303 316.00 0.9993443 1076.8727  782.2336
#> 16: 271.00 279 286.00 292.0 297.00 305 320.05 1.0000165 1040.5391  767.7477
#> 17: 263.00 275 283.00 291.0 298.35 309 331.00 1.0007083  950.4381  731.4047
#> 18: 251.95 270 282.00 291.0 302.00 318 347.05 1.0031517 1230.4660  982.6347
#> 19: 237.00 269 287.65 307.0 327.00 351 403.15 1.0029817 1076.9075  871.5941
#> 20: 239.95 292 327.00 363.0 401.00 451 576.15 1.0034715 1394.6975  915.4979