Summarise the posterior nowcast prediction

A generic wrapper around enw_posterior() with opinionated defaults to extract the posterior prediction for the nowcast ("pp_inf_obs" from the stan code). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_nowcast_summary(fit, obs, probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95))

Arguments

fit

A cmdstanr fit object.

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

Value

A data.frame summarising the model posterior nowcast prediction. This uses observed data where available and the posterior prediction where not.

See also

Examples

fit <- enw_example("nowcast")
enw_nowcast_summary(fit$fit[[1]], fit$latest[[1]])
#>     reference_date report_date .group max_confirm location age_group confirm
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median        sd     mad  q5
#>  1:                 1    19   0.000000000 149.000  149.0  0.000000  0.0000 149
#>  2:                 1    18   0.000000000 167.453  167.0  1.291301  1.4826 166
#>  3:                 1    17   0.000000000 135.677  135.0  1.853674  1.4826 133
#>  4:                 1    16   0.000000000 140.896  141.0  2.169768  2.9652 138
#>  5:                 1    15   0.007194245 145.222  145.0  2.760133  2.9652 141
#>  6:                 1    14   0.000000000 103.248  103.0  2.893591  2.9652  99
#>  7:                 1    13   0.000000000  62.742   62.0  2.412313  2.9652  59
#>  8:                 1    12   0.000000000 185.065  185.0  3.687054  4.4478 180
#>  9:                 1    11   0.000000000 255.574  255.0  6.079925  5.9304 246
#> 10:                 1    10   0.004219409 266.480  266.0  7.480345  7.4130 255
#> 11:                 1     9   0.000000000 235.082  235.0  7.671123  7.4130 224
#> 12:                 1     8   0.015873016 229.150  229.0  9.532280 10.3782 215
#> 13:                 1     7   0.040000000 163.327  163.0  8.941478  8.8956 149
#> 14:                 1     6   0.010204082 130.404  130.0  8.343284  8.8956 118
#> 15:                 1     5   0.012396694 299.890  299.0 12.795923 13.3434 280
#> 16:                 1     4   0.017937220 303.138  301.0 17.375755 17.0499 278
#> 17:                 1     3   0.019801980 311.131  309.0 24.503638 23.7216 275
#> 18:                 1     2   0.070175439 318.522  315.0 34.923992 34.0998 270
#> 19:                 1     1   0.383928571 334.231  327.5 53.967621 51.1497 258
#> 20:                 1     0   1.000000000 322.430  306.0 93.868922 80.8017 198
#>       q20    q35   q50    q65   q80    q95      rhat  ess_bulk  ess_tail
#>  1: 149.0 149.00 149.0 149.00 149.0 149.00        NA        NA        NA
#>  2: 166.0 167.00 167.0 168.00 168.0 170.00 1.0002481  858.3035  887.5883
#>  3: 134.0 135.00 135.0 136.00 137.0 139.00 0.9986872  784.2638  908.5946
#>  4: 139.0 140.00 141.0 142.00 143.0 145.00 0.9999757  986.1909  904.5245
#>  5: 143.0 144.00 145.0 146.00 147.0 150.00 1.0007676  865.5054  947.2221
#>  6: 101.0 102.00 103.0 104.00 106.0 109.00 0.9991973  893.3750  835.9871
#>  7:  61.0  62.00  62.0  63.00  65.0  67.00 1.0036034  926.5402  861.1900
#>  8: 182.0 183.00 185.0 186.00 188.0 191.00 0.9989872 1097.4089  937.1445
#>  9: 251.0 253.00 255.0 258.00 261.0 266.00 0.9991648 1111.3917 1021.2936
#> 10: 260.0 263.00 266.0 269.00 272.0 279.05 1.0073591 1059.0036  983.0723
#> 11: 228.0 232.00 235.0 237.35 241.0 248.00 1.0012317  975.8115  927.8796
#> 12: 221.0 225.00 229.0 232.00 237.0 246.00 0.9999872 1084.7105  951.3275
#> 13: 155.0 159.00 163.0 166.00 171.0 178.00 0.9994459 1041.2300  974.4463
#> 14: 123.0 127.00 130.0 133.00 137.0 145.00 1.0012796  935.7652 1050.2310
#> 15: 289.0 294.00 299.0 304.00 310.2 322.00 1.0017171 1130.4067  958.2443
#> 16: 288.8 295.00 301.0 308.00 317.0 334.05 1.0055266  848.0051  763.8073
#> 17: 291.0 300.65 309.0 317.35 329.0 356.00 1.0105072 1083.0271 1028.2641
#> 18: 288.8 303.00 315.0 330.00 343.2 383.00 1.0047446 1063.6034  939.8372
#> 19: 288.8 308.00 327.5 347.00 375.0 436.00 1.0005068 1007.8998  874.2774
#> 20: 247.8 276.00 306.0 340.35 390.2 493.05 1.0003640 1366.8407  840.3280