Summarise the posterior nowcast prediction

A generic wrapper around enw_posterior() with opinionated defaults to extract the posterior prediction for the nowcast ("pp_inf_obs" from the stan code). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_nowcast_summary(
  fit,
  obs,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  timestep = "day"
)

Arguments

fit

A cmdstanr fit object.

obs

An observation data.frame containing reference_date columns of the same length as the number of rows in the posterior and the most up to date observation for each date. This is used to align the posterior with the observations. The easiest source of this data is the output of latest output of enw_preprocess_data() or enw_latest_data().

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

timestep

The timestep to used. This can be a string ("day", "week", "month") or a numeric whole number representing the number of days.

Value

A data.frame summarising the model posterior nowcast prediction. This uses observed data where available and the posterior prediction where not.

See also

Examples

fit <- enw_example("nowcast")
enw_nowcast_summary(fit$fit[[1]], fit$latest[[1]])
#>     reference_date report_date .group max_confirm location age_group confirm
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>  1:                 1    19   0.000000000 149.000  149.0   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.478  167.0   1.354012  1.4826
#>  3:                 1    17   0.000000000 135.796  136.0   1.844393  1.4826
#>  4:                 1    16   0.000000000 141.299  141.0   2.436706  2.2239
#>  5:                 1    15   0.007194245 145.864  146.0   3.079771  2.9652
#>  6:                 1    14   0.000000000 103.719  103.0   3.084080  2.9652
#>  7:                 1    13   0.000000000  62.823   63.0   2.365855  2.9652
#>  8:                 1    12   0.000000000 185.185  185.0   3.830855  4.4478
#>  9:                 1    11   0.000000000 256.267  256.0   6.546187  5.9304
#> 10:                 1    10   0.004219409 267.641  267.0   7.781560  7.4130
#> 11:                 1     9   0.000000000 236.055  235.0   8.434163  7.4130
#> 12:                 1     8   0.015873016 230.953  231.0   9.894883 10.3782
#> 13:                 1     7   0.040000000 164.830  164.0   9.288131  8.8956
#> 14:                 1     6   0.010204082 129.266  129.0   8.201250  7.4130
#> 15:                 1     5   0.012396694 293.081  292.0  11.891503 11.8608
#> 16:                 1     4   0.017937220 293.110  291.5  15.184152 14.0847
#> 17:                 1     3   0.019801980 291.826  289.0  21.584476 20.7564
#> 18:                 1     2   0.070175439 294.681  291.0  29.825640 29.6520
#> 19:                 1     1   0.383928571 311.024  305.0  52.222856 47.4432
#> 20:                 1     0   1.000000000 378.046  361.0 112.730351 99.3342
#>         q5   q20 q35   q50    q65   q80    q95      rhat  ess_bulk  ess_tail
#>  1: 149.00 149.0 149 149.0 149.00 149.0 149.00        NA        NA        NA
#>  2: 166.00 166.0 167 167.0 168.00 168.2 170.00 1.0000029  989.6437  972.4510
#>  3: 133.00 134.0 135 136.0 136.00 137.0 139.00 1.0014191  936.7059  776.1750
#>  4: 138.00 139.0 140 141.0 142.00 143.0 146.00 1.0006007  940.7707  993.4910
#>  5: 142.00 143.0 144 146.0 147.00 148.0 151.05 0.9996456 1119.1434  979.5307
#>  6:  99.00 101.0 102 103.0 105.00 106.0 109.00 0.9999319  986.4520  879.3803
#>  7:  59.00  61.0  62  63.0  64.00  65.0  67.00 1.0023759  948.6815  955.5744
#>  8: 180.00 182.0 183 185.0 186.00 188.0 192.00 0.9996973  920.3613  870.9005
#>  9: 246.00 251.0 253 256.0 258.00 262.0 268.00 0.9984765  943.1807  931.9972
#> 10: 256.00 261.0 264 267.0 270.00 274.0 281.05 0.9998729  966.7512  899.5473
#> 11: 223.00 229.0 232 235.0 239.00 243.0 251.00 0.9985037  880.5708  958.2759
#> 12: 216.00 222.0 226 231.0 234.00 239.0 248.00 1.0046207 1035.1435  907.6846
#> 13: 151.00 157.0 161 164.0 168.00 173.0 181.00 1.0027599  947.9732 1012.4335
#> 14: 117.00 122.8 126 129.0 132.00 135.0 144.00 1.0014890  983.9472  977.1206
#> 15: 275.00 283.0 288 292.0 297.00 302.0 314.05 1.0000113 1074.1573  815.9249
#> 16: 271.95 281.0 286 291.5 298.00 306.0 320.00 1.0016347 1009.7806 1062.5233
#> 17: 261.00 273.0 281 289.0 298.00 310.0 330.05 1.0009863 1130.7556  986.6337
#> 18: 251.00 269.8 281 291.0 304.00 319.0 347.00 0.9992332 1122.6892  797.0210
#> 19: 237.00 267.0 288 305.0 323.00 349.0 404.05 0.9994896 1028.4657  890.2360
#> 20: 228.00 288.0 323 361.0 404.35 456.4 575.20 1.0036345 1170.1286  946.0206