Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean       median        sd       mad         q5
#>            <char>       <num>        <num>     <num>     <num>      <num>
#>  1:  expr_beta[1] -0.32019135 -0.313468545 0.4719893 0.4663760 -1.0901889
#>  2:  expr_beta[2] -0.74038314 -0.746882835 0.4829204 0.4788528 -1.5245174
#>  3:  expr_beta[3]  0.47401653  0.476436800 0.4999795 0.4899696 -0.3409476
#>  4:  expr_beta[4] -0.53425310 -0.541361975 0.4991747 0.4675105 -1.3446587
#>  5:  expr_beta[5] -0.35283471 -0.341547660 0.5165949 0.5278692 -1.2310439
#>  6:  expr_beta[6]  1.77518278  1.754966550 0.5060366 0.5130603  0.9927312
#>  7:  expr_beta[7]  0.22836379  0.217275465 0.4508645 0.4535358 -0.4862596
#>  8:  expr_beta[8]  0.18068041  0.193796645 0.4640738 0.4661506 -0.5669878
#>  9:  expr_beta[9] -0.25700065 -0.232491580 0.4639707 0.4396751 -1.0507673
#> 10: expr_beta[10] -0.05452652 -0.050068295 0.4470491 0.4585587 -0.7697181
#> 11: expr_beta[11] -0.60611875 -0.588781615 0.4525201 0.4402069 -1.3719542
#> 12: expr_beta[12] -1.43996538 -1.429942450 0.4856450 0.5108136 -2.2526915
#> 13: expr_beta[13]  1.65605524  1.650733550 0.5083926 0.4807774  0.8648800
#> 14: expr_beta[14]  1.78115178  1.769585200 0.4660973 0.4446532  1.0340851
#> 15: expr_beta[15] -0.26055568 -0.254206415 0.3928030 0.3663731 -0.9213899
#> 16: expr_beta[16] -0.24088953 -0.230386475 0.4200619 0.4002978 -0.9331860
#> 17: expr_beta[17] -0.04674257 -0.042740430 0.4426445 0.4647786 -0.7711823
#> 18: expr_beta[18] -0.79005255 -0.770506800 0.4135231 0.4122316 -1.5107553
#> 19: expr_beta[19] -0.50469051 -0.525717030 0.4466045 0.4680735 -1.2246207
#> 20: expr_beta[20]  1.85365959  1.841799600 0.4635994 0.4710307  1.1299419
#> 21: expr_beta[21]  0.44523352  0.437180875 0.4101772 0.3906846 -0.2060474
#> 22: expr_beta[22] -0.40155226 -0.396541450 0.4049779 0.3855741 -1.0797208
#> 23: expr_beta[23] -0.14751856 -0.138619460 0.3947424 0.4015433 -0.7782497
#> 24: expr_beta[24]  0.01524095  0.022815620 0.4003055 0.3981528 -0.6282335
#> 25: expr_beta[25] -0.89662706 -0.880105220 0.4300562 0.3976072 -1.6441800
#> 26: expr_beta[26] -0.88534784 -0.868967455 0.4855118 0.4726506 -1.7002608
#> 27: expr_beta[27]  2.19272602  2.161698200 0.5120513 0.4946392  1.3246077
#> 28: expr_beta[28]  1.40346138  1.382971550 0.4346950 0.4176852  0.7150258
#> 29: expr_beta[29]  0.03935298  0.033121206 0.3872781 0.3845059 -0.6045337
#> 30: expr_beta[30] -0.48027994 -0.476700975 0.3883970 0.3696986 -1.1111875
#> 31: expr_beta[31] -0.14443150 -0.133373980 0.4004516 0.3702782 -0.7924091
#> 32: expr_beta[32] -0.73859154 -0.730767310 0.4013771 0.3925696 -1.4601488
#> 33: expr_beta[33] -0.37664696 -0.379308865 0.4548716 0.4446331 -1.1258026
#> 34: expr_beta[34]  1.75065755  1.751819650 0.4767639 0.4624977  0.9655303
#> 35: expr_beta[35]  0.16176505  0.131223590 0.4356223 0.4371837 -0.4987789
#> 36: expr_beta[36]  0.01565616  0.003472225 0.4654906 0.4670889 -0.7248555
#> 37: expr_beta[37]  0.03007161  0.019158597 0.5086772 0.4935142 -0.8295010
#> 38: expr_beta[38]  0.12404208  0.119429835 0.5631639 0.5364774 -0.7775894
#> 39: expr_beta[39] -0.03548987 -0.052523940 0.6687768 0.6601507 -1.0990928
#>          variable        mean       median        sd       mad         q5
#>            <char>       <num>        <num>     <num>     <num>      <num>
#>             q20          q80         q95      rhat  ess_bulk ess_tail
#>           <num>        <num>       <num>     <num>     <num>    <num>
#>  1: -0.71010843  0.058567508  0.41945177 1.0007712  835.6741 561.3815
#>  2: -1.16253746 -0.326596042  0.04203365 1.0025927  880.0079 845.4472
#>  3:  0.06557318  0.890336242  1.30881321 1.0055579  991.1887 674.8935
#>  4: -0.94538736 -0.117524678  0.26939569 0.9989676 1032.5903 630.5401
#>  5: -0.77596812  0.092587062  0.50968498 1.0014722  984.5286 772.9448
#>  6:  1.36004822  2.193550580  2.63890411 1.0024238 1016.8840 614.4088
#>  7: -0.14832090  0.599790666  0.97021847 1.0006250 1126.7534 663.9769
#>  8: -0.21380562  0.570561918  0.91447361 1.0008365 1328.5046 665.5682
#>  9: -0.61267389  0.119263988  0.45624515 1.0034953 1025.9958 734.6290
#> 10: -0.44583336  0.312164210  0.67065326 1.0018849  710.7928 681.1261
#> 11: -0.98975035 -0.242347394  0.11701029 1.0015651 1036.2840 699.5415
#> 12: -1.84692536 -1.026897720 -0.65122445 1.0038096  921.8887 821.6462
#> 13:  1.25483994  2.053339140  2.51297444 1.0018836 1055.2115 770.1263
#> 14:  1.39904528  2.172014700  2.54832912 1.0004343  929.0432 469.2282
#> 15: -0.58731034  0.044126859  0.37542706 1.0143370  910.6557 631.2003
#> 16: -0.59132504  0.101614036  0.43093022 0.9999458  968.4144 736.3714
#> 17: -0.41391116  0.348147508  0.65707408 1.0008488 1060.3276 652.0607
#> 18: -1.12859854 -0.444877930 -0.13630677 1.0038297  925.5502 785.7024
#> 19: -0.88929617 -0.106097164  0.23734938 0.9995649 1079.3438 730.2426
#> 20:  1.47099142  2.251628080  2.62457361 1.0012680  754.2583 800.1464
#> 21:  0.09955755  0.771813376  1.14975883 1.0051884 1213.7036 754.2567
#> 22: -0.72976544 -0.065233724  0.27740958 1.0010283 1175.9185 699.5991
#> 23: -0.47369388  0.177560254  0.50205920 1.0046186 1186.0125 654.5199
#> 24: -0.31379081  0.341613922  0.67991934 1.0010743 1220.2946 558.3089
#> 25: -1.23488934 -0.552364908 -0.20616873 1.0037727  952.0349 680.4777
#> 26: -1.29223532 -0.493947396 -0.12037232 1.0017135  815.8363 652.5990
#> 27:  1.78047748  2.631200060  3.07574713 1.0010461  830.8996 650.3065
#> 28:  1.03609534  1.750644160  2.16905649 1.0022789  770.2758 755.5526
#> 29: -0.27115932  0.372319078  0.67183833 0.9988493  956.0999 602.3863
#> 30: -0.79442503 -0.161764074  0.15704531 1.0039528  913.0985 561.7087
#> 31: -0.46259570  0.184354710  0.51896905 1.0009000 1045.6107 678.0340
#> 32: -1.04543758 -0.386321636 -0.09256058 1.0066655 1094.4003 567.5733
#> 33: -0.74034924 -0.006286013  0.35990866 1.0064612  853.8593 489.5167
#> 34:  1.35269714  2.135911400  2.57496522 1.0040032  829.9975 689.4263
#> 35: -0.20950913  0.534516186  0.90324470 1.0006855 1083.7066 747.8989
#> 36: -0.38824638  0.395963932  0.77578717 1.0005970 1073.5696 585.2625
#> 37: -0.38240918  0.445608564  0.89886011 1.0004866 1099.0376 825.1805
#> 38: -0.30792165  0.572560448  1.01445988 0.9993830 1144.8665 605.4792
#> 39: -0.60550686  0.506056744  1.09178222 1.0045362 1272.0269 695.2699
#>             q20          q80         q95      rhat  ess_bulk ess_tail
#>           <num>        <num>       <num>     <num>     <num>    <num>