Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean      median        sd       mad         q5
#>  1:  expr_beta[1]  0.32457432  0.31099550 0.5645264 0.5696865 -0.5677573
#>  2:  expr_beta[2] -0.33302066 -0.33591150 0.5381189 0.5467934 -1.1882865
#>  3:  expr_beta[3] -0.80486524 -0.80434850 0.5637706 0.5332460 -1.7023055
#>  4:  expr_beta[4]  0.47590999  0.46739600 0.5280723 0.5377813 -0.4094209
#>  5:  expr_beta[5] -0.47896320 -0.48517550 0.5438857 0.5514056 -1.3680625
#>  6:  expr_beta[6] -0.22947987 -0.23216000 0.5573074 0.5477140 -1.1357110
#>  7:  expr_beta[7]  1.70172299  1.71357500 0.5549077 0.5691331  0.7911440
#>  8:  expr_beta[8]  0.35894785  0.33959750 0.5306142 0.5163593 -0.5137593
#>  9:  expr_beta[9]  0.23435790  0.23426150 0.5341069 0.5285617 -0.6586791
#> 10: expr_beta[10] -0.34426439 -0.34075200 0.5231955 0.5513084 -1.1748875
#> 11: expr_beta[11] -0.16960673 -0.16708950 0.5187156 0.4991506 -1.0260180
#> 12: expr_beta[12] -0.56626942 -0.54756000 0.5080131 0.5054465 -1.4437395
#> 13: expr_beta[13] -1.20873598 -1.18783500 0.5478744 0.5551966 -2.1671070
#> 14: expr_beta[14]  1.51316188  1.50818000 0.5646819 0.5682065  0.5919151
#> 15: expr_beta[15]  1.78787993  1.78017500 0.5441534 0.4922825  0.9068597
#> 16: expr_beta[16] -0.22721866 -0.22783900 0.4720398 0.4981158 -1.0420520
#> 17: expr_beta[17] -0.23041097 -0.22539750 0.4930500 0.4951350 -1.0229105
#> 18: expr_beta[18] -0.19327557 -0.18214050 0.4698171 0.4687425 -0.9901994
#> 19: expr_beta[19] -0.54966432 -0.55493700 0.4631250 0.4568061 -1.3375335
#> 20: expr_beta[20] -0.33606753 -0.34552550 0.5122267 0.5149909 -1.1758645
#> 21: expr_beta[21]  1.65282279  1.62484000 0.5515100 0.5342253  0.7995906
#> 22: expr_beta[22]  0.41085922  0.38758300 0.4944919 0.4824506 -0.3638883
#> 23: expr_beta[23] -0.29539851 -0.27766750 0.4628180 0.4362411 -1.1087000
#> 24: expr_beta[24] -0.33220604 -0.31648050 0.4657746 0.4673935 -1.1021165
#> 25: expr_beta[25] -0.09979849 -0.10645800 0.4647000 0.4436814 -0.8542074
#> 26: expr_beta[26] -0.86578167 -0.85861450 0.4778653 0.4842668 -1.6395145
#> 27: expr_beta[27] -0.63014717 -0.62152500 0.5394971 0.5423736 -1.5336220
#> 28: expr_beta[28]  2.05061451  2.03900000 0.5691224 0.5644851  1.1813515
#> 29: expr_beta[29]  1.61983323  1.58360000 0.5229551 0.5135801  0.8326538
#> 30: expr_beta[30]  0.08737375  0.06453220 0.4644408 0.4368889 -0.6309870
#> 31: expr_beta[31] -0.61096245 -0.60194200 0.4774027 0.4855033 -1.4044255
#> 32: expr_beta[32] -0.26292769 -0.26390750 0.4789259 0.4482092 -1.0303225
#> 33: expr_beta[33] -0.63810873 -0.62634300 0.5066253 0.4989913 -1.4842600
#> 34: expr_beta[34] -0.14185639 -0.14035100 0.5349006 0.4966651 -1.0357070
#> 35: expr_beta[35]  1.46544003  1.44865000 0.5591331 0.5458414  0.5730308
#> 36: expr_beta[36]  0.16609941  0.16173400 0.5568830 0.5433596 -0.7456388
#> 37: expr_beta[37]  0.01886907  0.01027121 0.5724857 0.5639176 -0.9294588
#> 38: expr_beta[38]  0.09991892  0.10889800 0.6075559 0.5903735 -0.9179446
#> 39: expr_beta[39]  0.19172175  0.18276600 0.6642032 0.6520201 -0.9213319
#> 40: expr_beta[40]  0.36361993  0.36345950 0.7706377 0.7547672 -0.9022501
#>          variable        mean      median        sd       mad         q5
#>              q20         q80         q95      rhat  ess_bulk ess_tail
#>  1: -0.136141600  0.80692140  1.27757900 0.9988082 1108.9443 755.6332
#>  2: -0.793877400  0.11581040  0.54702325 0.9986453 1450.3561 852.1223
#>  3: -1.294246000 -0.32721480  0.10426610 1.0026340 1373.6902 948.7417
#>  4:  0.030596120  0.92586160  1.34165400 1.0009814 1214.7480 880.2151
#>  5: -0.942981200 -0.02479340  0.42608590 1.0091304 1179.8479 713.5425
#>  6: -0.703136800  0.20665280  0.71433600 1.0013455 1248.0921 857.1823
#>  7:  1.235270000  2.17079600  2.60301550 1.0015553 1108.5963 622.2671
#>  8: -0.081404880  0.80810640  1.22218700 1.0001607 1183.2485 651.4696
#>  9: -0.189486200  0.68962420  1.11916700 1.0070877 1655.5854 828.9620
#> 10: -0.822558800  0.11156920  0.57136345 1.0006784 1219.9232 665.7028
#> 11: -0.588090200  0.26638320  0.64708830 1.0025242 1171.7353 780.3423
#> 12: -0.978719000 -0.14531380  0.23348870 1.0036211 1297.3555 863.6776
#> 13: -1.668116000 -0.75293860 -0.29711590 1.0058026 1088.5952 758.1927
#> 14:  1.023372000  1.97502800  2.43391950 1.0007559  944.4527 684.0765
#> 15:  1.371042000  2.20963400  2.67874500 0.9999402  981.8985 707.3602
#> 16: -0.628712400  0.18389380  0.53730040 1.0024671 1374.8828 814.5561
#> 17: -0.657083400  0.18605160  0.57336560 1.0029970 1194.7481 779.0059
#> 18: -0.590000400  0.18691660  0.53116640 1.0039788 1466.8872 806.3921
#> 19: -0.938454200 -0.15676700  0.20032325 1.0004648 1370.9410 909.0425
#> 20: -0.778328400  0.08551790  0.50958830 1.0031855 1250.6961 617.8546
#> 21:  1.202362000  2.11126000  2.55703250 0.9988434 1393.6240 551.4185
#> 22:  0.004064258  0.81757900  1.24409150 0.9988965 1160.3880 732.9565
#> 23: -0.668138400  0.07099878  0.44281460 1.0003619  944.9087 524.8983
#> 24: -0.721107000  0.04275818  0.41663475 0.9993887 1012.1569 779.4768
#> 25: -0.477020600  0.28416940  0.65423135 1.0007390  895.9667 557.7331
#> 26: -1.281252000 -0.46335840 -0.05739187 1.0060605 1013.0612 747.5057
#> 27: -1.078326000 -0.16179480  0.20551740 1.0089392  949.0663 761.4667
#> 28:  1.568542000  2.52828000  3.01045300 1.0032705 1120.1935 611.3021
#> 29:  1.190700000  2.04668800  2.52273700 1.0059037  943.7398 478.8401
#> 30: -0.298242400  0.45113340  0.93601070 1.0011549 1088.4195 799.2504
#> 31: -1.030416000 -0.20060700  0.16069040 1.0013267 1177.6426 880.9254
#> 32: -0.645288800  0.11523120  0.51691720 0.9999005 1452.0304 792.9190
#> 33: -1.076654000 -0.24118520  0.19523890 1.0019986 1245.8648 644.8195
#> 34: -0.564462000  0.27911940  0.74949545 0.9988417 1535.9467 675.2636
#> 35:  1.002228000  1.90363800  2.41435900 1.0064370 1369.4019 703.1286
#> 36: -0.290180600  0.60540900  1.12348950 1.0011714 1558.4254 828.0634
#> 37: -0.456937400  0.51018400  0.95001770 1.0000028 1385.7004 671.9348
#> 38: -0.395061800  0.61097420  1.08651650 1.0011711 1416.8174 797.0406
#> 39: -0.367529800  0.75110360  1.28875800 1.0067156 1165.0971 741.1174
#> 40: -0.271500400  0.99036940  1.68385400 0.9989737 1778.0469 809.2057
#>              q20         q80         q95      rhat  ess_bulk ess_tail