Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean      median        sd       mad         q5
#>  1:  expr_beta[1]  0.35691445  0.36498800 0.5764547 0.5484506 -0.5943839
#>  2:  expr_beta[2] -0.34761195 -0.32384900 0.5760045 0.5536463 -1.2701980
#>  3:  expr_beta[3] -0.83816753 -0.83024800 0.5693749 0.5711331 -1.8262760
#>  4:  expr_beta[4]  0.51801887  0.52348450 0.5678107 0.5556399 -0.3951924
#>  5:  expr_beta[5] -0.49017229 -0.49719900 0.5278500 0.4999846 -1.3714010
#>  6:  expr_beta[6] -0.23446383 -0.21519950 0.5755229 0.5527518 -1.2024595
#>  7:  expr_beta[7]  1.72076461  1.70401500 0.5556995 0.5449370  0.8715682
#>  8:  expr_beta[8]  0.34962057  0.34943250 0.5410623 0.5203097 -0.5621784
#>  9:  expr_beta[9]  0.23732838  0.23436150 0.5341452 0.5232866 -0.6420201
#> 10: expr_beta[10] -0.34902716 -0.34409350 0.5217350 0.5295615 -1.1965940
#> 11: expr_beta[11] -0.17188953 -0.16680400 0.4845065 0.4743660 -1.0095885
#> 12: expr_beta[12] -0.55853461 -0.54901350 0.4921987 0.4969349 -1.3611635
#> 13: expr_beta[13] -1.23213415 -1.23010500 0.5682347 0.5755424 -2.1864280
#> 14: expr_beta[14]  1.52471513  1.52728000 0.6255768 0.6206238  0.4838366
#> 15: expr_beta[15]  1.80820361  1.80485500 0.5600774 0.5653524  0.9292188
#> 16: expr_beta[16] -0.22058912 -0.22138500 0.4908559 0.4650221 -1.0230965
#> 17: expr_beta[17] -0.25068399 -0.25479950 0.4912554 0.4890450 -1.0482845
#> 18: expr_beta[18] -0.17175111 -0.16917250 0.4886694 0.5068935 -0.9776898
#> 19: expr_beta[19] -0.56449312 -0.55253050 0.4844249 0.4558706 -1.3610185
#> 20: expr_beta[20] -0.33829765 -0.34597550 0.5274359 0.5072512 -1.1949605
#> 21: expr_beta[21]  1.68238934  1.67413500 0.5279810 0.5121642  0.8400429
#> 22: expr_beta[22]  0.41138546  0.41694050 0.4848998 0.4530284 -0.4269672
#> 23: expr_beta[23] -0.28407707 -0.27718050 0.4429254 0.4290170 -1.0027790
#> 24: expr_beta[24] -0.33421052 -0.32154200 0.4541127 0.4562668 -1.0654940
#> 25: expr_beta[25] -0.12360429 -0.13059150 0.4913108 0.5112924 -0.9331567
#> 26: expr_beta[26] -0.87457779 -0.87055900 0.5120680 0.4747708 -1.7559085
#> 27: expr_beta[27] -0.65372879 -0.64744800 0.5288792 0.5360251 -1.4965265
#> 28: expr_beta[28]  2.09713455  2.08991000 0.5327654 0.5280651  1.2489815
#> 29: expr_beta[29]  1.62818297  1.60410000 0.5205884 0.4904886  0.7600817
#> 30: expr_beta[30]  0.11566660  0.14057550 0.4731800 0.4516771 -0.6606485
#> 31: expr_beta[31] -0.65684394 -0.65804050 0.4933335 0.4917769 -1.4553930
#> 32: expr_beta[32] -0.22655365 -0.21496750 0.4812528 0.4896998 -1.0142790
#> 33: expr_beta[33] -0.65826420 -0.66979550 0.5243565 0.5240576 -1.5382910
#> 34: expr_beta[34] -0.13190444 -0.13824350 0.5499114 0.5404900 -1.0432585
#> 35: expr_beta[35]  1.47507558  1.46098000 0.5497911 0.5541440  0.5757841
#> 36: expr_beta[36]  0.16886656  0.16426950 0.5513899 0.5517659 -0.7193582
#> 37: expr_beta[37]  0.03530783  0.04368235 0.5600475 0.5563963 -0.8796558
#> 38: expr_beta[38]  0.12571370  0.12635600 0.6185879 0.6069401 -0.8701020
#> 39: expr_beta[39]  0.16884623  0.17235550 0.6772904 0.6992572 -0.9112642
#> 40: expr_beta[40]  0.37352032  0.35352300 0.7936915 0.7642136 -0.8954799
#>          variable        mean      median        sd       mad         q5
#>             q20         q80          q95      rhat  ess_bulk ess_tail
#>  1: -0.11922640  0.83821680  1.310382000 1.0010635  660.0161 535.7780
#>  2: -0.81108380  0.11612020  0.593448150 1.0026682 1335.8326 740.6334
#>  3: -1.33949200 -0.35110620  0.070130075 1.0016985 1163.2292 789.4890
#>  4:  0.04686534  0.99532180  1.444933500 1.0018444 1102.4074 879.6314
#>  5: -0.91605840 -0.07005216  0.404101100 1.0004425 1113.6900 668.9304
#>  6: -0.71488600  0.22962420  0.693560750 1.0056901  996.5392 648.7603
#>  7:  1.26151800  2.17562000  2.652352000 1.0037668  954.4084 652.8714
#>  8: -0.08204750  0.80780620  1.245801000 1.0008162 1020.0426 902.6647
#>  9: -0.20514160  0.65418700  1.131625000 1.0037875  968.1302 796.1627
#> 10: -0.77210180  0.09092864  0.491583300 1.0037540 1141.1396 732.7558
#> 11: -0.55306220  0.22969240  0.592171100 1.0019680 1318.3628 809.8154
#> 12: -0.96768420 -0.13333260  0.250620500 1.0014743 1134.6824 914.3102
#> 13: -1.69079400 -0.73640560 -0.358825500 1.0030722  946.6902 795.5025
#> 14:  0.99659740  2.04084800  2.598101000 1.0069396 1075.2809 742.3032
#> 15:  1.31686200  2.27177800  2.741562000 1.0015112  951.8782 859.8358
#> 16: -0.61209140  0.19275560  0.572072800 1.0028789 1225.7750 773.4261
#> 17: -0.63679360  0.16880200  0.560836600 1.0029651  910.8575 766.3416
#> 18: -0.58390260  0.28426180  0.607000950 1.0018456 1247.1149 840.7685
#> 19: -0.94331960 -0.17481360  0.193365300 1.0014472 1189.8584 862.9854
#> 20: -0.77095000  0.09148314  0.525546600 0.9982799  962.2272 772.9448
#> 21:  1.24051200  2.10356000  2.606347000 1.0008667 1313.8286 877.8272
#> 22:  0.02028172  0.82299060  1.200706500 1.0037534 1000.4386 606.5635
#> 23: -0.64170760  0.09441038  0.404700000 1.0013013 1002.5619 815.0601
#> 24: -0.71855220  0.04867070  0.415618250 0.9996585 1265.1278 939.4297
#> 25: -0.55576900  0.29392660  0.668455550 1.0002861 1063.4436 663.0799
#> 26: -1.28327200 -0.46753760 -0.002321493 0.9985271 1289.3440 629.9431
#> 27: -1.10884800 -0.19467660  0.190780650 1.0019952 1009.7123 799.0850
#> 28:  1.63177400  2.53894600  3.005590000 1.0002712  763.4697 662.5946
#> 29:  1.18957800  2.04650400  2.504602000 0.9990166 1120.2857 680.8118
#> 30: -0.30102080  0.48244940  0.882347250 1.0010978 1201.0119 760.9960
#> 31: -1.05982800 -0.23950340  0.190652200 1.0006893  951.3861 887.4073
#> 32: -0.63355320  0.19755480  0.523404300 0.9985643 1172.4544 846.3230
#> 33: -1.07347800 -0.20559440  0.178272250 0.9991165 1298.1225 750.9543
#> 34: -0.56258620  0.32878740  0.755459800 1.0003530 1239.1159 681.0165
#> 35:  1.01486000  1.92881000  2.390124000 1.0008108  943.7494 618.1981
#> 36: -0.28282940  0.61264040  1.096016500 0.9996861  896.6749 732.1576
#> 37: -0.43939000  0.50305020  0.938247200 0.9997439  848.4870 841.4673
#> 38: -0.39462300  0.61521820  1.165504000 0.9998664  995.9477 724.1238
#> 39: -0.41606460  0.73891440  1.259385500 1.0009455 1395.7144 812.5458
#> 40: -0.27644920  1.03718400  1.698252500 1.0003707 1172.9385 744.6613
#>             q20         q80          q95      rhat  ess_bulk ess_tail