Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean      median        sd       mad         q5
#>  1:  expr_beta[1]  0.30921632  0.30965250 0.5605897 0.5633116 -0.6321378
#>  2:  expr_beta[2] -0.29880768 -0.30701750 0.5323509 0.5290411 -1.1575990
#>  3:  expr_beta[3] -0.80954527 -0.78988050 0.5457703 0.5661493 -1.7460535
#>  4:  expr_beta[4]  0.46714165  0.46400700 0.5408664 0.5416789 -0.4190951
#>  5:  expr_beta[5] -0.49467562 -0.49509500 0.5684288 0.5881498 -1.3897840
#>  6:  expr_beta[6] -0.23932038 -0.24144300 0.6015006 0.6014189 -1.2733970
#>  7:  expr_beta[7]  1.70810250  1.68871500 0.6097804 0.6073842  0.7575587
#>  8:  expr_beta[8]  0.34387857  0.34234250 0.5524105 0.5665278 -0.5956675
#>  9:  expr_beta[9]  0.25769668  0.26080050 0.5209613 0.5464056 -0.5930261
#> 10: expr_beta[10] -0.36115848 -0.35441050 0.5196194 0.5246912 -1.2383660
#> 11: expr_beta[11] -0.14261141 -0.14498700 0.5409863 0.5237848 -1.0081395
#> 12: expr_beta[12] -0.60116797 -0.59326500 0.5135157 0.4963730 -1.4483605
#> 13: expr_beta[13] -1.18987964 -1.20564000 0.5718288 0.5918784 -2.1076125
#> 14: expr_beta[14]  1.51341508  1.53526500 0.5902728 0.6027140  0.5249397
#> 15: expr_beta[15]  1.81099202  1.76425500 0.5552335 0.5393402  0.9243832
#> 16: expr_beta[16] -0.23318704 -0.23760400 0.5184560 0.5112072 -1.0894010
#> 17: expr_beta[17] -0.22654094 -0.23183300 0.5083324 0.5080663 -1.0664380
#> 18: expr_beta[18] -0.18313643 -0.20097200 0.5002845 0.5115719 -0.9912695
#> 19: expr_beta[19] -0.57670042 -0.59132450 0.4867117 0.4572264 -1.4040605
#> 20: expr_beta[20] -0.30974050 -0.29535150 0.5467479 0.5360563 -1.2080640
#> 21: expr_beta[21]  1.61957481  1.60733500 0.5256980 0.5338472  0.7880852
#> 22: expr_beta[22]  0.44790562  0.42086450 0.4934454 0.4638329 -0.3252101
#> 23: expr_beta[23] -0.27979669 -0.27792800 0.4784463 0.4615897 -1.0926040
#> 24: expr_beta[24] -0.36690516 -0.36518400 0.4903180 0.4393575 -1.2077230
#> 25: expr_beta[25] -0.07390903 -0.07730095 0.4511930 0.4291216 -0.8041371
#> 26: expr_beta[26] -0.89567792 -0.88663200 0.4923413 0.4714905 -1.6911695
#> 27: expr_beta[27] -0.65548125 -0.64203700 0.5608036 0.5726068 -1.5567150
#> 28: expr_beta[28]  2.10704814  2.07448000 0.5746041 0.5754490  1.2269450
#> 29: expr_beta[29]  1.61495468  1.61417500 0.5295164 0.5458637  0.7682298
#> 30: expr_beta[30]  0.11419270  0.09957430 0.4833925 0.4586082 -0.6314293
#> 31: expr_beta[31] -0.64690139 -0.65090450 0.5050231 0.5021477 -1.5025320
#> 32: expr_beta[32] -0.24423133 -0.24732550 0.4877905 0.4908431 -1.0529215
#> 33: expr_beta[33] -0.65281563 -0.62133800 0.5231371 0.5377590 -1.5094985
#> 34: expr_beta[34] -0.12705089 -0.12465750 0.5529788 0.5544642 -0.9986610
#> 35: expr_beta[35]  1.47106339  1.46909500 0.6030544 0.5824246  0.4842683
#> 36: expr_beta[36]  0.17286987  0.19988800 0.5604678 0.5529957 -0.7788352
#> 37: expr_beta[37]  0.05488720  0.05732775 0.6127427 0.6412797 -0.9688827
#> 38: expr_beta[38]  0.05829442  0.05740225 0.6029646 0.5847656 -0.9612945
#> 39: expr_beta[39]  0.18933237  0.16977700 0.6419841 0.6433061 -0.8415081
#> 40: expr_beta[40]  0.42423792  0.43613100 0.7906812 0.7630816 -0.8660402
#>          variable        mean      median        sd       mad         q5
#>              q20          q80         q95      rhat  ess_bulk ess_tail
#>  1: -0.154041000  0.783442200  1.19796300 1.0016007  922.0327 845.6659
#>  2: -0.737969600  0.134063000  0.56264515 1.0025292 1246.1983 616.9065
#>  3: -1.271726000 -0.347359000  0.07293548 1.0003989 1357.8097 831.1805
#>  4:  0.004766122  0.947309600  1.34425400 1.0012996 1217.0440 799.8847
#>  5: -0.990739200 -0.007058358  0.38090075 1.0021169 1210.6133 874.3571
#>  6: -0.739323600  0.282726200  0.71168410 0.9997518 1308.9179 526.8150
#>  7:  1.192698000  2.222740000  2.73101850 1.0011415 1405.8953 795.6888
#>  8: -0.110576800  0.831092000  1.23000450 0.9994161 1376.2157 829.1273
#>  9: -0.194165800  0.719065400  1.10173950 1.0070148 1477.4667 757.9880
#> 10: -0.821577000  0.083427520  0.45892260 1.0054224 1259.7922 845.7388
#> 11: -0.582931200  0.295951600  0.71318245 1.0100257 1259.2233 641.3915
#> 12: -1.028688000 -0.184618800  0.21756940 1.0014092  989.6907 700.2856
#> 13: -1.676586000 -0.725648800 -0.26848555 1.0007362  974.7010 608.8522
#> 14:  1.013742000  2.013630000  2.46039750 1.0009544 1076.1238 710.4487
#> 15:  1.351862000  2.300374000  2.74645100 1.0054567 1039.0789 770.9567
#> 16: -0.684064400  0.178737400  0.61317025 1.0013182 1129.7146 516.4527
#> 17: -0.647028400  0.213193800  0.62161885 1.0067373 1019.9608 676.1481
#> 18: -0.591701800  0.253172400  0.66015770 1.0045544 1368.0402 813.9703
#> 19: -0.963655800 -0.192352800  0.26299260 1.0019134 1105.0084 605.0290
#> 20: -0.769458600  0.131582800  0.58707095 1.0054156 1031.6844 732.8707
#> 21:  1.164680000  2.075858000  2.50006300 1.0046183 1036.0434 796.0915
#> 22:  0.052008640  0.849966200  1.33677650 0.9997362  848.8552 635.6972
#> 23: -0.661620600  0.131426800  0.50771435 1.0026946 1135.7843 724.4714
#> 24: -0.740454600  0.016016400  0.43012365 0.9990784 1188.7051 724.9150
#> 25: -0.433719600  0.292346800  0.68376430 1.0098189  931.9724 678.3496
#> 26: -1.300150000 -0.493622800 -0.10543945 1.0024508  913.3234 742.6848
#> 27: -1.159574000 -0.190779800  0.25438470 0.9995996 1185.2882 743.3417
#> 28:  1.613406000  2.584372000  3.06662450 1.0007770 1049.4843 541.8187
#> 29:  1.169150000  2.047714000  2.50560650 1.0020309  845.1937 805.5060
#> 30: -0.274356400  0.500179800  0.93658765 1.0078369  987.5066 738.3007
#> 31: -1.092032000 -0.214005200  0.19014685 1.0049760 1089.3756 877.2643
#> 32: -0.642243200  0.151280600  0.54995140 1.0021721 1472.7616 723.2467
#> 33: -1.102268000 -0.215524800  0.16654470 1.0011516 1693.0594 773.6382
#> 34: -0.611274400  0.319924800  0.78841360 0.9990922 1092.7570 792.7206
#> 35:  0.986148000  1.962298000  2.49068150 1.0012054 1038.9344 562.6188
#> 36: -0.298254000  0.640660800  1.07517200 0.9996742 1031.7711 774.8388
#> 37: -0.465815200  0.571496400  1.05543450 1.0041651 1287.1113 792.6725
#> 38: -0.424778200  0.564137800  1.04378350 1.0021169 1340.7507 750.2872
#> 39: -0.364340000  0.727304200  1.25941450 1.0032416 1230.9766 801.3727
#> 40: -0.215913000  1.057834000  1.67070550 1.0026953 1446.8922 646.8817
#>              q20          q80         q95      rhat  ess_bulk ess_tail