Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean      median        sd       mad         q5
#>            <char>       <num>       <num>     <num>     <num>      <num>
#>  1:  expr_beta[1]  0.29620901  0.29376100 0.5662902 0.5448918 -0.5992006
#>  2:  expr_beta[2] -0.31681488 -0.29993400 0.5317599 0.5434355 -1.2043135
#>  3:  expr_beta[3] -0.83187946 -0.82578350 0.5571812 0.5828708 -1.7111430
#>  4:  expr_beta[4]  0.48160358  0.47166350 0.5476592 0.5267300 -0.4414568
#>  5:  expr_beta[5] -0.47554266 -0.50153900 0.5565102 0.5590099 -1.3784790
#>  6:  expr_beta[6] -0.20960816 -0.22442150 0.5950978 0.5852141 -1.1447735
#>  7:  expr_beta[7]  1.72879611  1.71759500 0.5689084 0.5768945  0.7886897
#>  8:  expr_beta[8]  0.31773925  0.30209850 0.5271347 0.5264775 -0.5295946
#>  9:  expr_beta[9]  0.26798140  0.25330950 0.5097008 0.5001025 -0.5475638
#> 10: expr_beta[10] -0.36033269 -0.34907250 0.5301068 0.5182512 -1.2117775
#> 11: expr_beta[11] -0.18039784 -0.16596100 0.5033497 0.5060640 -1.0326625
#> 12: expr_beta[12] -0.54690847 -0.50757950 0.4942241 0.5006562 -1.3490105
#> 13: expr_beta[13] -1.23033541 -1.23327500 0.5502436 0.5771539 -2.1396530
#> 14: expr_beta[14]  1.51989151  1.51439000 0.6027142 0.6144487  0.5561230
#> 15: expr_beta[15]  1.80210036  1.77943000 0.5107207 0.5045658  1.0016505
#> 16: expr_beta[16] -0.22716887 -0.22277000 0.4867748 0.5169567 -1.0460475
#> 17: expr_beta[17] -0.22664140 -0.20906200 0.4709836 0.4724297 -1.0141115
#> 18: expr_beta[18] -0.20209197 -0.21238450 0.4797982 0.4674037 -0.9906518
#> 19: expr_beta[19] -0.55454047 -0.55181050 0.4910778 0.4874003 -1.3448100
#> 20: expr_beta[20] -0.34298956 -0.35151600 0.5067153 0.4810770 -1.2226325
#> 21: expr_beta[21]  1.64406910  1.63164500 0.4877930 0.4724972  0.7984210
#> 22: expr_beta[22]  0.43146950  0.40924200 0.4870717 0.4759646 -0.3637259
#> 23: expr_beta[23] -0.26668978 -0.27113200 0.4650391 0.4451694 -1.0261175
#> 24: expr_beta[24] -0.33973159 -0.31941100 0.4891249 0.4935352 -1.1107665
#> 25: expr_beta[25] -0.11112979 -0.09261535 0.4879186 0.4664482 -0.9222411
#> 26: expr_beta[26] -0.85376828 -0.83187300 0.4893379 0.4694816 -1.7256890
#> 27: expr_beta[27] -0.68370916 -0.67271400 0.5317492 0.5419874 -1.5617985
#> 28: expr_beta[28]  2.11201358  2.12078000 0.5515361 0.5646779  1.2184810
#> 29: expr_beta[29]  1.61215943  1.57959500 0.5049575 0.5042545  0.8324763
#> 30: expr_beta[30]  0.10040269  0.06869015 0.4741600 0.4606337 -0.6331166
#> 31: expr_beta[31] -0.62168928 -0.59641600 0.4816113 0.4637432 -1.4630915
#> 32: expr_beta[32] -0.24466525 -0.24001800 0.4964319 0.4897533 -1.0720000
#> 33: expr_beta[33] -0.66004222 -0.64518200 0.5004014 0.4723260 -1.5233410
#> 34: expr_beta[34] -0.12606312 -0.12501800 0.5350974 0.5406664 -1.0041360
#> 35: expr_beta[35]  1.44070932  1.46098500 0.5647499 0.5677691  0.5176760
#> 36: expr_beta[36]  0.19455571  0.19200300 0.5466678 0.5254431 -0.6910205
#> 37: expr_beta[37]  0.02063698  0.02360580 0.5460403 0.5187069 -0.8766409
#> 38: expr_beta[38]  0.07447285  0.07360340 0.6092489 0.6016248 -0.9110737
#> 39: expr_beta[39]  0.19051808  0.18916950 0.6575018 0.7062617 -0.8801058
#> 40: expr_beta[40]  0.39758301  0.40393550 0.7496790 0.7330337 -0.8752186
#>          variable        mean      median        sd       mad         q5
#>             q20          q80         q95      rhat  ess_bulk ess_tail
#>           <num>        <num>       <num>     <num>     <num>    <num>
#>  1: -0.17205680  0.765937200  1.18709650 1.0005590 1121.3239 652.8846
#>  2: -0.77418400  0.135515800  0.50251835 0.9990386 1981.3448 874.8668
#>  3: -1.30854000 -0.357424600  0.08786453 1.0065508 1748.3595 708.4019
#>  4:  0.04147834  0.944163600  1.41265250 1.0110507 1469.1454 612.3143
#>  5: -0.92154020  0.003286268  0.39687090 0.9995840 1693.0767 713.3629
#>  6: -0.70608800  0.260764400  0.80120415 0.9999039 1417.0141 817.1444
#>  7:  1.26349000  2.197528000  2.66801450 1.0016719 1177.8888 679.4667
#>  8: -0.11971800  0.770508600  1.25031000 1.0009287 1051.3639 691.6514
#>  9: -0.16888420  0.694341600  1.10570900 1.0019377 1303.9249 745.2780
#> 10: -0.78377360  0.083920660  0.47105400 1.0011447 1246.5636 655.8888
#> 11: -0.59299660  0.239857400  0.63863300 1.0126797 1485.3065 816.5606
#> 12: -0.96733380 -0.145935800  0.23739190 1.0058140 1557.4068 739.9153
#> 13: -1.69771600 -0.752380400 -0.35273215 1.0063687 1128.2273 753.4192
#> 14:  0.98651520  2.015004000  2.52092400 0.9998703 1465.4484 811.3731
#> 15:  1.37430400  2.231298000  2.67956600 1.0024314  846.5143 749.0349
#> 16: -0.67066700  0.196942800  0.57026905 1.0012467 1305.1597 789.6947
#> 17: -0.61300220  0.174541200  0.50812125 0.9989294 1270.7035 757.9277
#> 18: -0.58567200  0.187735200  0.60979995 1.0002819 1360.4162 703.5780
#> 19: -0.95456420 -0.155478000  0.23660675 0.9994164 1614.4079 812.4019
#> 20: -0.75989140  0.056627000  0.50518180 1.0018300 1693.8735 686.2069
#> 21:  1.24392200  2.046072000  2.44740750 1.0021632 1141.4184 359.6896
#> 22:  0.02367334  0.829093600  1.29439600 1.0032822 1698.6128 686.8256
#> 23: -0.63654080  0.108879400  0.48338575 1.0013704 1409.2022 944.6680
#> 24: -0.76404400  0.065487160  0.45456765 1.0017239 1224.7525 784.4916
#> 25: -0.51794980  0.301539800  0.66225930 1.0008011 1150.0287 578.6923
#> 26: -1.25178800 -0.441176400 -0.11352870 1.0009225 1322.7116 798.3919
#> 27: -1.14672800 -0.245874000  0.17046835 0.9992539 1385.6803 857.3920
#> 28:  1.62795400  2.581152000  3.00660450 1.0009978 1046.3796 615.0825
#> 29:  1.17702800  2.038976000  2.45862250 0.9997661 1096.1812 722.2563
#> 30: -0.28757180  0.498077600  0.89992380 0.9987015 1433.8879 684.7777
#> 31: -1.01825000 -0.232797200  0.14851350 0.9989266 1237.6908 680.6098
#> 32: -0.65153220  0.151675400  0.59678845 0.9995899 1204.7503 874.9570
#> 33: -1.06222400 -0.265084600  0.18037990 1.0006931 1150.1858 759.3438
#> 34: -0.55843940  0.308902000  0.77440290 1.0021679 1261.7007 697.5065
#> 35:  1.00067340  1.902414000  2.37522600 1.0017658 1527.5194 667.2508
#> 36: -0.25002400  0.617634600  1.10956550 0.9999257 2259.4488 695.6379
#> 37: -0.42020480  0.454609800  0.97964260 0.9995508 1468.3498 749.9510
#> 38: -0.42087460  0.576908200  1.05356950 1.0008158 1207.3397 611.5088
#> 39: -0.39625640  0.763409200  1.20785550 1.0015104 1769.1051 922.9919
#> 40: -0.24897040  1.049056000  1.60072250 0.9994174 1841.5218 815.9484
#>             q20          q80         q95      rhat  ess_bulk ess_tail