Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean      median        sd       mad         q5
#>            <char>       <num>       <num>     <num>     <num>      <num>
#>  1:  expr_beta[1] -0.36231237 -0.35151193 0.5850323 0.6006064 -1.3159711
#>  2:  expr_beta[2] -0.81956125 -0.81313828 0.5819066 0.5560748 -1.7628072
#>  3:  expr_beta[3]  0.46612639  0.49545094 0.5655799 0.5622846 -0.4949055
#>  4:  expr_beta[4] -0.48794876 -0.49916739 0.5589527 0.5466011 -1.3540056
#>  5:  expr_beta[5] -0.22884095 -0.24754695 0.5903614 0.5921100 -1.1685938
#>  6:  expr_beta[6]  1.73624712  1.70249710 0.5808644 0.5712751  0.7963793
#>  7:  expr_beta[7]  0.31247778  0.30180589 0.5438900 0.5319103 -0.5481479
#>  8:  expr_beta[8]  0.27411988  0.25767186 0.5339315 0.5099338 -0.6125495
#>  9:  expr_beta[9] -0.37902061 -0.38080532 0.5177745 0.5149070 -1.2208034
#> 10: expr_beta[10] -0.13800267 -0.13863863 0.5126915 0.4773898 -0.9891446
#> 11: expr_beta[11] -0.56535928 -0.54510019 0.5349959 0.4969682 -1.4540302
#> 12: expr_beta[12] -1.23616337 -1.23939545 0.5669137 0.5490113 -2.2006179
#> 13: expr_beta[13]  1.52430200  1.50412600 0.6083542 0.5898186  0.5706043
#> 14: expr_beta[14]  1.80409074  1.75511405 0.5630083 0.5551297  0.9537094
#> 15: expr_beta[15] -0.21820779 -0.20997694 0.4942602 0.4706036 -0.9992519
#> 16: expr_beta[16] -0.24226832 -0.24057451 0.4896806 0.4745631 -1.0390083
#> 17: expr_beta[17] -0.18186565 -0.18607635 0.4885592 0.4840573 -0.9573425
#> 18: expr_beta[18] -0.56179968 -0.55139067 0.5233220 0.5032514 -1.4664948
#> 19: expr_beta[19] -0.32415045 -0.34557810 0.5414081 0.4867231 -1.2015871
#> 20: expr_beta[20]  1.62616667  1.60915960 0.5308494 0.5013420  0.7826004
#> 21: expr_beta[21]  0.44747037  0.44964237 0.4927074 0.4847680 -0.3446012
#> 22: expr_beta[22] -0.28153522 -0.29583439 0.4731531 0.4539509 -1.0268855
#> 23: expr_beta[23] -0.36288210 -0.35493043 0.5065852 0.4978106 -1.2188500
#> 24: expr_beta[24] -0.11913917 -0.11253562 0.5030405 0.4948383 -0.9200924
#> 25: expr_beta[25] -0.85023634 -0.84443731 0.4808408 0.4926400 -1.6473357
#> 26: expr_beta[26] -0.65606987 -0.64917948 0.5315031 0.4869885 -1.5168655
#> 27: expr_beta[27]  2.09818328  2.07172435 0.6009010 0.5828347  1.1135353
#> 28: expr_beta[28]  1.61234161  1.61686790 0.5287609 0.5388000  0.7493838
#> 29: expr_beta[29]  0.09712640  0.10143074 0.4698238 0.4568917 -0.6671413
#> 30: expr_beta[30] -0.62193854 -0.61202998 0.4911750 0.4764071 -1.4610819
#> 31: expr_beta[31] -0.25666914 -0.24656807 0.5032483 0.4953429 -1.0916028
#> 32: expr_beta[32] -0.63714072 -0.62861597 0.5150330 0.4915710 -1.4770739
#> 33: expr_beta[33] -0.14580387 -0.16495427 0.5645721 0.5251862 -1.0306670
#> 34: expr_beta[34]  1.47075466  1.44993585 0.5893215 0.5336735  0.5148384
#> 35: expr_beta[35]  0.17575746  0.16139901 0.5545039 0.5552362 -0.7371526
#> 36: expr_beta[36]  0.04687313  0.03002159 0.5723066 0.5528774 -0.8561952
#> 37: expr_beta[37]  0.08724014  0.05556753 0.5980743 0.5931941 -0.8643870
#> 38: expr_beta[38]  0.17819387  0.16938760 0.6788880 0.6819758 -0.9490598
#> 39: expr_beta[39]  0.42262333  0.42980613 0.7639196 0.7547569 -0.8700442
#>          variable        mean      median        sd       mad         q5
#>             q20         q80         q95      rhat  ess_bulk ess_tail
#>           <num>       <num>       <num>     <num>     <num>    <num>
#>  1: -0.87133971  0.12555731  0.56508046 0.9989587 1034.8155 784.3484
#>  2: -1.27793500 -0.35128993  0.10315747 1.0012032 1219.4900 659.5294
#>  3: -0.01216879  0.93607003  1.35838524 1.0007621 1436.5899 628.6754
#>  4: -0.95466409 -0.03895863  0.49544196 1.0008615 1124.3042 688.5391
#>  5: -0.72499127  0.26687294  0.74611517 1.0020059 1257.3737 916.6946
#>  6:  1.25857018  2.23777830  2.70428653 1.0015001 1175.0875 937.0116
#>  7: -0.11994888  0.76649043  1.20764185 0.9999581 1422.5973 797.0406
#>  8: -0.16944823  0.72601092  1.14223692 1.0006645 1271.9763 827.1159
#>  9: -0.80719373  0.04970144  0.47068590 0.9997404 1253.8721 796.6055
#> 10: -0.55146267  0.26209196  0.70929364 0.9997342 1247.4815 812.6574
#> 11: -0.97577538 -0.13774556  0.24113021 0.9996867 1167.9192 731.0374
#> 12: -1.69440914 -0.76736135 -0.34887406 1.0017817 1058.8636 813.4761
#> 13:  1.02702416  2.03352926  2.52405496 1.0059446 1064.2396 770.2097
#> 14:  1.33421322  2.28054282  2.76657796 1.0085878  938.2817 644.0399
#> 15: -0.62194760  0.16917189  0.59835233 1.0066819 1141.9916 758.6467
#> 16: -0.63600774  0.16621049  0.54314613 1.0019863 1120.5716 562.4977
#> 17: -0.61136918  0.20751536  0.63155207 1.0029007  770.5252 563.1756
#> 18: -0.98965963 -0.11010908  0.26152720 1.0013634 1130.7799 692.9768
#> 19: -0.74655571  0.09641736  0.55604595 0.9999846 1088.1731 689.8266
#> 20:  1.18499112  2.04806982  2.52287568 1.0048840  666.1959 718.7960
#> 21:  0.01769111  0.84117455  1.29764955 0.9989889 1255.2466 720.4613
#> 22: -0.67767772  0.10544919  0.51079736 1.0008952 1042.3838 667.5575
#> 23: -0.76521636  0.04122409  0.47322174 1.0011551  927.7012 677.4681
#> 24: -0.55313653  0.27715696  0.71077097 1.0055471 1234.8771 642.6127
#> 25: -1.23696004 -0.43866941 -0.09357462 1.0005179 1003.9991 732.4665
#> 26: -1.08055646 -0.24322699  0.19951345 1.0009004 1022.0853 704.1599
#> 27:  1.61533832  2.59578078  3.12353828 1.0030772  699.4889 481.3586
#> 28:  1.16813958  2.07291802  2.44034505 1.0028804  997.7398 787.7762
#> 29: -0.29967880  0.48647851  0.87632563 1.0060800 1086.4967 804.1704
#> 30: -1.02467426 -0.21999886  0.16220810 1.0057719 1074.2416 908.8440
#> 31: -0.68746043  0.16230600  0.55967256 1.0021082 1338.2938 902.4538
#> 32: -1.05282160 -0.22627583  0.17844925 1.0011134 1012.3729 646.3955
#> 33: -0.59472951  0.30289449  0.74514686 1.0036005 1104.3458 855.6841
#> 34:  1.00136246  1.92662572  2.44258336 1.0009131 1219.4466 773.5495
#> 35: -0.29302284  0.65335792  1.09055307 1.0009167 1069.8313 540.1782
#> 36: -0.42408784  0.51637944  1.02373972 1.0088049 1271.9392 723.2950
#> 37: -0.41256106  0.58398682  1.08271750 1.0031225 1570.9752 868.4931
#> 38: -0.41310494  0.76460447  1.32156526 0.9987118 1650.8947 567.5275
#> 39: -0.19550220  1.07287042  1.63541228 1.0052628 1443.3048 638.0017
#>             q20         q80         q95      rhat  ess_bulk ess_tail