Summarise the posterior

A generic wrapper around posterior::summarise_draws() with opinionated defaults.

Usage

enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)

Arguments

fit

A cmdstanr fit object.

variables

A character vector of variables to return posterior summaries for. By default summaries for all parameters are returned.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

...

Additional arguments that may be passed but will not be used.

Value

A data.frame summarising the model posterior.

See also

Examples

fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
#>          variable        mean       median        sd       mad         q5
#>  1:  expr_beta[1]  0.32247472  0.338387000 0.5599104 0.5401784 -0.5884213
#>  2:  expr_beta[2] -0.33517948 -0.329458500 0.5515760 0.5133994 -1.2521760
#>  3:  expr_beta[3] -0.82458037 -0.830468000 0.5664859 0.5410890 -1.7149765
#>  4:  expr_beta[4]  0.50017845  0.507985500 0.5548193 0.5656964 -0.4416455
#>  5:  expr_beta[5] -0.49211128 -0.501705500 0.5506490 0.5623101 -1.3512810
#>  6:  expr_beta[6] -0.21819429 -0.206166000 0.5510473 0.5370378 -1.1559205
#>  7:  expr_beta[7]  1.71806636  1.729265000 0.5462492 0.5460119  0.8004145
#>  8:  expr_beta[8]  0.32300834  0.307873500 0.5308181 0.5497788 -0.5055089
#>  9:  expr_beta[9]  0.26073828  0.282569500 0.5012190 0.4990157 -0.6089277
#> 10: expr_beta[10] -0.37001865 -0.382536500 0.5316492 0.5411119 -1.2536980
#> 11: expr_beta[11] -0.16254693 -0.149461500 0.5293063 0.5118632 -1.0445990
#> 12: expr_beta[12] -0.57731811 -0.573319500 0.5481950 0.5329161 -1.4757775
#> 13: expr_beta[13] -1.19988748 -1.213040000 0.5883767 0.6189744 -2.1268635
#> 14: expr_beta[14]  1.53836898  1.563060000 0.5998243 0.5683769  0.4959207
#> 15: expr_beta[15]  1.77050210  1.748155000 0.5322216 0.5316604  0.9741486
#> 16: expr_beta[16] -0.23216867 -0.258784500 0.4707397 0.4703319 -0.9805193
#> 17: expr_beta[17] -0.21833087 -0.242596000 0.4769162 0.4754961 -0.9622282
#> 18: expr_beta[18] -0.19196526 -0.195510000 0.4794046 0.5002166 -0.9928552
#> 19: expr_beta[19] -0.57075420 -0.556655500 0.5066780 0.4952225 -1.3583375
#> 20: expr_beta[20] -0.34326445 -0.337585500 0.5160784 0.5057089 -1.2166530
#> 21: expr_beta[21]  1.65891978  1.651225000 0.5281567 0.5072419  0.7690231
#> 22: expr_beta[22]  0.42427424  0.412209500 0.4715924 0.4421669 -0.3293036
#> 23: expr_beta[23] -0.28568970 -0.295981000 0.4512625 0.4222486 -1.0215700
#> 24: expr_beta[24] -0.32155423 -0.324554500 0.4691444 0.4423797 -1.1283835
#> 25: expr_beta[25] -0.09773168 -0.120236000 0.4602765 0.4401335 -0.8718705
#> 26: expr_beta[26] -0.88791656 -0.865780000 0.4925783 0.4473286 -1.7588625
#> 27: expr_beta[27] -0.66798334 -0.659061000 0.5198224 0.5064465 -1.5324715
#> 28: expr_beta[28]  2.08750148  2.055500000 0.5381331 0.5322608  1.2505040
#> 29: expr_beta[29]  1.61456236  1.605955000 0.5373865 0.5060633  0.7811348
#> 30: expr_beta[30]  0.11775303  0.100548500 0.4773478 0.4689397 -0.6268959
#> 31: expr_beta[31] -0.65807533 -0.642011000 0.4920458 0.4700153 -1.4376195
#> 32: expr_beta[32] -0.24204215 -0.240013000 0.4854801 0.4898703 -1.0065790
#> 33: expr_beta[33] -0.63662301 -0.630302500 0.5123658 0.4981299 -1.4756840
#> 34: expr_beta[34] -0.13428000 -0.133333500 0.5681001 0.5320414 -1.0641440
#> 35: expr_beta[35]  1.45873527  1.446325000 0.5648977 0.5513641  0.5702454
#> 36: expr_beta[36]  0.17596728  0.167394000 0.5543489 0.5432009 -0.6964099
#> 37: expr_beta[37]  0.01502973  0.008104375 0.5782420 0.5842489 -0.8798406
#> 38: expr_beta[38]  0.10388639  0.071395650 0.6512795 0.6383030 -0.9309701
#> 39: expr_beta[39]  0.18580796  0.168813500 0.6895821 0.6621166 -0.9171666
#> 40: expr_beta[40]  0.39259101  0.377317500 0.7767691 0.7296290 -0.8495955
#>          variable        mean       median        sd       mad         q5
#>             q20         q80         q95      rhat  ess_bulk ess_tail
#>  1: -0.14270780  0.78065200  1.20346400 1.0021603  714.6681 609.9378
#>  2: -0.77719040  0.10379780  0.54302050 1.0036855 1262.6375 504.0825
#>  3: -1.29391200 -0.36525580  0.07783156 0.9999359 1116.2168 697.3545
#>  4:  0.03849216  0.97584900  1.40315350 1.0086691 1135.5286 615.6157
#>  5: -0.97335260 -0.03204564  0.41000290 1.0010007 1258.1778 861.2881
#>  6: -0.67434520  0.22287280  0.71500850 0.9999690 1048.9039 516.2610
#>  7:  1.25414000  2.16936400  2.62522350 0.9983015 1021.2042 699.9287
#>  8: -0.14248400  0.77871360  1.18683750 1.0003833  848.9079 770.8900
#>  9: -0.18209020  0.68633040  1.06368250 1.0044866 1007.6991 870.8076
#> 10: -0.80714900  0.08960668  0.46413145 1.0015626 1292.0185 740.9535
#> 11: -0.59539100  0.25483080  0.72313170 0.9995037 1004.0812 669.2008
#> 12: -1.01078400 -0.13058960  0.31369895 1.0026888 1111.7633 603.5306
#> 13: -1.73638600 -0.70285240 -0.22549755 1.0017319  918.0660 668.8652
#> 14:  1.04193200  2.00727400  2.52476650 1.0039410  811.5404 715.9051
#> 15:  1.31238000  2.21457000  2.69267250 1.0081784  671.7252 606.7500
#> 16: -0.63643420  0.14273940  0.58648660 0.9993838  902.3891 664.8678
#> 17: -0.61403460  0.19564420  0.59908500 1.0001883  942.6207 649.6672
#> 18: -0.59803760  0.22412580  0.57244125 1.0084288  955.0925 559.5933
#> 19: -0.98200240 -0.14096460  0.22648935 1.0031153  831.1255 650.9774
#> 20: -0.75997380  0.10367300  0.45689665 1.0016436 1029.9612 488.1878
#> 21:  1.22571600  2.09416000  2.51556550 0.9999525  913.1640 697.0549
#> 22:  0.03576674  0.82900780  1.19219400 1.0082323  845.4144 740.5203
#> 23: -0.63054160  0.06901830  0.43821110 1.0018118  964.1409 840.2770
#> 24: -0.69097480  0.05950174  0.43832860 1.0013270 1062.1889 837.0867
#> 25: -0.44475380  0.28423200  0.62355940 0.9997798  884.0591 541.5468
#> 26: -1.25291000 -0.49283920 -0.11463860 0.9997994  870.8012 604.8573
#> 27: -1.09106400 -0.22992720  0.17963435 1.0035222 1005.9316 869.3363
#> 28:  1.63731600  2.53761000  2.99754950 1.0008203  771.7465 705.7897
#> 29:  1.17005600  2.02111600  2.56786150 1.0045334  630.0200 529.4048
#> 30: -0.27930620  0.50096000  0.94497975 0.9987949 1054.5768 768.7437
#> 31: -1.07643800 -0.26095880  0.15197150 1.0001976  895.8805 739.0679
#> 32: -0.64508600  0.16545180  0.56508735 1.0022309  846.4045 609.9356
#> 33: -1.06041600 -0.19554720  0.20240210 1.0002922 1078.1239 756.8602
#> 34: -0.59636400  0.31713240  0.78808490 1.0014451 1021.8652 754.7591
#> 35:  0.99095740  1.93170800  2.39164050 1.0010707 1143.6083 677.2956
#> 36: -0.29490800  0.62365040  1.06787450 1.0016283 1050.9204 712.3969
#> 37: -0.48485900  0.49282200  0.98846190 1.0064016 1140.9861 709.5052
#> 38: -0.43604580  0.62166340  1.22410250 1.0010480  948.8004 695.3760
#> 39: -0.37672560  0.76397180  1.35751650 1.0008524 1133.5732 822.8682
#> 40: -0.22108880  1.04695400  1.75608150 0.9997342 1052.5630 747.7963
#>             q20         q80         q95      rhat  ess_bulk ess_tail