Posterior predictive summary

This function summarises posterior predictives for observed data (by report and reference date). The functionality of this function can be used directly on the output of epinowcast() using the supplied summary.epinowcast() method.

Usage

enw_pp_summary(fit, diff_obs, probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95))

Arguments

fit

A cmdstanr fit object.

diff_obs

A data.frame of observed data with at least a date variable reference_date, and a grouping variable .group.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

Value

A data.table summarising the posterior predictions.

See also

Examples

fit <- enw_example("nowcast")
enw_pp_summary(fit$fit[[1]], fit$new_confirm[[1]], probs = c(0.5))
#>      reference_date report_date .group max_confirm location age_group confirm
#>   1:     2021-07-13  2021-07-13      1          59       DE       00+      21
#>   2:     2021-07-13  2021-07-14      1          59       DE       00+      33
#>   3:     2021-07-13  2021-07-15      1          59       DE       00+      36
#>   4:     2021-07-13  2021-07-16      1          59       DE       00+      40
#>   5:     2021-07-13  2021-07-17      1          59       DE       00+      43
#>  ---                                                                         
#> 626:     2021-08-20  2021-08-21      1         171       DE       00+     159
#> 627:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 628:     2021-08-21  2021-08-21      1         112       DE       00+      69
#> 629:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 630:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>      cum_prop_reported delay new_confirm prop_reported   mean median        sd
#>   1:         0.3559322     0          21    0.35593220 19.340     18  8.788381
#>   2:         0.5593220     1          12    0.20338983 20.633     19  9.243579
#>   3:         0.6101695     2           3    0.05084746  6.047      5  3.458432
#>   4:         0.6779661     3           4    0.06779661  4.113      4  2.548475
#>   5:         0.7288136     4           3    0.05084746  2.400      2  1.809219
#>  ---                                                                          
#> 626:         0.9298246     1          61    0.35672515 80.093     75 32.268927
#> 627:         1.0000000     2          12    0.07017544 12.498     12  6.274501
#> 628:         0.6160714     0          69    0.61607143 74.934     69 32.428217
#> 629:         1.0000000     1          43    0.38392857 46.497     43 19.085761
#> 630:         1.0000000     0          45    1.00000000 40.198     37 19.651946
#>          mad q50      rhat  ess_bulk  ess_tail
#>   1:  8.8956  18 0.9991744  999.8591 1027.3024
#>   2:  8.8956  19 1.0023251  948.1680  978.4886
#>   3:  2.9652   5 0.9993604  734.7846  751.2877
#>   4:  2.9652   4 1.0030090 1100.5175 1009.5913
#>   5:  1.4826   2 0.9992011  965.0750  994.7849
#>  ---                                          
#> 626: 31.1346  75 1.0009581 1047.6045 1020.4985
#> 627:  5.9304  12 1.0010833  881.1689  967.4896
#> 628: 29.6520  69 1.0009965 1131.3266  881.7254
#> 629: 17.7912  43 0.9991313 1028.5843  916.3896
#> 630: 17.7912  37 1.0007027  975.8059  874.9003