Summarise posterior samples

This function summarises posterior samples for arbitrary strata. It optionally holds out the observed data (variables that are not ".draw", ".iteration", ".sample", ".chain" ) joins this to the summarised posterior.

enw_summarise_samples(
  samples,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  by = c("reference_date", ".group"),
  link_with_obs = TRUE
)

Arguments

samples

A data.frame of posterior samples with at least a numeric sample variable.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

by

A character vector of variables to summarise by. Defaults to c("reference_date", ".group").

link_with_obs

Logical, should the observed data be linked to the posterior summary? This is useful for plotting the posterior against the observed data. Defaults to TRUE.

Value

A data.frame summarising the posterior samples.

See also

Examples

fit <- enw_example("nowcast")
samples <- summary(fit, type = "nowcast_sample")
enw_summarise_samples(samples, probs = c(0.05, 0.5, 0.95))
#> Key: <reference_date, .group>
#>     reference_date .group report_date max_confirm location age_group confirm
#>             <IDat>  <num>      <IDat>       <int>   <fctr>    <fctr>   <int>
#>  1:     2021-08-03      1  2021-08-22         149       DE       00+     149
#>  2:     2021-08-04      1  2021-08-22         166       DE       00+     166
#>  3:     2021-08-05      1  2021-08-22         133       DE       00+     133
#>  4:     2021-08-06      1  2021-08-22         137       DE       00+     137
#>  5:     2021-08-07      1  2021-08-22         139       DE       00+     139
#>  6:     2021-08-08      1  2021-08-22          97       DE       00+      97
#>  7:     2021-08-09      1  2021-08-22          58       DE       00+      58
#>  8:     2021-08-10      1  2021-08-22         175       DE       00+     175
#>  9:     2021-08-11      1  2021-08-22         233       DE       00+     233
#> 10:     2021-08-12      1  2021-08-22         237       DE       00+     237
#> 11:     2021-08-13      1  2021-08-22         204       DE       00+     204
#> 12:     2021-08-14      1  2021-08-22         189       DE       00+     189
#> 13:     2021-08-15      1  2021-08-22         125       DE       00+     125
#> 14:     2021-08-16      1  2021-08-22          98       DE       00+      98
#> 15:     2021-08-17      1  2021-08-22         242       DE       00+     242
#> 16:     2021-08-18      1  2021-08-22         223       DE       00+     223
#> 17:     2021-08-19      1  2021-08-22         202       DE       00+     202
#> 18:     2021-08-20      1  2021-08-22         171       DE       00+     171
#> 19:     2021-08-21      1  2021-08-22         112       DE       00+     112
#> 20:     2021-08-22      1  2021-08-22          45       DE       00+      45
#>     reference_date .group report_date max_confirm location age_group confirm
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>                 <num> <num>         <num>   <num>  <num>      <num>   <num>
#>  1:                 1    19   0.000000000 149.000    149   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.426    167   1.318432  1.4826
#>  3:                 1    17   0.000000000 135.789    136   1.854162  1.4826
#>  4:                 1    16   0.000000000 141.288    141   2.378805  1.4826
#>  5:                 1    15   0.007194245 146.019    146   3.079630  2.9652
#>  6:                 1    14   0.000000000 103.853    104   2.969884  2.9652
#>  7:                 1    13   0.000000000  62.766     63   2.432110  2.9652
#>  8:                 1    12   0.000000000 185.035    185   3.847932  4.4478
#>  9:                 1    11   0.000000000 256.031    256   6.498790  5.9304
#> 10:                 1    10   0.004219409 267.692    267   7.713536  7.4130
#> 11:                 1     9   0.000000000 236.057    235   8.508004  7.4130
#> 12:                 1     8   0.015873016 231.407    231  10.083304 10.3782
#> 13:                 1     7   0.040000000 164.876    164   9.848127 10.3782
#> 14:                 1     6   0.010204082 128.858    128   8.134495  7.4130
#> 15:                 1     5   0.012396694 293.309    292  11.938427 11.8608
#> 16:                 1     4   0.017937220 293.042    292  15.799489 14.8260
#> 17:                 1     3   0.019801980 291.478    290  21.108128 19.2738
#> 18:                 1     2   0.070175439 293.178    291  29.440398 28.1694
#> 19:                 1     1   0.383928571 308.518    302  51.151522 46.7019
#> 20:                 1     0   1.000000000 375.644    363 103.424281 91.1799
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>         q5   q50    q95
#>      <num> <num>  <num>
#>  1: 149.00   149 149.00
#>  2: 166.00   167 170.00
#>  3: 133.00   136 139.00
#>  4: 138.00   141 146.00
#>  5: 141.00   146 152.00
#>  6: 100.00   104 109.00
#>  7:  59.00    63  67.00
#>  8: 179.00   185 192.00
#>  9: 246.00   256 268.00
#> 10: 256.00   267 282.00
#> 11: 223.00   235 251.00
#> 12: 216.00   231 250.00
#> 13: 150.00   164 181.00
#> 14: 117.00   128 143.00
#> 15: 276.00   292 315.00
#> 16: 268.00   292 320.00
#> 17: 260.00   290 330.00
#> 18: 250.00   291 344.00
#> 19: 239.00   302 399.10
#> 20: 233.95   363 565.05
#>         q5   q50    q95