Summarise posterior samples

This function summarises posterior samples for arbitrary strata. It optionally holds out the observed data (variables that are not ".draw", ".iteration", ".sample", ".chain" ) joins this to the summarised posterior.

Usage

enw_summarise_samples(
  samples,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  by = c("reference_date", ".group"),
  link_with_obs = TRUE
)

Arguments

samples

A data.frame of posterior samples with at least a numeric sample variable.

probs

A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work.

by

A character vector of variables to summarise by. Defaults to c("reference_date", ".group").

link_with_obs

Logical, should the observed data be linked to the posterior summary? This is useful for plotting the posterior against the observed data. Defaults to TRUE.

Value

A data.frame summarising the posterior samples.

See also

Examples

fit <- enw_example("nowcast")
samples <- summary(fit, type = "nowcast_sample")
enw_summarise_samples(samples, probs = c(0.05, 0.5, 0.95))
#>     reference_date .group report_date max_confirm location age_group confirm
#>  1:     2021-08-03      1  2021-08-22         149       DE       00+     149
#>  2:     2021-08-04      1  2021-08-22         166       DE       00+     166
#>  3:     2021-08-05      1  2021-08-22         133       DE       00+     133
#>  4:     2021-08-06      1  2021-08-22         137       DE       00+     137
#>  5:     2021-08-07      1  2021-08-22         139       DE       00+     139
#>  6:     2021-08-08      1  2021-08-22          97       DE       00+      97
#>  7:     2021-08-09      1  2021-08-22          58       DE       00+      58
#>  8:     2021-08-10      1  2021-08-22         175       DE       00+     175
#>  9:     2021-08-11      1  2021-08-22         233       DE       00+     233
#> 10:     2021-08-12      1  2021-08-22         237       DE       00+     237
#> 11:     2021-08-13      1  2021-08-22         204       DE       00+     204
#> 12:     2021-08-14      1  2021-08-22         189       DE       00+     189
#> 13:     2021-08-15      1  2021-08-22         125       DE       00+     125
#> 14:     2021-08-16      1  2021-08-22          98       DE       00+      98
#> 15:     2021-08-17      1  2021-08-22         242       DE       00+     242
#> 16:     2021-08-18      1  2021-08-22         223       DE       00+     223
#> 17:     2021-08-19      1  2021-08-22         202       DE       00+     202
#> 18:     2021-08-20      1  2021-08-22         171       DE       00+     171
#> 19:     2021-08-21      1  2021-08-22         112       DE       00+     112
#> 20:     2021-08-22      1  2021-08-22          45       DE       00+      45
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>  1:                 1    19   0.000000000 149.000    149   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.374    167   1.289103  1.4826
#>  3:                 1    17   0.000000000 135.815    136   1.837975  1.4826
#>  4:                 1    16   0.000000000 141.216    141   2.345815  2.9652
#>  5:                 1    15   0.007194245 145.893    145   3.111468  2.9652
#>  6:                 1    14   0.000000000 103.790    103   3.037619  2.9652
#>  7:                 1    13   0.000000000  62.841     62   2.658719  2.9652
#>  8:                 1    12   0.000000000 184.990    185   3.790814  2.9652
#>  9:                 1    11   0.000000000 256.377    256   6.672285  5.9304
#> 10:                 1    10   0.004219409 267.731    267   8.319727  8.8956
#> 11:                 1     9   0.000000000 236.758    236   8.200285  7.4130
#> 12:                 1     8   0.015873016 231.538    231  10.108053 10.3782
#> 13:                 1     7   0.040000000 165.269    164  10.333800 10.3782
#> 14:                 1     6   0.010204082 129.858    129   8.445660  7.4130
#> 15:                 1     5   0.012396694 293.125    292  12.075811 11.8608
#> 16:                 1     4   0.017937220 293.747    292  16.541784 14.8260
#> 17:                 1     3   0.019801980 291.156    289  21.078804 19.2738
#> 18:                 1     2   0.070175439 294.404    291  29.526710 28.1694
#> 19:                 1     1   0.383928571 311.311    307  50.224743 47.4432
#> 20:                 1     0   1.000000000 372.540    360 105.260887 91.9212
#>         q5 q50    q95
#>  1: 149.00 149 149.00
#>  2: 166.00 167 170.00
#>  3: 133.00 136 139.00
#>  4: 138.00 141 145.00
#>  5: 141.00 145 151.05
#>  6:  99.00 103 109.00
#>  7:  59.00  62  68.00
#>  8: 179.00 185 192.00
#>  9: 247.00 256 268.00
#> 10: 255.00 267 282.00
#> 11: 224.00 236 251.00
#> 12: 216.00 231 250.00
#> 13: 149.95 164 183.00
#> 14: 117.00 129 145.00
#> 15: 275.00 292 315.00
#> 16: 269.00 292 325.05
#> 17: 260.00 289 327.00
#> 18: 251.00 291 350.05
#> 19: 238.00 307 401.05
#> 20: 232.95 360 565.05