epinowcast: A Bayesian Framework for Real-time Infectious Disease Surveillance

A modular Bayesian framework for real-time infectious disease surveillance. Provides tools for nowcasting, reproduction number estimation, delay estimation, and forecasting from data subject to reporting delays, right-truncation, missing data, and incomplete ascertainment. Users can build models suited to their setting using a flexible formula interface supporting fixed effects, random effects, random walks, and time-varying parameters, with options including parametric and non-parametric delay distributions with optional modifiers (via discrete-time hazard models), renewal processes, observation models, missing data imputation, and stratified analyses with partial pooling. By jointly estimating disease dynamics and reporting patterns, our framework enables earlier and more reliable detection of trends. While designed with epidemiological applications in mind, the framework can be applied to any right-truncated time series count data.

See also

Author

Maintainer: Sam Abbott contact@samabbott.co.uk (ORCID)

Authors:

Adrian Lison adrian.lison@bsse.ethz.ch (ORCID)

Sebastian Funk sebastian.funk@lshtm.ac.uk

Carl Pearson carl.ab.pearson@gmail.com (ORCID)

Hugo Gruson hugo.gruson@normalesup.org (ORCID)

Felix Guenther felixguenther1@gmx.de (ORCID)

Michael DeWitt me.dewitt.jr@gmail.com (ORCID)

James Mba Azam james.azam@lshtm.ac.uk (ORCID)

Jessalyn Sebastian jessalynsebastian@gmail.com (ORCID)

Other contributors:

Hannah Choi hannah.choi1@lshtm.ac.uk [contributor]

Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [contributor]

Joel Hellewell joel@ebi.ac.uk (ORCID) [contributor]

Luis Rivas luisnicolasrivas@gmail.com [contributor]

Sang Woo Park swp2@princeton.edu (ORCID) [contributor]

Nathan McIntosh natemcintosh@gmail.com [contributor]

Kath Sherratt katharine.sherratt@lshtm.ac.uk (ORCID) [contributor]

Nikos Bosse nikos.bosse@lshtm.ac.uk (ORCID) [contributor]

Adam Howes adamthowes@gmail.com (ORCID) [contributor]

Kaitlyn Johnson johnsonkaitlyne9@gmail.com (ORCID) [contributor]

Barbora Nemcova barbora.nemcova@kit.edu (ORCID) [contributor]