Simulate daily double censored PMF

This function simulates the probability mass function of a daily double-censored process. The process involves two distributions: a primary distribution which represents the censoring process for the primary event and another distribution (which is offset by the primary).

Usage

simulate_double_censored_pmf(
  max,
  fun_primary = stats::runif,
  primary_args = list(),
  fun_dist = stats::rlnorm,
  dist_args = list(...),
  n = 1e+06,
  ...
)

Arguments

max

Maximum value for the computed CDF. If not specified, the maximum value is the maximum simulated delay.

fun_primary

Primary distribution function (default is runif).

primary_args

List of additional arguments to be passed to the primary distribution function.

fun_dist

Distribution function to be added to the primary (default is rlnorm).

dist_args

List of additional arguments to be passed to the distribution function.

n

Number of simulations (default is 1e6).

...

Additional arguments to be passed to the distribution function. This is an alternative to dist_args.

Value

A numeric vector representing the PMF.

See also

Examples

simulate_double_censored_pmf(10, meanlog = 0, sdlog = 1)
#>  [1] 0.238554 0.409493 0.169201 0.076588 0.039356 0.022459 0.013600 0.008697
#>  [9] 0.005946 0.003933 0.002796