Summary method for epinowcast

summary method for class "epinowcast".

Usage

# S3 method for class 'epinowcast'
summary(
  object,
  type = c("nowcast", "nowcast_samples", "fit", "posterior_prediction"),
  max_delay = object$max_delay,
  ...
)

Arguments

object

A data.table output from epinowcast().

type

Character string indicating the summary to return; enforced by base::match.arg(). Supported options are:

"nowcast" which summarises nowcast posterior with enw_nowcast_summary(),

"nowcast_samples" which samples latest with enw_nowcast_samples(),

"fit" which returns the summarised cmdstanr fit with enw_posterior(),

"posterior_prediction" which returns summarised posterior predictions for the observations after fitting using enw_pp_summary().

max_delay

Maximum delay to which nowcasts should be summarised. Must be equal (default) or larger than the modelled maximum delay. If it is larger, then nowcasts for unmodelled dates are added by assuming that case counts beyond the modelled maximum delay are fully observed.

...

Additional arguments passed to summary specified by type.

Value

A summary data.frame

See also

Examples

nowcast <- enw_example("nowcast")

# Summarise nowcast posterior
summary(nowcast, type = "nowcast")
#>     reference_date report_date .group max_confirm location age_group confirm
#>             <IDat>      <IDat>  <num>       <int>   <fctr>    <fctr>   <int>
#>  1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>  2:     2021-08-04  2021-08-22      1         166       DE       00+     166
#>  3:     2021-08-05  2021-08-22      1         133       DE       00+     133
#>  4:     2021-08-06  2021-08-22      1         137       DE       00+     137
#>  5:     2021-08-07  2021-08-22      1         139       DE       00+     139
#>  6:     2021-08-08  2021-08-22      1          97       DE       00+      97
#>  7:     2021-08-09  2021-08-22      1          58       DE       00+      58
#>  8:     2021-08-10  2021-08-22      1         175       DE       00+     175
#>  9:     2021-08-11  2021-08-22      1         233       DE       00+     233
#> 10:     2021-08-12  2021-08-22      1         237       DE       00+     237
#> 11:     2021-08-13  2021-08-22      1         204       DE       00+     204
#> 12:     2021-08-14  2021-08-22      1         189       DE       00+     189
#> 13:     2021-08-15  2021-08-22      1         125       DE       00+     125
#> 14:     2021-08-16  2021-08-22      1          98       DE       00+      98
#> 15:     2021-08-17  2021-08-22      1         242       DE       00+     242
#> 16:     2021-08-18  2021-08-22      1         223       DE       00+     223
#> 17:     2021-08-19  2021-08-22      1         202       DE       00+     202
#> 18:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 19:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 20:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>     reference_date report_date .group max_confirm location age_group confirm
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>                 <num> <num>         <num>   <num>  <num>      <num>   <num>
#>  1:                 1    19   0.000000000 149.000  149.0   0.000000  0.0000
#>  2:                 1    18   0.000000000 167.424  167.0   1.293018  1.4826
#>  3:                 1    17   0.000000000 135.760  135.5   1.798787  2.2239
#>  4:                 1    16   0.000000000 141.290  141.0   2.340807  2.9652
#>  5:                 1    15   0.007194245 145.842  145.5   3.092345  3.7065
#>  6:                 1    14   0.000000000 103.852  104.0   3.069775  2.9652
#>  7:                 1    13   0.000000000  62.653   62.0   2.382912  2.9652
#>  8:                 1    12   0.000000000 185.247  185.0   3.988722  4.4478
#>  9:                 1    11   0.000000000 256.324  256.0   6.795970  5.9304
#> 10:                 1    10   0.004219409 267.501  267.0   7.764038  7.4130
#> 11:                 1     9   0.000000000 236.484  236.0   8.613358  7.4130
#> 12:                 1     8   0.015873016 231.475  231.0  10.333642  8.8956
#> 13:                 1     7   0.040000000 165.095  164.0   9.970927 10.3782
#> 14:                 1     6   0.010204082 130.024  129.0   8.503278  8.8956
#> 15:                 1     5   0.012396694 292.568  292.0  11.588255 11.1195
#> 16:                 1     4   0.017937220 292.827  292.0  15.568798 14.8260
#> 17:                 1     3   0.019801980 291.069  289.0  20.306417 19.2738
#> 18:                 1     2   0.070175439 292.768  290.0  28.243546 26.6868
#> 19:                 1     1   0.383928571 310.151  303.0  50.513957 45.9606
#> 20:                 1     0   1.000000000 374.970  359.0 105.934202 93.4038
#>     cum_prop_reported delay prop_reported    mean median         sd     mad
#>         q5   q20    q35   q50    q65   q80    q95      rhat  ess_bulk  ess_tail
#>      <num> <num>  <num> <num>  <num> <num>  <num>     <num>     <num>     <num>
#>  1: 149.00   149 149.00 149.0 149.00   149 149.00        NA        NA        NA
#>  2: 166.00   166 167.00 167.0 168.00   168 170.00 0.9987896 1188.2125 1063.3471
#>  3: 133.00   134 135.00 135.5 136.00   137 139.00 1.0008364  963.5041  951.3370
#>  4: 138.00   139 140.00 141.0 142.00   143 145.00 0.9996689 1102.2710  925.0777
#>  5: 142.00   143 144.00 145.5 147.00   148 152.00 1.0038336  987.5966  824.7528
#>  6:  99.00   101 102.00 104.0 105.00   106 109.00 1.0006551 1101.8532  991.0711
#>  7:  59.00    61  62.00  62.0  63.00    64  67.00 1.0010114  751.4622  773.2821
#>  8: 179.00   182 183.00 185.0 186.00   189 192.00 0.9997242 1059.1415  883.2260
#>  9: 246.00   251 253.00 256.0 258.00   262 268.00 0.9997658 1039.1865  998.6170
#> 10: 256.00   261 264.00 267.0 270.00   274 281.00 0.9996562 1039.1951  936.1291
#> 11: 223.00   229 233.00 236.0 239.00   243 252.00 1.0009820 1187.7267 1046.0769
#> 12: 216.00   223 227.00 231.0 234.35   240 250.00 1.0012825  863.1990  964.5881
#> 13: 150.00   157 161.00 164.0 168.00   173 182.05 1.0020750 1113.7200  714.2937
#> 14: 117.00   123 126.00 129.0 133.00   137 144.00 1.0003030 1148.2821  886.6410
#> 15: 275.00   283 288.00 292.0 296.00   302 313.00 1.0002163 1089.7858 1059.6357
#> 16: 269.95   279 287.00 292.0 297.00   305 320.00 0.9993465 1014.5908  974.8269
#> 17: 262.00   274 281.65 289.0 297.00   307 326.00 1.0053979  946.5257  909.0082
#> 18: 253.00   269 280.00 290.0 300.00   314 344.00 1.0017054 1208.7533  912.9634
#> 19: 240.00   267 286.00 303.0 322.00   348 403.00 0.9990505 1404.1200  860.4009
#> 20: 237.00   290 324.00 359.0 400.00   450 555.25 1.0005813 1458.2840  771.2588
#>         q5   q20    q35   q50    q65   q80    q95      rhat  ess_bulk  ess_tail

# Nowcast posterior samples
summary(nowcast, type = "nowcast_samples")
#>        reference_date report_date .group max_confirm location age_group confirm
#>                <IDat>      <IDat>  <num>       <int>   <fctr>    <fctr>   <int>
#>     1:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>     2:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>     3:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>     4:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>     5:     2021-08-03  2021-08-22      1         149       DE       00+     149
#>    ---                                                                         
#> 19996:     2021-08-22  2021-08-22      1          45       DE       00+      45
#> 19997:     2021-08-22  2021-08-22      1          45       DE       00+      45
#> 19998:     2021-08-22  2021-08-22      1          45       DE       00+      45
#> 19999:     2021-08-22  2021-08-22      1          45       DE       00+      45
#> 20000:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>        cum_prop_reported delay prop_reported .chain .iteration .draw sample
#>                    <num> <num>         <num>  <int>      <int> <int>  <num>
#>     1:                 1    19             0      1          1     1    149
#>     2:                 1    19             0      1          2     2    149
#>     3:                 1    19             0      1          3     3    149
#>     4:                 1    19             0      1          4     4    149
#>     5:                 1    19             0      1          5     5    149
#>    ---                                                                     
#> 19996:                 1     0             1      2        496   996    376
#> 19997:                 1     0             1      2        497   997    496
#> 19998:                 1     0             1      2        498   998    268
#> 19999:                 1     0             1      2        499   999    464
#> 20000:                 1     0             1      2        500  1000    409

# Nowcast model fit
summary(nowcast, type = "fit")
#>                    variable          mean        median          sd        mad
#>                      <char>         <num>         <num>       <num>      <num>
#>   1:                   lp__ -1374.5173500 -1374.1300000   7.6002999  7.8281280
#>   2: expr_lelatent_int[1,1]     4.1850615     4.1897800   0.1634419  0.1573261
#>   3:           expr_beta[1]     0.2962090     0.2937610   0.5662902  0.5448918
#>   4:           expr_beta[2]    -0.3168149    -0.2999340   0.5317599  0.5434355
#>   5:           expr_beta[3]    -0.8318795    -0.8257835   0.5571812  0.5828708
#>  ---                                                                          
#> 852:       pp_inf_obs[16,1]   292.8270000   292.0000000  15.5687976 14.8260000
#> 853:       pp_inf_obs[17,1]   291.0690000   289.0000000  20.3064174 19.2738000
#> 854:       pp_inf_obs[18,1]   292.7680000   290.0000000  28.2435457 26.6868000
#> 855:       pp_inf_obs[19,1]   310.1510000   303.0000000  50.5139571 45.9606000
#> 856:       pp_inf_obs[20,1]   374.9700000   359.0000000 105.9342020 93.4038000
#>                 q5           q20           q80           q95      rhat
#>              <num>         <num>         <num>         <num>     <num>
#>   1: -1387.5400000 -1380.9240000 -1368.0040000 -1.362980e+03 1.0044678
#>   2:     3.9234605     4.0517220     4.3216100  4.457815e+00 1.0000011
#>   3:    -0.5992006    -0.1720568     0.7659372  1.187096e+00 1.0005590
#>   4:    -1.2043135    -0.7741840     0.1355158  5.025184e-01 0.9990386
#>   5:    -1.7111430    -1.3085400    -0.3574246  8.786453e-02 1.0065508
#>  ---                                                                  
#> 852:   269.9500000   279.0000000   305.0000000  3.200000e+02 0.9993465
#> 853:   262.0000000   274.0000000   307.0000000  3.260000e+02 1.0053979
#> 854:   253.0000000   269.0000000   314.0000000  3.440000e+02 1.0017054
#> 855:   240.0000000   267.0000000   348.0000000  4.030000e+02 0.9990505
#> 856:   237.0000000   290.0000000   450.0000000  5.552500e+02 1.0005813
#>       ess_bulk ess_tail
#>          <num>    <num>
#>   1:  222.7663 383.0502
#>   2: 1553.9126 680.7740
#>   3: 1121.3239 652.8846
#>   4: 1981.3448 874.8668
#>   5: 1748.3595 708.4019
#>  ---                   
#> 852: 1014.5908 974.8269
#> 853:  946.5257 909.0082
#> 854: 1208.7533 912.9634
#> 855: 1404.1200 860.4009
#> 856: 1458.2840 771.2588

# Posterior predictions
summary(nowcast, type = "posterior_prediction")
#>      reference_date report_date .group max_confirm location age_group confirm
#>              <IDat>      <IDat>  <num>       <int>   <fctr>    <fctr>   <int>
#>   1:     2021-07-13  2021-07-13      1          59       DE       00+      21
#>   2:     2021-07-13  2021-07-14      1          59       DE       00+      33
#>   3:     2021-07-13  2021-07-15      1          59       DE       00+      36
#>   4:     2021-07-13  2021-07-16      1          59       DE       00+      40
#>   5:     2021-07-13  2021-07-17      1          59       DE       00+      43
#>  ---                                                                         
#> 626:     2021-08-20  2021-08-21      1         171       DE       00+     159
#> 627:     2021-08-20  2021-08-22      1         171       DE       00+     171
#> 628:     2021-08-21  2021-08-21      1         112       DE       00+      69
#> 629:     2021-08-21  2021-08-22      1         112       DE       00+     112
#> 630:     2021-08-22  2021-08-22      1          45       DE       00+      45
#>      cum_prop_reported delay new_confirm prop_reported   mean median        sd
#>                  <num> <num>       <int>         <num>  <num>  <num>     <num>
#>   1:         0.3559322     0          21    0.35593220 19.811     18  9.274983
#>   2:         0.5593220     1          12    0.20338983 20.022     19  9.171567
#>   3:         0.6101695     2           3    0.05084746  6.041      6  3.315465
#>   4:         0.6779661     3           4    0.06779661  4.248      4  2.691792
#>   5:         0.7288136     4           3    0.05084746  2.417      2  1.850009
#>  ---                                                                          
#> 626:         0.9298246     1          61    0.35672515 79.802     76 31.840078
#> 627:         1.0000000     2          12    0.07017544 12.666     12  6.022185
#> 628:         0.6160714     0          69    0.61607143 73.979     69 33.181947
#> 629:         1.0000000     1          43    0.38392857 46.518     43 20.106864
#> 630:         1.0000000     0          45    1.00000000 40.202     37 19.857127
#>          mad    q5   q20   q35   q50   q65   q80    q95      rhat  ess_bulk
#>        <num> <num> <num> <num> <num> <num> <num>  <num>     <num>     <num>
#>   1:  8.8956  7.00  12.0    15    18 22.00  27.0  37.00 0.9993374 1160.7699
#>   2:  8.8956  7.00  12.0    16    19 22.00  27.2  36.05 1.0025466  939.1310
#>   3:  2.9652  1.00   3.0     4     6  7.00   9.0  12.00 1.0008153  965.8831
#>   4:  2.9652  1.00   2.0     3     4  5.00   6.0   9.00 0.9998566  949.7067
#>   5:  1.4826  0.00   1.0     1     2  3.00   4.0   6.00 1.0073205  911.5356
#>  ---                                                                       
#> 626: 29.6520 35.00  53.0    64    76 88.35 104.0 138.05 0.9987847  995.1738
#> 627:  5.9304  4.00   7.0    10    12 14.00  17.0  24.00 1.0006961  799.3681
#> 628: 29.6520 31.95  46.0    58    69 80.00  97.0 136.00 0.9998700 1132.3077
#> 629: 17.7912 20.00  29.8    37    43 51.00  61.0  83.00 0.9993141 1062.3664
#> 630: 17.7912 15.00  24.0    30    37 44.00  55.0  74.10 1.0013381 1214.5188
#>       ess_tail
#>          <num>
#>   1:  984.8785
#>   2:  884.2327
#>   3:  994.7491
#>   4:  735.2735
#>   5:  910.5869
#>  ---          
#> 626:  867.8137
#> 627:  813.8822
#> 628: 1025.8119
#> 629:  998.8127
#> 630:  954.4698